Pakistan Demographic and Health Survey 2017-18 Births Recode Data

PKBR71 data of Births Recode from Pakistan Demographic and Health Survey 2017-18.

CASEID

Case Identification

BIDX

Birth column number

V000

Country code and phase

V001

Cluster number

V002

Household number

V003

Respondent's line number

V004

Ultimate area unit

V005

Women's individual sample weight (6 decimals)

V006

Month of interview

V007

Year of interview

V008

Date of interview (CMC)

V008A

Date of interview Century Day Code (CDC)

V009

Respondent's month of birth

V010

Respondent's year of birth

V011

Date of birth (CMC)

V012

Respondent's current age

V013

Age in 5-year groups

V014

Completeness of age information

V015

Result of individual interview

V016

Day of interview

V017

CMC start of calendar

V018

Row of month of interview

V019

Length of calendar

V019A

Number of calendar columns

V020

Ever-married sample

V021

Primary sampling unit

V022

Sample strata for sampling errors

V023

Stratification used in sample design

V024

Region

V025

Type of place of residence

V026

NA - De facto place of residence

V027

Number of visits

V028

Interviewer identification

V029

Keyer identification

V030

Field supervisor

V031

Field editor

V032

NA - Office editor

V034

Line number of husband

V040

NA - Cluster altitude in meters

V042

Household selected for hemoglobin

V044

Selected for Domestic Violence module

V045A

Language of questionnaire

V045B

Language of interview

V045C

Native language of respondent

V046

Translator used

V101

Region

V102

Type of place of residence

V103

NA - Childhood place of residence

V104

NA - Years lived in place of residence

V105

NA - Type of place of previous residence

V105A

NA - Region of previous residence

V106

Highest educational level

V107

Highest year of education

V113

Source of drinking water

V115

Time to get to water source

V116

Type of toilet facility

V119

Household has: electricity

V120

Household has: radio

V121

Household has: television

V122

Household has: refrigerator

V123

Household has: bicycle

V124

Household has: motorcycle/scooter

V125

Household has: car/truck

V127

Main floor material

V128

Main wall material

V129

Main roof material

V130

NA - Religion

V131

NA - Ethnicity

V133

Education in single years

V134

NA - De facto place of residence

V135

Usual resident or visitor

V136

Number of household members (listed)

V137

Number of children 5 and under in household (de jure)

V138

Number of eligible women in household (de facto)

V139

De jure region of residence

V140

De jure type of place of residence

V141

NA - De jure place of residence

V149

Educational attainment

V150

Relationship to household head

V151

Sex of household head

V152

Age of household head

V153

Household has: telephone (land-line)

AWFACTT

All woman factor - total

AWFACTU

All woman factor - urban/rural

AWFACTR

All woman factor - regional

AWFACTE

All woman factor - educational

AWFACTW

All woman factor - wealth index

V155

Literacy

V156

NA - Ever participated in a literacy program outside of primary

V157

Frequency of reading newspaper or magazine

V158

Frequency of listening to radio

V159

Frequency of watching television

V160

Toilet facilities shared with other households

V161

Type of cooking fuel

V166

NA - Result of salt test for iodine

V167

NA - Number of trips in last 12 months

V168

NA - Away for more than one month in last 12 months

V169A

Owns a mobile telephone

V169B

Use mobile telephone for financial transactions

V170

Has an account in a bank or other financial institution

V171A

Use of internet

V171B

Frequency of using internet last month

V190

Wealth index combined

V191

Wealth index factor score combined (5 decimals)

V190A

Wealth index for urban/rural

V191A

Wealth index factor score for urban/rural (5 decimals)

ML101

Type of mosquito bed net(s) slept under last night

V201

Total children ever born

V202

Sons at home

V203

Daughters at home

V204

Sons elsewhere

V205

Daughters elsewhere

V206

Sons who have died

V207

Daughters who have died

V208

Births in last five years

V209

Births in past year

V210

Births in month of interview

V211

Date of first birth (CMC)

V212

Age of respondent at 1st birth

V213

Currently pregnant

V214

Duration of current pregnancy

V215

Time since last menstrual period

V216

Menstruated in last six weeks

V217

Knowledge of ovulatory cycle

V218

Number of living children

V219

Living children + current pregnancy

V220

Living children + current pregnancy (grouped)

V221

Marriage to first birth interval (months)

V222

Last birth to interview (months)

V223

Completeness of current pregnancy information

V224

Entries in birth history

V225

Current pregnancy wanted

V226

Time since last period (comp) (months)

V227

Flag for last period

V228

Ever had a terminated pregnancy

V229

Month pregnancy ended

V230

Year pregnancy ended

V231

CMC pregnancy ended

V232

Completeness of last termination information

V233

Months when pregnancy ended

V234

Other such pregnancies

V235

Index last child prior to maternity-health (calendar)

V237

Birth between last and interview

V238

Births in last three years

V239

Pregnancies terminated before calendar beginning

V240

Month of last termination prior to calendar

V241

Year of last termination prior to calendar

V242

CMC termination ended prior to calendar

V243

Completeness of last termination date prior to calendar

V244

Can women get pregnant after birth and before period

V310

NA - Living children at first use

V311

NA - Children at first use (grouped)

V312

Current contraceptive method

V313

Current use by method type

V315

Month of start of use of method

V316

Year of start of use of method

V317

Date of start of use of method (CMC)

V318

Completeness of information

V319

Years since sterilization

V320

Age at sterilization

V321

Marital duration at sterilization

V322

Parity at sterilization

V323

Brand of pill used

V323A

Brand of condom used

V325A

NA - Cost of current method (in CURRENCY)

V326

Last source for current users

V327

Last source for users by type

V337

Months of use of current method

V359

Last method discontinued in last 5 years

V360

Reason of last discontinuation

V361

Pattern of use

V362

Intention to use

V363

NA - Preferred future method

V364

Contraceptive use and intention

V367

Wanted last child

V372

NA - Shown pill package

V372A

NA - Shown condom package

V375A

NA - Main reason not using a method

V376

NA - Main reason not to use a method

V376A

NA - Would ever use method if married

V379

NA - Source known for any method

V380

NA - Source known for any method

V384A

Heard family planning on radio last few months

V384B

Heard family planning on TV last few months

V384C

Heard family planning in newspaper/magazine last few months

V384D

Heard family planning by text messages on mobile phone

V393

Visited by fieldworker in last 12 months

V393A

Did fieldworker talk about family planning

V394

Visited health facility last 12 months

V395

At health facility, told of family planning

V3A00A

NA - Source of family planning for non-users: government hospital

V3A00B

NA - Source of family planning for non-users: CS public sector

V3A00C

NA - Source of family planning for non-users: CS public sector

V3A00D

NA - Source of family planning for non-users: CS public sector

V3A00E

NA - Source of family planning for non-users: CS public sector

V3A00F

NA - Source of family planning for non-users: CS public sector

V3A00G

NA - Source of family planning for non-users: CS public sector

V3A00H

NA - Source of family planning for non-users: CS public sector

V3A00I

NA - Source of family planning for non-users: CS public sector

V3A00J

NA - Source of family planning for non-users: private hospital/clinic

V3A00K

NA - Source of family planning for non-users: pharmacy

V3A00L

NA - Source of family planning for non-users: private doctor

V3A00M

NA - Source of family planning for non-users: CS private sector

V3A00N

NA - Source of family planning for non-users: CS private sector

V3A00O

NA - Source of family planning for non-users: CS private sector

V3A00P

NA - Source of family planning for non-users: CS private sector

V3A00Q

NA - Source of family planning for non-users: CS private sector

V3A00R

NA - Source of family planning for non-users: CS private sector

V3A00S

NA - Source of family planning for non-users: CS other

V3A00T

NA - Source of family planning for non-users: CS other

V3A00U

NA - Source of family planning for non-users: CS other

V3A00V

NA - Source of family planning for non-users: CS other

V3A00W

NA - Source of family planning for non-users: CS other

V3A00X

NA - Source of family planning for non-users: other

V3A00Y

Source of family planning for non-users: no source

V3A00Z

Source of family planning for non-users: any source

V3A01

NA - Told sterilization would mean no more children

V3A02

Told about side effects

V3A03

Told about side effects by health or family planning worker

V3A04

Told how to deal with side effects

V3A05

Told about other family planning methods

V3A06

Told about other family planning methods by health or family planning worker

V3A07

First source for current method

V3A08A

NA - Reason not using: not married

V3A08B

Reason not using: not having sex

V3A08C

Reason not using: infrequent sex

V3A08D

Reason not using: menopausal/hysterectomy

V3A08E

Reason not using: subfecund/infecund

V3A08F

Reason not using: postpartum amenorrheic

V3A08G

Reason not using: breastfeeding

V3A08H

Reason not using: fatalistic

V3A08I

Reason not using: respondent opposed

V3A08J

Reason not using: husband/partner opposed

V3A08K

Reason not using: others opposed

V3A08L

Reason not using: religious prohibition

V3A08M

Reason not using: knows no method

V3A08N

Reason not using: knows no source

V3A08O

NA - Reason not using: health concerns

V3A08P

Reason not using: fear of side effects/health concerns

V3A08Q

Reason not using: lack of access/too far

V3A08R

Reason not using: costs too much

V3A08S

Reason not using: inconvenient to use

V3A08T

Reason not using: interferes with body's processes

V3A08U

Reason not using: preferred method not available

V3A08V

Reason not using: no method available

V3A08W

NA - Reason not using: CS

V3A08AA

NA - Reason not using: CS

V3A08AB

NA - Reason not using: CS

V3A08AC

NA - Reason not using: CS

V3A08AD

NA - Reason not using: CS

V3A08X

Reason not using: other

V3A08Z

Reason not using: don't know

V3A09A

NA - How many pill cycles did get last time

V3A09B

NA - How many condoms did get last time

V401

Last birth a caesarean section

V404

Currently breastfeeding

V405

Currently amenorrheic

V406

Currently abstaining

V407

NA - Number of times breastfed during night

V408

NA - Number of times breastfed during day

V409

Gave child plain water

V409A

NA - Gave child sugar water

V410

Gave child juice

V410A

NA - Gave child tea or coffee

V411

Gave child tinned, powdered or fresh milk

V411A

Gave child baby formula

V412

NA - Gave child fresh milk

V412A

Gave child fortified baby food (cerelac, etc)

V412B

NA - Gave child other porridge/gruel

V412C

Gave child soup/clear broth

V413

Gave child other liquid

V413A

NA - Gave child CS liquid

V413B

NA - Gave child CS liquid

V413C

NA - Gave child CS liquid

V413D

NA - Gave child CS liquid

V414A

NA - Gave child CS foods

V414B

NA - Gave child CS foods

V414C

NA - Gave child CS foods

V414D

NA - Gave child CS foods

V414E

Gave child bread, noodles, other made from grains

V414F

Gave child potatoes, cassava, or other tubers

V414G

Gave child eggs

V414H

Gave child meat (beef, pork, lamb, chicken, etc)

V414I

Gave child pumpkin, carrots, squash (yellow or orange inside)

V414J

Gave child any dark green leafy vegetables

V414K

Gave child mangoes, papayas, other vitamin A fruits

V414L

Gave child any other fruits

V414M

Gave child liver, heart, other organs

V414N

Gave child fish or shellfish

V414O

Gave child food made from beans, peas, lentils, nuts

V414P

Gave child cheese, yogurt, other milk products

V414Q

NA - Gave child oil, fats, butter, products made of them

V414R

NA - Gave child chocolates, sweets, candies, pastries, etc

V414S

Gave child other solid-semisolid food

V414T

NA - Gave child CS foods

V414U

NA - Gave child CS foods

V414V

Gave child yogurt

V414W

NA - Gave child CS foods

V415

Drank from bottle with nipple yesterday/last night

V416

Heard of oral rehydration

V417

Entries in pregnancy and postnatal care roster

V418

Entries in immunization roster

V418A

Entries in child health roster

V419

Entries in height/weight roster

V420

Measurer's code

V421

NA - Assistant measurer's code

V426

When child put to breast

V437

Respondent's weight in kilograms (1 decimal)

V438

Respondent's height in centimeters (1 decimal)

V439

Height/Age percentile

V440

Height/Age standard deviation

V441

Height/Age percent ref. median

V442

Weight/Height percent ref. median (DHS)

V443

Weight/Height percent ref. median (Fog)

V444

Weight/Height percent ref. median (WHO)

V444A

Weight/Height standard deviation (DHS)

V445

Body Mass Index

V446

Rohrer's index

V447

Result of measurement - height/weight

V447A

Women's age in years (from household questionnaire)

V452A

NA - Under age 18 (from household questionnaire)

V452B

NA - Line number of parent/caretaker

V452C

NA - Read consent statement - hemoglobin

V453

NA - Hemoglobin level (g/dl - 1 decimal)

V454

NA - Currently pregnant (from household questionnaire)

V455

NA - Result of measurement - hemoglobin

V456

NA - Hemoglobin level adjusted for altitude and smoking (g/dl - 1 decimal)

V457

NA - Anemia level

V458

NA - Agrees to referral - anemia

V459

Have mosquito bed net for sleeping (from household questionnaire)

V460

Children under 5 slept under mosquito bed net last night (household questionnaire)

V461

Respondent slept under mosquito bed net

V462

NA - Washed hands before preparing last meal

V463A

Smokes cigarettes

V463B

Smokes pipe full of tobacco

V463C

Chews tobacco

V463D

Snuffs by nose

V463E

NA - Smokes kreteks

V463F

NA - Smokes cigars, cheroots or cigarillos

V463G

Smokes water pipe

V463H

Snuff by mouth

V463I

Chews betel quid with tobacco

V463J

NA - Smokes/uses country specific

V463K

NA - Smokes/uses country specific

V463L

NA - Smokes/uses country specific

V463X

Smokes/uses other

V463Z

Does not use cigarettes and tobacco

V463AA

Frequency smokes cigarettes

V463AB

Frequency currently uses other type of tobacco

V464

Number of cigarettes in last 24 hours

V465

Disposal of youngest child's stools when not using toilet

V466

NA - When child is seriously ill, can decide whether medical treatment sought

V467A

NA - Getting medical help for self: know where to go

V467B

Getting medical help for self: getting permission to go

V467C

Getting medical help for self: getting money needed for treatment

V467D

Getting medical help for self: distance to health facility

V467E

NA - Getting medical help for self: having to take transport

V467F

Getting medical help for self: not wanting to go alone

V467G

NA - Getting medical help for self: concern no female health provider

V467H

NA - Getting medical help for self: concern no provider

V467I

NA - Getting medical help for self: concern no drugs available

V467J

NA - Getting medical help for self: CS

V467K

NA - Getting medical help for self: CS

V467L

NA - Getting medical help for self: CS

V467M

NA - Getting medical help for self: CS

V468

Record for Last Birth

V469E

Times gave child powdered/tinned/fresh milk

V469F

Times gave child infant formula

V469X

Times gave child yogurt

V471A

NA - Mother had tinned, powdered or fresh milk

V471B

NA - Mother had tea or coffee

V471C

NA - Mother had any other liquid

V471D

NA - Mother had CS liquid

V471E

NA - Mother had CS liquid

V471F

NA - Mother had CS liquid

V471G

NA - Mother had CS liquid

V472A

NA - Mother had CS foods

V472B

NA - Mother had CS foods

V472C

NA - Mother had CS foods

V472D

NA - Mother had CS foods

V472E

NA - Mother had bread, noodles, other made from grains

V472F

NA - Mother had potatoes, cassava, or other tubers

V472G

NA - Mother had eggs

V472H

NA - Mother had meat (beef, pork, lamb, chicken, etc)

V472I

NA - Mother had pumpkin, carrots, squash (yellow or orange inside)

V472J

NA - Mother had any dark green leafy vegetables

V472K

NA - Mother had mangoes, papayas, other vitamin A fruits

V472L

NA - Mother had any other fruits

V472M

NA - Mother had liver, heart, other organs

V472N

NA - Mother had fish or shellfish

V472O

NA - Mother had food made from beans, peas, lentils, nuts

V472P

NA - Mother had cheese, yogurt, other milk products

V472Q

NA - Mother had oil, fats, butter, products made of them

V472R

NA - Mother had chocolates, sweets, candies, pastries, etc

V472S

NA - Mother had other solid-semisolid food

V472T

NA - Mother had CS foods

V472U

NA - Mother had CS foods

V473A

NA - Read consent statement - HIV

V473B

NA - Result of measurement - HIV

V474

Heard of tuberculosis or TB

V474A

Tuberculosis spread by: air when coughing or sneezing

V474B

Tuberculosis spread by: sharing utensils

V474C

Tuberculosis spread by: touching a person with TB

V474D

Tuberculosis spread by: food

V474E

Tuberculosis spread by: sexual contact

V474F

Tuberculosis spread by: mosquito bites

V474G

NA - Tuberculosis spread by: CS

V474H

NA - Tuberculosis spread by: CS

V474I

NA - Tuberculosis spread by: CS

V474J

NA - Tuberculosis spread by: CS

V474X

Tuberculosis spread by: other

V474Z

Tuberculosis spread by: don't know

V475

Tuberculosis can be cured

V476

NA - Keep secret when family member gets TB

V477

Number of injections in last 12 months

V478

Number of injections administered by a health worker

V479

NA - Place of injection by health worker

V480

Syringe and needle from new, unopened package

V481

Covered by health insurance

V481A

Health insurance type: mutual/community organization

V481B

Health insurance type: provided by employer

V481C

Health insurance type: social security (sehat sahulat)

V481D

Health insurance type: private/commercially purchased

V481E

NA - Health insurance type: CS

V481F

NA - Health insurance type: CS

V481G

NA - Health insurance type: CS

V481H

NA - Health insurance type: CS

V481X

Health insurance type: other

V482A

NA - Arrange care of biological children under age 18

V482B

NA - Primary caregiver of children under age 18

V482C

NA - Arrange care of non-biological children < 18

V501

Current marital status

V502

Currently/formerly/never in union

V503

Number of unions

V504

Currently residing with husband/partner

V505

Number of other wives

V506

Respondent's rank among wives

V507

Month of first cohabitation

V508

Year of first cohabitation

V509

Date of first cohabitation (CMC)

V510

Completeness of date information in V509

V511

Age at first cohabitation

V512

Years since first cohabitation

V513

Cohabitation duration (grouped)

V525

Age at first sex

V527

Time since last sex

V528

Time since last sex (in days)

V529

Time since last sex (in months)

V530

Flag for V529

V531

Age at first sex (imputed)

V532

Flag for V531

V535

NA - Ever been married or in union

V536

Recent sexual activity

V537

Months of abstinence

V538

NA - How previous marriage or union ended

V539

NA - Person who received most of late husband's property

V540

NA - Respondent received any of late husband's assets or valuables

V541

NA - Intends to wait until marriage to have sex

V602

Fertility preference

V603

Preferred waiting time for birth of a/another child

V604

Preferred waiting time for birth of a/another child (grouped)

V605

Desire for more children

V613

Ideal number of children

V614

Ideal number of children (grouped)

V616

Time for future birth

V621

Husband's desire for children

V623

Exposure

V624

Unmet need

V625

Exposure (definition 2)

V626

Unmet need (definition 2)

V625A

Exposure to need for contraception (definition 3)

V626A

Unmet need for contraception (definition 3)

V627

Ideal number of boys

V628

Ideal number of girls

V629

Ideal number of either sex

V631

NA - Problem if became pregnant

V632

Decision maker for using contraception

V632A

Decision maker for not using contraception

V633A

NA - Reason for not having sex: husband has STI

V633B

Reason for not having sex: husband has other women

V633C

NA - Reason for not having sex: recent birth

V633D

NA - Reason for not having sex: tired, not in mood

V633E

NA - Reason for not having sex: CS

V633F

NA - Reason for not having sex: CS

V633G

NA - Reason for not having sex: CS

V634

NA - Husband knows that respondent is using contraception

V701

Husband/partner's education level

V702

Husband/partner's highest year of education (at level in V701)

V704

Husband/partner's occupation

V704A

Husband/partner worked in last 7 days/12 months

V705

Husband/partner's occupation (grouped)

V714

Respondent currently working

V714A

Respondent has a job, but currently absent

V715

Husband/partner's total number of years of education

V716

Respondent's occupation

V717

Respondent's occupation (grouped)

V719

Respondent works for family, others, self

V721

NA - Respondent works at home or away

V729

Husband/partner's educational attainment

V730

Husband/partner's age

V731

Respondent worked in last 12 months

V732

Respondent employed all year/seasonal

V739

Person who usually decides how to spend respondent's earnings

V740

NA - Type of land where respondent works

V741

Type of earnings from respondent's work

V743A

Person who usually decides on respondent's health care

V743B

Person who usually decides on large household purchases

V743C

NA - Person who usually decides on household purchases for daily needs

V743D

Person who usually decides on visits to family or relatives

V743E

NA - Person who usually decides on food to be cooked each day

V743F

Person who usually decides what to do with money husband earns

V744A

Beating justified if wife goes out without telling husband

V744B

Beating justified if wife neglects the children

V744C

Beating justified if wife argues with husband

V744D

Beating justified if wife refuses to have sex with husband

V744E

Beating justified if wife burns the food

V745A

Owns a house alone or jointly

V745B

Owns land alone or jointly

V745C

Title deed on house own by respondent

V745D

Title deed on land own by respondent

V746

Respondent earns more than husband/partner

BORD

Birth order number

B0

Child is twin

B1

Month of birth

B2

Year of birth

B3

Date of birth (CMC)

B4

Sex of child

B5

Child is alive

B6

Age at death

B7

Age at death (months, imputed)

B8

Current age of child

B9

Child lives with whom

B10

Completeness of information

B11

Preceding birth interval (months)

B12

Succeeding birth interval (months)

B13

Flag for age at death

B15

Live birth between births

B16

Child's line number in household

B17

Day of birth

B18

Century Day Code of birth (CDC)

B19

Current age of child in months (months since birth for dead children)

B20

NA - Duration of pregnancy

MIDX

Index to Birth History

M1

Number of tetanus injections before birth

M1A

Number of tetanus injections before pregnancy

M1B

NA - Month of last tetanus injection before pregnancy

M1C

NA - Year of last tetanus injection before pregnancy

M1D

Years ago received last tetanus injection before pregnancy

M1E

Last tetanus injection before pregnancy (CMC)

M2A

Prenatal: doctor

M2B

Prenatal: Nurse/Midwife/Lady health visitor

M2C

NA - Prenatal: CS health professional

M2D

NA - Prenatal: CS health professional

M2E

NA - Prenatal: CS health professional

M2F

NA - Prenatal: CS health professional

M2G

Prenatal: traditional birth attendant (DAI-TBA)

M2H

Prenatal: Lady H. Worker

M2I

Prenatal: Homeopath

M2J

Prenatal: Hakim

M2K

Prenatal: other

M2L

Prenatal: Dispenser/compounder

M2M

NA - Prenatal: CS other

M2N

Prenatal: no one

M3A

Assistance: doctor

M3B

Assistance: Nurse

M3C

Assistance: Midwife

M3D

Assistance: Lady health visitor

M3E

Assistance: Community midwife

M3F

NA - Assistance: CS health professional

M3G

Assistance: dai/traditional birth attendant

M3H

Assistance: Family welfare wk

M3I

Assistance: Lady h worker

M3J

Assistance: Homeopath

M3K

Assistance: other

M3L

Assistance: Hakim

M3M

Assistance: Relative/friend

M3N

Assistance: no one

M4

Duration of breastfeeding

M5

Months of breastfeeding

M6

Duration of amenorrhea

M7

Months of amenorrhea

M8

Duration of abstinence

M9

Months of abstinence

M10

Wanted pregnancy when became pregnant

M11

Desired time would have waited

M13

Timing of 1st antenatal check (months)

M14

Number of antenatal visits during pregnancy

M15

Place of delivery

M17

Delivery by caesarean section

M17A

Timing on decision for caesarean section

M18

Size of child at birth

M19

Birth weight in kilograms (3 decimals)

M19A

Weight at birth/recall

M27

NA - Flag for breastfeeding

M28

Flag for amenorrhea

M29

Flag for abstinence

M34

When child put to breast

M35

NA - Number of times breastfed during previous night

M36

NA - Number of times breastfed during previous day

M38

Drank from bottle with nipple yesterday/last night

M39A

Did eat any solid, semi-solid or soft foods yesterday

M39

Number of times ate solid, semi-solid or soft food yesterday

M42A

NA - During pregnancy: weighed

M42B

NA - During pregnancy: height measured

M42C

During pregnancy: blood pressure taken

M42D

During pregnancy: urine sample taken

M42E

During pregnancy: blood sample taken

M43

NA - Told about pregnancy complications

M44

NA - Told where to go for pregnancy complications

M45

During pregnancy, given or bought iron tablets/syrup

M46

Days tablets or syrup taken

M47

NA - During pregnancy, had difficulty with daylight vision

M48

NA - During pregnancy, had difficulty with night blindness

M49A

NA - During pregnancy took: SP/fansidar for malaria

M49B

NA - During pregnancy took: chloroquine for malaria

M49C

NA - During pregnancy took: CS drug for malaria

M49D

NA - During pregnancy took: CS drug for malaria

M49E

NA - During pregnancy took: CS drug for malaria

M49F

NA - During pregnancy took: CS drug for malaria

M49G

NA - During pregnancy took: CS drug for malaria

M49X

NA - During pregnancy took: other drug for malaria

M49Z

NA - During pregnancy took: don't know

M49Y

NA - During pregnancy took: no drug for malaria

M54

NA - Received Vitamin A dose in first 2 months after delivery

M55

Given child anything other than breast milk

M55A

NA - First 3 days, given milk (other than breast milk)

M55B

NA - First 3 days, given plain water

M55C

NA - First 3 days, given sugar/glucose water

M55D

NA - First 3 days, given gripe water

M55E

NA - First 3 days, given sugar/salt solution

M55F

NA - First 3 days, given fruit juice

M55G

NA - First 3 days, given infant formula

M55H

NA - First 3 days, given tea/infusions

M55I

NA - First 3 days, given honey

M55J

NA - First 3 days, given coffee

M55K

NA - First 3 days, given country specific

M55L

NA - First 3 days, given country specific

M55M

NA - First 3 days, given country specific

M55N

NA - First 3 days, given country specific

M55O

NA - First 3 days, given country specific

M55X

NA - First 3 days, given other

M55Z

NA - First 3 days, given nothing

M57A

Antenatal care: respondent's home

M57B

Antenatal care: other home

M57C

NA - Antenatal care: CS home

M57D

NA - Antenatal care: CS home

M57E

Antenatal care: government hospital

M57F

Antenatal care: Rural health centre/Mother child health centre

M57G

Antenatal care: BHU(Basic Health Unit)

M57H

Antenatal care: Community midwife

M57I

NA - Antenatal care: CS public health

M57J

NA - Antenatal care: CS public health

M57K

NA - Antenatal care: CS public health

M57L

Antenatal care: Other public sector

M57M

Antenatal care: private hospital/clinic

M57N

Antenatal care: private doctor

M57O

Antenatal care: Homeopath

M57P

Antenatal care: Dispenser/compounder

M57Q

Antenatal care: Other private medical sector

M57R

NA - Antenatal care: CS private medical

M57S

Antenatal care: Hakim

M57T

NA - Antenatal care: CS other

M57U

NA - Antenatal care: CS other

M57V

NA - Antenatal care: CS other

M57X

Antenatal care: other

M60

Drugs for intestinal parasites during pregnancy

M61

Time spent at place of delivery

M62

Respondent's health checked before discharge

M63

How long before discharging respondent health check took place

M64

Who checked respondent health before discharge

M65A

NA - Reason didn't deliver at health facility: cost too much

M65B

NA - Reason didn't deliver at health facility: facility not open

M65C

NA - Reason didn't deliver at health facility: too far/no transport

M65D

NA - Reason didn't deliver at health facility: don't trust facility/poor service

M65E

NA - Reason didn't deliver at health facility: no female provider

M65F

NA - Reason didn't deliver at health facility: husband/family didn't allow

M65G

NA - Reason didn't deliver at health facility: not necessary

M65H

NA - Reason didn't deliver at health facility: not customary

M65I

NA - Reason didn't deliver at health facility: CS

M65J

NA - Reason didn't deliver at health facility: CS

M65K

NA - Reason didn't deliver at health facility: CS

M65L

NA - Reason didn't deliver at health facility: CS

M65X

NA - Reason didn't deliver at health facility: other

M66

Respondent's health checked after discharge/delivery at home

M67

How long after discharge/delivery at home respondent health check took place

M68

Who checked respondent health after discharge/delivery at home

M69

Where respondent check after discharge/delivery at home took place

M70

Baby postnatal check within 2 months

M71

Time after delivery postnatal check took place

M72

Person who performed postnatal checkup

M73

Place baby was first checked

M74

Child's health checked before discharge

M75

How long before discharging child health check took place

M76

Who checked child health before discharge

M77

Was child put on mother's chest and bare skin after birth

M78A

During first 2 days health provider: examine cord

M78B

During first 2 days health provider: measure temperature

M78C

During first 2 days health provider: counsel on newborn dangers

M78D

During first 2 days health provider: counsel on breastfeeding

M78E

During first 2 days health provider: observe breastfeeding

M78F

NA - During first 2 days health provider: CS

M78G

NA - During first 2 days health provider: CS

M78H

NA - During first 2 days health provider: CS

M78I

NA - During first 2 days health provider: CS

M78J

NA - During first 2 days health provider: CS

HIDX

Index to birth history

H1

Has health card

H1A

Has health card and or other vaccination document

H2

Received BCG

H2D

BCG day

H2M

BCG month

H2Y

BCG year

H3

Received DPT 1

H3D

DPT 1 day

H3M

DPT 1 month

H3Y

DPT 1 year

H4

Received POLIO 1

H4D

POLIO 1 day

H4M

POLIO 1 month

H4Y

POLIO 1 year

H5

Received DPT 2

H5D

DPT 2 day

H5M

DPT 2 month

H5Y

DPT 2 year

H6

Received POLIO 2

H6D

POLIO 2 day

H6M

POLIO 2 month

H6Y

POLIO 2 year

H7

Received DPT 3

H7D

DPT 3 day

H7M

DPT 3 month

H7Y

DPT 3 year

H8

Received POLIO 3

H8D

POLIO 3 day

H8M

POLIO 3 month

H8Y

POLIO 3 year

H9

Received MEASLES 1

H9D

MEASLES 1 day

H9M

MEASLES 1 month

H9Y

MEASLES 1 year

H9A

Received MEASLES 2 (MMR)

H9AD

MEASLES 2 (MMR) day

H9AM

MEASLES 2 (MMR) month

H9AY

MEASLES 2 (MMR) year

H0

Received POLIO 0

H0D

POLIO 0 day

H0M

POLIO 0 month

H0Y

POLIO 0 year

H10

Ever had vaccination

H33

Received Vitamin A1 (most recent)

H33D

Vitamin A1 day

H33M

Vitamin A1 month

H33Y

Vitamin A1 year

H35

NA - Any vaccinations in last 2 years part of campaign

H36A

NA - Vaccinated during Campaign A

H36B

NA - Vaccinated during Campaign B

H36C

NA - Vaccinated during Campaign C

H36D

NA - Vaccinated during Campaign D

H36E

NA - Vaccinated during Campaign E

H36F

NA - Vaccinated during Campaign F

H40

NA - Received Vitamin A2 (2nd most recent)

H40D

NA - Vitamin A2 day

H40M

NA - Vitamin A2 month

H40Y

NA - Vitamin A2 year

H41A

NA - Received vitamin A after most recent in card

H41B

NA - Ever received Vitamin A dose

H50

NA - Received Hepatitis B at birth

H50D

NA - Hepatitis B at birth day

H50M

NA - Hepatitis B at birth month

H50Y

NA - Hepatitis B at birth year

H51

Received Pentavalent 1

H51D

Pentavalent 1 day

H51M

Pentavalent 1 month

H51Y

Pentavalent 1 year

H52

Received Pentavalent 2

H52D

Pentavalent 2 day

H52M

Pentavalent 2 month

H52Y

Pentavalent 2 year

H53

Received Pentavalent 3

H53D

Pentavalent 3 day

H53M

Pentavalent 3 month

H53Y

Pentavalent 3 year

H54

Received Pneumococcal 1

H54D

Pneumococcal 1 day

H54M

Pneumococcal 1 month

H54Y

Pneumococcal 1 year

H55

Received Pneumococcal 2

H55D

Pneumococcal 2 day

H55M

Pneumococcal 2 month

H55Y

Pneumococcal 2 year

H56

Received Pneumococcal 3

H56D

Pneumococcal 3 day

H56M

Pneumococcal 3 month

H56Y

Pneumococcal 3 year

H57

NA - Received Rotavirus 1

H57D

NA - Rotavirus 1 day

H57M

NA - Rotavirus 1 month

H57Y

NA - Rotavirus 1 year

H58

NA - Received Rotavirus 2

H58D

NA - Rotavirus 2 day

H58M

NA - Rotavirus 2 month

H58Y

NA - Rotavirus 2 year

H59

NA - Received Rotavirus 3

H59D

NA - Rotavirus 3 day

H59M

NA - Rotavirus 3 month

H59Y

NA - Rotavirus 3 year

H60

Received Polio inactive

H60D

Polio inactive day

H60M

Polio inactive month

H60Y

Polio inactive year

H61

Received Hepatitis B 1

H61D

Hepatitis B 1 day

H61M

Hepatitis B 1 month

H61Y

Hepatitis B 1 year

H62

Received Hepatitis B 2

H62D

Hepatitis B 2 day

H62M

Hepatitis B 2 month

H62Y

Hepatitis B 2 year

H63

Received Hepatitis B 3

H63D

Hepatitis B 3 day

H63M

Hepatitis B 3 month

H63Y

Hepatitis B 3 year

H64

Received HiB 1

H64D

HiB 1 day

H64M

HiB 1 month

H64Y

HiB 1 year

H65

Received HiB 2

H65D

HiB 2 day

H65M

HiB 2 month

H65Y

HiB 2 year

H66

Received HiB 3

H66D

HiB 3 day

H66M

HiB 3 month

H66Y

HiB 3 year

H80A

In the last 7 days given: Baby Active (multiple micronutrient powder)

H80B

In the last 7 days given: Plumpy'NUT (ready to use therapeutic food)

H80C

In the last 7 days given: Plumpy'DOZ (ready to use supplemental food)

H80D

NA - In the last 7 days given: CS

H80E

NA - In the last 7 days given: CS

H80F

NA - In the last 7 days given: CS

H80G

NA - In the last 7 days given: CS

HIDXA

Index to birth history

H11

Had diarrhea recently

H11B

NA - Blood in the stools

H12A

Diarrhea: government hospital

H12B

Diarrhea: Rural health centre/Mother child health centre

H12C

Diarrhea: BHU(Basic Health Unit)

H12D

Diarrhea: Lady h.worker

H12E

Diarrhea: Other public sector

H12F

NA - Diarrhea: CS public sector

H12G

NA - Diarrhea: CS public sector

H12H

NA - Diarrhea: CS public sector

H12I

NA - Diarrhea: CS public sector

H12J

Diarrhea: private hospital/clinic

H12K

Diarrhea: private pharmacy (chemist/medical store)

H12L

Diarrhea: private doctor

H12M

Diarrhea: Homeopath

H12N

Diarrhea: Dispenser/compounder

H12O

Diarrhea: Other private medical sector

H12P

NA - Diarrhea: CS private medical

H12Q

NA - Diarrhea: CS private medical

H12R

NA - Diarrhea: CS private medical

H12S

Diarrhea: Shop

H12T

Diarrhea: Hakim

H12U

Diarrhea: DAI, TBA

H12V

Diarrhea: Community midwife

H12W

NA - Diarrhea: CS other sector

H12X

Diarrhea: Other

H12Y

Diarrhea: no treatment

H12Z

Diarrhea: medical treatment

H13

Given oral rehydration

H13B

Given pre-packaged ORS liquid

H14

Given recommended home solution

H15

Given antibiotic pills or syrups

H15A

Given antimotility

H15B

Given antibiotic injection

H15C

Given intravenous (IV)

H15D

Given home remedy, herbal medicine

H15E

Given zinc

H15F

Given other (not antibiotic, antimotility, zinc)

H15G

Given unknown pill or syrup

H15H

Given non-antibiotic injection

H15I

Given unknown injection

H15J

Given Rice Starch

H15K

NA - Given CS

H15L

NA - Given CS

H15M

NA - Given CS

H20

Given other treatment

H21A

Given no treatment

H21

Received any treatment

H22

Had fever in last two weeks

H31

Had cough in last two weeks

H31B

Short, rapid breaths

H31C

Problem in the chest or blocked or running nose

H31D

NA - Had fever/cough in last 2 weeks: amount offered to drink

H31E

NA - Had fever/cough in last 2 weeks: amount offered to eat

H32A

Fever/cough: government hospital

H32B

Fever/cough: Rural health centre/Mother child health centre

H32C

Fever/cough: BHU(Basic Health Unit)

H32D

Fever/cough: Lady h.worker

H32E

Fever/cough: Other public sector

H32F

NA - Fever/cough: CS public sector

H32G

NA - Fever/cough: CS public sector

H32H

NA - Fever/cough: CS public sector

H32I

NA - Fever/cough: CS public sector

H32J

Fever/cough: private hospital/ clinic

H32K

Fever/cough: private pharmacy (chemist/medical store)

H32L

Fever/cough: private doctor

H32M

Fever/cough: Homeopath

H32N

Fever/cough: Dispenser/compounder

H32O

Fever/cough: Other private medical sector

H32P

NA - Fever/cough: CS private medical

H32Q

NA - Fever/cough: CS private medical

H32R

NA - Fever/cough: CS private medical

H32S

Fever/cough: Shop

H32T

Fever/cough: Hakim

H32U

Fever/cough: DAI/TBA

H32V

Fever/cough: Community midwife

H32W

NA - Fever/cough: CS other sector

H32X

Fever/cough: other

H32Y

Fever/cough: no treatment

H32Z

Fever/cough: medical treatment

H34

Vitamin A in last 6 months

H37A

SP/Fansidar taken for fever

H37B

Chloroquine taken for fever

H37C

Amodiaquine taken for fever

H37D

Quinine taken for fever

H37DA

Quinine injection/IV taken for fever

H37E

Combination with artemisinin taken for fever

H37AA

Artesunate rectal taken for fever

H37AB

Artesunate injection/IV taken for fever

H37F

NA - CS antimalarial taken for fever

H37G

NA - CS antimalarial taken for fever

H37H

Other antimalarial taken for fever

H37I

Antibiotic pill/syrup taken for fever

H37J

Antibiotic injection taken for fever

H37K

NA - Aspirin taken for fever

H37L

Acetaminophen/paracetamol/panadol taken for fever

H37M

Ibuprofen taken for fever

H37N

Ponstan taken for fever/cough

H37O

Cough syrup taken for fever/cough

H37P

NA - CS other for fever/cough

H37X

Other taken for fever

H37Y

Nothing taken and didn't seek treatment for fever/cough

H37Z

Don't know if or what was taken for fever

H38

Diarrhea in last 2 weeks: amount offered to drink

H39

Diarrhea in last 2 weeks: amount offered to eat

H42

Taking iron pills, sprinkles or syrup

H43

Drugs for intestinal parasites in last 6 months

H44A

Place first sought treatment for diarrhea

H44B

NA - Days after diarrhea sought advice or treatment

H44C

NA - Still has diarrhea

H45

NA - Times zinc was given

H46A

Place first sought treatment for fever

H46B

Number of days after fever began sought advice or treatment

H47

NA - Blood taken from child's finger/heel for testing

HWIDX

Index to Birth History

HW1

Child's age in months

HW2

Child's weight in kilograms (1 decimal)

HW3

Child's height in centimeters (1 decimal)

HW4

Height/Age percentile

HW5

Height/Age standard deviation

HW6

Height/Age percent of ref. median

HW7

Weight/Age percentile

HW8

Weight/Age standard deviation

HW9

Weight/Age percent of ref. median

HW10

Weight/Height percentile

HW11

Weight/Height standard deviation

HW12

Weight/Height percent of ref. median

HW13

Result of measurement - height/weight

HW15

Height: lying or standing

HW16

Day of birth of child

HW17

Date measured (day)

HW18

Date measured (month)

HW19

Date measured (year)

HW51

NA - Line number of parent/caretaker

HW52

NA - Read consent statement - hemoglobin

HW53

NA - Hemoglobin level (g/dl - 1 decimal)

HW55

NA - Result of measurement - hemoglobin

HW56

NA - Hemoglobin level adjusted for altitude (g/dl - 1 decimal)

HW57

NA - Anemia level

HW58

NA - Agrees to referral - anemia

HW70

Height/Age standard deviation (new WHO)

HW71

Weight/Age standard deviation (new WHO)

HW72

Weight/Height standard deviation (new WHO)

HW73

BMI standard deviation (new WHO)

IDXML

Index to Birth History

ML0

Type of mosquito bed net(s) child slept under last night

ML1

NA - Number of times took fansidar during pregnancy

ML2

NA - Source of antimalarial during pregnancy

ML11

NA - Child has fever/cough now

ML12

NA - Child has had convulsions in last 2 weeks

ML13A

Fansidar taken for fever

ML13B

Chloroquine taken for fever

ML13C

Amodiaquine taken for fever

ML13D

Quinine taken for fever

ML13DA

Quinine injection/IV taken for fever

ML13E

Combination with artemisinin taken for fever

ML13AA

Artesunate rectal taken for fever

ML13AB

Artesunate injection/IV taken for fever

ML13F

NA - CS antimalarial taken for fever

ML13G

NA - CS antimalarial taken for fever

ML13H

Other antimalarial taken for fever

ML13I

Antibiotic pill/syrup taken for fever

ML13J

Antibiotic injection taken for fever

ML13K

NA - Aspirin taken for fever

ML13L

Acetaminophen/paracetamol/panadol taken for fever

ML13M

Ibuprofen taken for fever

ML13N

Ponstan taken for fever

ML13O

Cough syrup taken for fever

ML13P

NA - CS other taken for fever

ML13X

Other taken for fever

ML13Y

Nothing taken and didn't seek treatment for fever/cough

ML13Z

Don't know if or what was taken for fever

ML14A

NA - Injection given for fever/convulsion

ML14B

NA - Suppository given for fever/convulsion

ML14Y

NA - No suppository or injection given for fever/convulsion

ML14Z

NA - Don't know if suppository or injection given for fever/convulsion

ML15A

NA - Timing of start of fansidar

ML15B

NA - Number of days child took fansidar

ML15C

NA - First source for fansidar

ML16A

NA - Timing of start of chloroquine

ML16B

NA - Number of days child took chloroquine

ML16C

NA - First source for chloroquine

ML17A

NA - Timing of start of amodiaquine

ML17B

NA - Number of days child took amodiaquine

ML17C

NA - First source for amodiaquine

ML18A

NA - Timing of start of quinine

ML18B

NA - Number of days child took quinine

ML18C

NA - First source for quinine

ML19A

NA - For fever/convulsion: consulted traditional healer

ML19B

NA - For fever/convulsion: gave tepid sponging

ML19C

NA - For fever/convulsion: gave herbs

ML19D

NA - For fever/convulsion: CS

ML19E

NA - For fever/convulsion: CS

ML19F

NA - For fever/convulsion: CS

ML19X

NA - For fever/convulsion: other

ML19Y

NA - For fever/convulsion: gave/did nothing

ML19Z

NA - For fever/convulsion: don't know if something else was done

ML20A

NA - Timing of start of combination with artemisinin

ML20B

NA - Number of days child took combination with artemisinin

ML20C

NA - First source for combination with artemisinin

ML21A

NA - Timing of start of CS antimalarial use

ML21B

NA - Number of days child took CS antimalarial

ML21C

NA - First source for CS antimalarial

ML22A

NA - Timing of start of CS antimalarial use

ML22B

NA - Number of days child took CS antimalarial

ML22C

NA - First source for CS antimalarial

ML23A

NA - Timing of start of other anti-malarial use

ML23B

NA - Number of days child took other anti-malarial

ML23C

NA - First source for other anti-malarial

ML24C

NA - First source for antibiotic pill/syrup

ML25A

NA - Timing of start of artesunate

SDIST

District

SEDUC

Woman's educational level for report

SWFEDUC1

All woman factor - education as will be used in the report

SWFREGUR

All woman factor - region combined with urban/rural

S109

Highest class completed

S121A

Mother tongue

S207BB

Number of pregnancy losses

S229B

Sought health care after pregnancy ended

S229CA

From whom sought help: Doctor

S229CB

From whom sought help: Nurse/midwife/Lady health visitor

S229CC

From whom sought help: DAI-TBA

S229CD

From whom sought help: Lady h. worker

S229CE

From whom sought help: Homeopath

S229CF

From whom sought help: Hakim

S229CG

From whom sought help: Dispenser/compounder

S229CX

From whom sought help: Other

S301A

Heard FP before marriage

S316

Source for current method when started

S323AA

Advised by health of FP worker: Help select method

S323AB

Advised by health of FP worker: Explain method use

S325

Last source of FP

S708A

Blood relationship with husband

S708B

Type of relationship with husband

S709A

Had a say in choosing the husband

S816AA

Messages conveyed: Limiting the family size

S816B

Messages conveyed: Disadvantages of getting married at young age

S816C

Messages conveyed: Spacing births

S816D

Messages conveyed: Use of contraception

S816E

Messages conveyed: Family welfare

S816F

Messages conveyed: Maternal and child health

S816G

Messages conveyed: Less children means prosperous life

S816H

Messages conveyed: More children means poverty and starvation

S816I

Messages conveyed: Importance of breastfeeding

S816J

Messages conveyed: Importance of girls' education

S816K

Messages conveyed: Reduction in maternal deaths

S816X

Messages conveyed: Other

S816Z

Messages conveyed: Don't know/do not remember

S816A

Message effectivity

SPEDUC

Husband/partner's education level for report

S908A

Worked before got married

S908B

Worked after marriage

S908C

Years when start working after marriage

S915A

Work at home or away

S924A

Inherited any land or house

S927A

Have autonomy to sell the house

S930A

Have autonomy to sell the land

S932F

Justifies domestic violence: Neglected In-laws

S1032A

Think there is treatment for HIV

S1032B

Knows where HIV treatment can be received

S1039A

People with HIV get discriminatory treatment

S1107A

Currently use any type of drugs

S1107E

Duration of treatment of TB now a days

S1107F

Ever been told by a doctor/nurse/Lady health visitor that god forbid have TB

S1107G

Ever heard of Hepatitis B or C

S1107H

Can a person do something to avoid getting Hepatitis B or C

S1107IA

Avoid getting hepatitis B or C: Practice safe sex

S1107IB

Avoid getting hepatitis B or C: Safe blood transfer

S1107IC

Avoid getting hepatitis B or C: Use disposable syringe

S1107ID

Avoid getting hepatitis B or C: Avoid contaminated food/water

S1107IE

Avoid getting hepatitis B or C: Avoid contact with infected person

S1107IF

Avoid getting hepatitis B or C: Ensure instruments of dentists are properly sterilized

S1107IX

Avoid getting hepatitis B or C: Other

S1107IZ

Avoid getting hepatitis B or C: Don't know

S1107J

Ever been tested for Hepatitis B or C

S1107K

Months ago tested for Hepatitis B or C

S1110A

Receive any cash/kind benefit from Benazir Income Support Program

S1221A

Suffered abortion, miscarriage, stillbirth or any other health problem

S1228P

Govt social welfare center: Person respondent went to seek help

S1228Q

Local leaders (wadera, khan): Person respondent went to seek help

S1228R

Public representatives: Person respondent went to seek help

S1228Y

Don't know/don't remember: Person respondent went to seek help

S1228Z

Refuse to answer: Person respondent went to seek help

S1228A

Satisfied with the help provided

S1228BA

Reason sought help: Encouraged by friends/family

S1228BB

Reason sought help: Public campaign

S1228BC

Reason sought help: Knowledge about helpline

S1228BD

Reason sought help: Service provider's attitude

S1228BE

Reason sought help: Could not endure more

S1228BF

Reason sought help: Badly injured

S1228BG

Reason sought help: Threatened or tried to kill her

S1228BH

Reason sought help: Threatened to hit children

S1228BI

Reason sought help: Saw children suffering

S1228BJ

Reason sought help: Thrown out of the home

S1228BK

Reason sought help: Afraid of more violence

S1228BX

Reason sought help: Other

S1228BY

Reason sought help: Don't know/don't remember

S1228BZ

Reason sought help: Refused/no answer

S1228CA

Consequences when sought help: Got threats

S1228CB

Consequences when sought help: Embarrassed/ashamed

S1228CC

Consequences when sought help: Blamed

S1228CD

Consequences when sought help: Marriage breakup

S1228CE

Consequences when sought help: Lost children

S1228CF

Consequences when sought help: Exploitation the person who approached to help

S1228CG

Consequences when sought help: Faced no consequences

S1228CX

Consequences when sought help: Other

S1228CY

Consequences when sought help: Don't know

S1228CZ

Consequences when sought help: Refused/no answer

S1229AA

Reason not sought help: Fear of threats

S1229AB

Reason not sought help: Fear of consequences

S1229AC

Reason not sought help: More violence

S1229AD

Reason not sought help: Violence was not serious

S1229AE

Reason not sought help: Embarrassed/ashamed

S1229AF

Reason not sought help: Or would be blamed

S1229AG

Reason not sought help: No trust on any one to help

S1229AH

Reason not sought help: Unaware if anyone can help

S1229AI

Reason not sought help: Unaware if a female can help

S1229AJ

Reason not sought help: Afraid of marriage breakup

S1229AK

Reason not sought help: Afraid would lose children

S1229AL

Reason not sought help: Bring bad name to family

S1229AM

Reason not sought help: Did not know her options/way out

S1229AX

Reason not sought help: Other

S1229AY

Reason not sought help: Don't know

S1229AZ

Reason not sought help: Refused/no answer

S1229BA

Consequences due to violence: Isolated self

S1229BB

Consequences due to violence: Fired from job

S1229BC

Consequences due to violence: Quit job

S1229BD

Consequences due to violence: Stop participating in decision making

S1229BE

Consequences due to violence: Got divorced

S1229BF

Consequences due to violence: Not faced any consequences

S1229BX

Consequences due to violence: Other

S1229BY

Consequences due to violence: Don't know

S1229BZ

Consequences due to violence: Refused/no answer

SV005

Women's individual sample weight (AJK/GB, 6 decimals)

SD005

Weight for Domestic Violence (AJK/GB, 6 decimals)

IDX94

Index to Birth History

S409A

Satisfied with service provided

S413AA

During antenatal care advised on: Early breastfeeding

S413AB

During antenatal care advised on: Exclusive breastfeeding

S413AC

During antenatal care advised on: Balanced diet

S437A

Talked about using FP method when was in facility

S443A

Talked about using FP method after left facility

S451A

Talked about using FP method when delivery was at home or other place non private/public

S469A

Reason stop breastfeeding

References

National Institute of Population Studies (NIPS), Pakistan (http://nips.org.pk/Home.htm).

Examples

load("PKBR71")

#> Warning: cannot open compressed file 'PKBR71', probable reason 'No such file or directory'

#> Error in readChar(con, 5L, useBytes = TRUE): cannot open the connection

PKBR71

#> # A tibble: 50,495 x 1,186
#>    CASEID  BIDX V000   V001  V002  V003  V004   V005  V006  V007  V008 V008A
#>    <fct>  <int> <fct> <int> <int> <int> <int>  <int> <int> <int> <int> <int>
#>  1 "    …     1 PK7       1     1     2     1 868683     2  2018  1418 43135
#>  2 "    …     2 PK7       1     1     2     1 868683     2  2018  1418 43135
#>  3 "    …     3 PK7       1     1     2     1 868683     2  2018  1418 43135
#>  4 "    …     4 PK7       1     1     2     1 868683     2  2018  1418 43135
#>  5 "    …     5 PK7       1     1     2     1 868683     2  2018  1418 43135
#>  6 "    …     1 PK7       1     4     1     1 868683     2  2018  1418 43134
#>  7 "    …     2 PK7       1     4     1     1 868683     2  2018  1418 43134
#>  8 "    …     3 PK7       1     4     1     1 868683     2  2018  1418 43134
#>  9 "    …     4 PK7       1     4     1     1 868683     2  2018  1418 43134
#> 10 "    …     5 PK7       1     4     1     1 868683     2  2018  1418 43134
#> # … with 50,485 more rows, and 1,174 more variables: V009 <int>, V010 <int>,
#> #   V011 <int>, V012 <int>, V013 <fct>, V014 <fct>, V015 <fct>, V016 <int>,
#> #   V017 <int>, V018 <fct>, V019 <fct>, V019A <fct>, V020 <fct>, V021 <int>,
#> #   V022 <fct>, V023 <fct>, V024 <fct>, V025 <fct>, V026 <fct>, V027 <int>,
#> #   V028 <int>, V029 <int>, V030 <int>, V031 <int>, V032 <int>, V034 <fct>,
#> #   V040 <int>, V042 <fct>, V044 <fct>, V045A <fct>, V045B <fct>, V045C <fct>,
#> #   V046 <fct>, V101 <fct>, V102 <fct>, V103 <fct>, V104 <fct>, V105 <fct>,
#> #   V105A <fct>, V106 <fct>, V107 <fct>, V113 <fct>, V115 <fct>, V116 <fct>,
#> #   V119 <fct>, V120 <fct>, V121 <fct>, V122 <fct>, V123 <fct>, V124 <fct>,
#> #   V125 <fct>, V127 <fct>, V128 <fct>, V129 <fct>, V130 <fct>, V131 <int>,
#> #   V133 <fct>, V134 <fct>, V135 <fct>, V136 <int>, V137 <int>, V138 <int>,
#> #   V139 <fct>, V140 <fct>, V141 <fct>, V149 <fct>, V150 <fct>, V151 <fct>,
#> #   V152 <fct>, V153 <fct>, AWFACTT <int>, AWFACTU <int>, AWFACTR <int>,
#> #   AWFACTE <int>, AWFACTW <int>, V155 <fct>, V156 <fct>, V157 <fct>,
#> #   V158 <fct>, V159 <fct>, V160 <fct>, V161 <fct>, V166 <fct>, V167 <fct>,
#> #   V168 <fct>, V169A <fct>, V169B <fct>, V170 <fct>, V171A <fct>, V171B <fct>,
#> #   V190 <fct>, V191 <int>, V190A <fct>, V191A <int>, ML101 <fct>, V201 <int>,
#> #   V202 <int>, V203 <int>, V204 <int>, V205 <int>, …