Pakistan Demographic and Health Survey 2017-18 Fieldworker Questionnaire Data

PKFW71 data of Fieldworker Questionnaire from Pakistan Demographic and Health Survey 2017-18.

FW101

Fieldworker code

FW000

Country code and phase

FW102

Fieldworker region of residence

FW103

Fieldworker type of place of residence

FW104

Fieldworker age

FW105

Fieldworker sex

FW106

Fieldworker marital status

FW107

Fieldworker number of living children

FW108

Fieldworker ever had a child who died

FW109

NA- Highest level of school attended by fieldworker

FW110

Highest grade/form/year at that level completed by fieldworker

FW111

Fieldworker religion

FW112

Fieldworker ethnicity

FW113A

English

FW113B

Urdu

FW113C

Sindhi

FW113D

Punjabi

FW113E

Sariaki

FW113F

Baluchi

FW113G

Pashto

FW113X

Other language

FW114

Fieldworker's mother tongue/native language

FW115

Has fieldworker ever worked on a DHS

FW116

Has fieldworker ever worked on any other survey

FW117

Was fieldworker working for NIPS at the time employed for DHS

FW118

Is fieldworker a permanent or temporary employee of agency

References

National Institute of Population Studies (NIPS), Pakistan (http://nips.org.pk/Home.htm).

Examples

load("PKFW71")

#> Warning: cannot open compressed file 'PKFW71', probable reason 'No such file or directory'

#> Error in readChar(con, 5L, useBytes = TRUE): cannot open the connection

PKFW71

#> # A tibble: 147 x 26
#>    FW101 FW000 FW102 FW103 FW104 FW105 FW106 FW107 FW108 FW109 FW110 FW111 FW112
#>    <int> <fct> <fct> <fct> <int> <fct> <fct> <int> <fct> <fct> <fct> <fct> <fct>
#>  1   100 PK7   Punj… Rural    26 Male  Neve…     0 No    Prim… Mast… Musl… Sara…
#>  2   101 PK7   Punj… Urban    40 Fema… Curr…     3 No    NA    14    Musl… Sara…
#>  3   102 PK7   Punj… Rural    26 Fema… Neve…     0 No    NA    Mast… Musl… Sara…
#>  4   103 PK7   Punj… Urban    36 Fema… Curr…     0 No    NA    Mast… Musl… Sara…
#>  5   104 PK7   Punj… Urban    30 Male  Neve…     0 No    NA    Mast… Musl… Punj…
#>  6   105 PK7   Punj… Urban    38 Male  Curr…     2 No    NA    Mast… Musl… Sara…
#>  7   107 PK7   Punj… Urban    26 Fema… Neve…     0 No    NA    Mast… Musl… Muha…
#>  8   108 PK7   Punj… Urban    35 Male  Curr…     2 No    NA    Mast… Musl… Sara…
#>  9   200 PK7   Punj… Urban    37 Male  Curr…     0 No    NA    Mast… Musl… Path…
#> 10   201 PK7   Punj… Rural    27 Fema… Neve…     0 No    NA    Mast… Musl… Sara…
#> # … with 137 more rows, and 13 more variables: FW113A <fct>, FW113B <fct>,
#> #   FW113C <fct>, FW113D <fct>, FW113E <fct>, FW113F <fct>, FW113G <fct>,
#> #   FW113X <fct>, FW114 <fct>, FW115 <fct>, FW116 <fct>, FW117 <fct>,
#> #   FW118 <fct>