Pakistan Demographic and Health Survey 2006-07 Household Recode Data

PKHR52 data of Household Recode from Pakistan Demographic and Health Survey 2006-07.

hhid

case identification

hv000

country code and phase

hv001

cluster number

hv002

household number

hv003

respondent's line number

hv004

ultimate area unit

hv005

sample weight

hv006

month of interview

hv007

year of interview

hv008

date of interview (cmc)

hv009

number of household members

hv010

number of eligible women in hh

hv011

number of eligible men in hh

hv012

number of de jure members

hv013

number of de facto members

hv014

number of children 5 and under

hv015

result of household interview

hv016

day of interview

hv017

number of visits

hv018

interviewer identification

hv019

keyer identification

hv020

ever-married sample

hv021

primary sampling unit

hv022

sample stratum number

hv023

sample domain

hv024

region

hv025

type of place of residence

hv026

place of residence

hv027

selection for male/husb. int.

hv028

sample weight for male subsamp

hv030

field supervisor

hv031

field editor

hv032

office editor

hv033

ultimate area selection pr-na

hv035

number of eligible children for height & weight

hv040

cluster altitude in meters -na

hv041

total adults measured

hv042

selection for hemoglobin

hv043

selection for women's statu-na

hv044

selection for domestic viol-na

hv201

source of drinking water

hv202

source of non-drinking wate-na

hv204

time to get to water source

hv205

type of toilet facility

hv206

has electricity

hv207

has radio

hv208

has television

hv209

has refrigerator

hv210

has bicycle

hv211

has motorcycle/scooter

hv212

has car/truck

hv213

main floor material

hv214

main wall material

hv215

main roof material

hv216

rooms used for sleeping

hv217

relationship structure

hv218

line number of head of househ.

hv219

sex of head of household

hv220

age of head of household

hv221

has telephone

hv225

share toilet with other households

hv226

type of cooking fuel

hv227

have bednet for sleeping

hv228

children under 5 slept under bednet last night

hv234

test salt for iodine -na

hv235

location of source for wate-na

hv236

person fetching water -na

hv237

anything done to water to make safe to drink

hv237a

water usually treated by: boil

hv237b

water usually treated by: add bleach/chlorine

hv237c

water usually treated by: strain through a cloth

hv237d

water usually treated by: use water filter

hv237e

water usually treated by: solar disinfection

hv237f

water usually treated by: let it stand and settle

hv237g

water usually treated by: c-na

hv237h

water usually treated by: c-na

hv237i

water usually treated by: c-na

hv237j

water usually treated by: c-na

hv237k

water usually treated by: c-na

hv237x

water usually treated by: other

hv237z

water usually treated by: don't know

hv238

number of households sharin-na

hv239

food cooked on stove or ope-na

hv240

household has a chimney, ho-na

hv241

food cooked in the house / -na

hv242

household has separate room-na

hv243a

has a mobile telephone -na

hv243b

has a watch

hv243c

has an animal-drawn cart

hv243d

has a boat with a motor

hv244

own land usable for agriculture

hv245

hectares for agricultural l-na

hv246

livestock, herds or farm animals

hv246a

cattle own -na

hv246b

cows, bulls own

hv246c

horses, donkeys, mules own

hv246d

goats own

hv246e

sheep own

hv246f

chickens own

hv246g

buffalo own

hv246h

camels own

hv246i

cs own -na

hv246j

cs own -na

hv246k

cs own -na

hv247

owns a bank account -na

hv248

number of sick people 18-59-na

hv249

member of the hh died last -na

hv250

number of members who died -na

hv251

number of orphans and vulne-na

hv270

wealth index

hv271

wealth index factor score (5 decimals)

hml1

number of mosquito nets

hml1a

number of mosquito nets with specific information

hml2

number of children under bednet previous night

shdist

district

shtype

type of questionnaire

shv005

hh weight for all households

shv218

line number of head of household (all households).

shv219

sex of head of household (all households)

shv220

age of head of household (all households)

shbirths

total births in household since january 2003

shdeaths

total deaths in household since january 2003

shcdeath

total deaths < 5/sbs from q38

shfdeath

total female deaths age 12-49 from q39

sh107f

room cooler, air conditioner owns

sh107g

washing machine owns

sh107h

water pump owns

sh107i

bed owns

sh107j

chairs owns

sh107k

almirah/cabinet owns

sh107l

clock owns

sh107m

sofa owns

sh107n

sewing machine owns

sh107o

camera owns

sh107p

personal computer owns

sh113

status of house

sh127a

hh uses coil to avoid mosquitos

sh127b

hh uses mats to avoid mosquitos

sh127c

hh uses spray to avoid mosquitos

sh127d

hh uses electric spray repellant to avoid mosquitos

sh127e

hh uses insect repellant to avoid mosquitos

sh127f

hh uses smoke to avoid mosquitos

sh127g

hh uses fan to avoid mosquitos

sh127h

hh uses covered windows/screens to avoid mosquitos

sh127x

hh uses other to avoid mosquitos

sh127y

hh uses nothing to avoid mosquitos

sh128

currently have any medicines for treating malaria in the hou

hmlidx.1

net number

hmlidx.2

net number

hml3.1

net observed -na

hml3.2

net observed -na

hml4.1

months ago obtained mosquit-na

hml4.2

months ago obtained mosquit-na

hml5.1

was net treated with insecticide when bought

hml5.2

was net treated with insecticide when bought

hml6.1

mosquito net treated with i-na

hml6.2

mosquito net treated with i-na

hml7.1

brand of bednet -na

hml7.2

brand of bednet -na

hml8.1

bednet treated since receiving

hml8.2

bednet treated since receiving

hml9.1

months ago re-treated

hml9.2

months ago re-treated

hml10.1

itn net

hml10.2

itn net

hml11.1

number of persons who slept under this net

hml11.2

number of persons who slept under this net

hml21.1

did someone sleep under bednet last night

hml21.2

did someone sleep under bednet last night

hmla.1

person who slept in mosquito net

hmla.2

person who slept in mosquito net

hmlb.1

person who slept in mosquito net

hmlb.2

person who slept in mosquito net

hmlc.1

person who slept in mosquito net

hmlc.2

person who slept in mosquito net

hmld.1

person who slept in mosquit-na

hmld.2

person who slept in mosquit-na

hmle.1

person who slept in mosquit-na

hmle.2

person who slept in mosquit-na

sh20.1

line number

sh20.2

line number

sh20.3

line number

sh20.4

line number

sh20.5

line number

sh20.6

line number

sh20.7

line number

sh20.8

line number

sh20.9

line number

sh22.1

gender

sh22.2

gender

sh22.3

gender

sh22.4

gender

sh22.5

gender

sh22.6

gender

sh22.7

gender

sh22.8

gender

sh22.9

gender

sh23m.1

month of birth

sh23m.2

month of birth

sh23m.3

month of birth

sh23m.4

month of birth

sh23m.5

month of birth

sh23m.6

month of birth

sh23m.7

month of birth

sh23m.8

month of birth

sh23m.9

month of birth

sh23y.1

year of birth

sh23y.2

year of birth

sh23y.3

year of birth

sh23y.4

year of birth

sh23y.5

year of birth

sh23y.6

year of birth

sh23y.7

year of birth

sh23y.8

year of birth

sh23y.9

year of birth

sh24.1

live birth

sh24.2

live birth

sh24.3

live birth

sh24.4

live birth

sh24.5

live birth

sh24.6

live birth

sh24.7

live birth

sh24.8

live birth

sh24.9

live birth

sh25.1

currently alive

sh25.2

currently alive

sh25.3

currently alive

sh25.4

currently alive

sh25.5

currently alive

sh25.6

currently alive

sh25.7

currently alive

sh25.8

currently alive

sh25.9

currently alive

sh26.1

line number in household roster

sh26.2

line number in household roster

sh26.3

line number in household roster

sh26.4

line number in household roster

sh26.5

line number in household roster

sh26.6

line number in household roster

sh26.7

line number in household roster

sh26.8

line number in household roster

sh26.9

line number in household roster

sh23c.1

cmc birth

sh23c.2

cmc birth

sh23c.3

cmc birth

sh23c.4

cmc birth

sh23c.5

cmc birth

sh23c.6

cmc birth

sh23c.7

cmc birth

sh23c.8

cmc birth

sh23c.9

cmc birth

sh23f.1

cmc flag

sh23f.2

cmc flag

sh23f.3

cmc flag

sh23f.4

cmc flag

sh23f.5

cmc flag

sh23f.6

cmc flag

sh23f.7

cmc flag

sh23f.8

cmc flag

sh23f.9

cmc flag

sh29.1

line number

sh29.2

line number

sh29.3

line number

sh29.4

line number

sh29.5

line number

sh29.6

line number

sh31a.1

usual resident or visitor

sh31a.2

usual resident or visitor

sh31a.3

usual resident or visitor

sh31a.4

usual resident or visitor

sh31a.5

usual resident or visitor

sh31a.6

usual resident or visitor

sh31.1

gender

sh31.2

gender

sh31.3

gender

sh31.4

gender

sh31.5

gender

sh31.6

gender

sh32m.1

month of death

sh32m.2

month of death

sh32m.3

month of death

sh32m.4

month of death

sh32m.5

month of death

sh32m.6

month of death

sh32y.1

year of death

sh32y.2

year of death

sh32y.3

year of death

sh32y.4

year of death

sh32y.5

year of death

sh32y.6

year of death

sh33.1

age at death

sh33.2

age at death

sh33.3

age at death

sh33.4

age at death

sh33.5

age at death

sh33.6

age at death

sh34.1

death was female, 12-49 years of age

sh34.2

death was female, 12-49 years of age

sh34.3

death was female, 12-49 years of age

sh34.4

death was female, 12-49 years of age

sh34.5

death was female, 12-49 years of age

sh34.6

death was female, 12-49 years of age

sh35.1

pregnant at the time of death

sh35.2

pregnant at the time of death

sh35.3

pregnant at the time of death

sh35.4

pregnant at the time of death

sh35.5

pregnant at the time of death

sh35.6

pregnant at the time of death

sh36.1

died during childbirth

sh36.2

died during childbirth

sh36.3

died during childbirth

sh36.4

died during childbirth

sh36.5

died during childbirth

sh36.6

died during childbirth

sh37.1

died within 6-weeks after delivery

sh37.2

died within 6-weeks after delivery

sh37.3

died within 6-weeks after delivery

sh37.4

died within 6-weeks after delivery

sh37.5

died within 6-weeks after delivery

sh37.6

died within 6-weeks after delivery

sh32c.1

cmc death

sh32c.2

cmc death

sh32c.3

cmc death

sh32c.4

cmc death

sh32c.5

cmc death

sh32c.6

cmc death

sh32f.1

cmc flag

sh32f.2

cmc flag

sh32f.3

cmc flag

sh32f.4

cmc flag

sh32f.5

cmc flag

sh32f.6

cmc flag

References

National Institute of Population Studies (NIPS), Pakistan (http://nips.org.pk/Home.htm).

Examples

load("PKHR52")

#> Warning: cannot open compressed file 'PKHR52', probable reason 'No such file or directory'

#> Error in readChar(con, 5L, useBytes = TRUE): cannot open the connection

PKHR52

#> # A tibble: 95,441 x 332
#>    hhid  hv000 hv001 hv002 hv003 hv004  hv005 hv006 hv007 hv008 hv009 hv010
#>    <chr> <chr> <int> <int> <int> <int>  <int> <int> <int> <int> <int> <int>
#>  1 "   … PK5       1     1     2     1 0.         2  2007  1286     7     0
#>  2 "   … PK5       1     2     2     1 0.         2  2007  1286     7     0
#>  3 "   … PK5       1     3     1     1 0.         2  2007  1286     1     0
#>  4 "   … PK5       1     4     1     1 0.         2  2007  1286     7     0
#>  5 "   … PK5       1     5     4     1 0.         2  2007  1286     4     0
#>  6 "   … PK5       1     6     2     1 0.         2  2007  1286     5     0
#>  7 "   … PK5       1     7     3     1 1.17e6     2  2007  1286     7     1
#>  8 "   … PK5       1     8     2     1 0.         2  2007  1286    19     0
#>  9 "   … PK5       1     9     2     1 0.         2  2007  1286     6     0
#> 10 "   … PK5       1    11     1     1 0.         2  2007  1286     8     0
#> # … with 95,431 more rows, and 320 more variables: hv011 <int>, hv012 <int>,
#> #   hv013 <int>, hv014 <int>, hv015 <fct>, hv016 <int>, hv017 <int>,
#> #   hv018 <int>, hv019 <int>, hv020 <fct>, hv021 <int>, hv022 <int>,
#> #   hv023 <fct>, hv024 <fct>, hv025 <fct>, hv026 <fct>, hv027 <fct>,
#> #   hv028 <int>, hv030 <int>, hv031 <int>, hv032 <int>, hv033 <int>,
#> #   hv035 <int>, hv040 <int>, hv041 <int>, hv042 <fct>, hv043 <fct>,
#> #   hv044 <fct>, hv201 <fct>, hv202 <fct>, hv204 <int>, hv205 <fct>,
#> #   hv206 <fct>, hv207 <fct>, hv208 <fct>, hv209 <fct>, hv210 <fct>,
#> #   hv211 <fct>, hv212 <fct>, hv213 <fct>, hv214 <fct>, hv215 <fct>,
#> #   hv216 <int>, hv217 <fct>, hv218 <int>, hv219 <fct>, hv220 <int>,
#> #   hv221 <fct>, hv225 <fct>, hv226 <fct>, hv227 <fct>, hv228 <fct>,
#> #   hv234 <fct>, hv235 <fct>, hv236 <fct>, hv237 <fct>, hv237a <fct>,
#> #   hv237b <fct>, hv237c <fct>, hv237d <fct>, hv237e <fct>, hv237f <fct>,
#> #   hv237g <fct>, hv237h <fct>, hv237i <fct>, hv237j <fct>, hv237k <fct>,
#> #   hv237x <fct>, hv237z <fct>, hv238 <fct>, hv239 <fct>, hv240 <fct>,
#> #   hv241 <fct>, hv242 <fct>, hv243a <fct>, hv243b <fct>, hv243c <fct>,
#> #   hv243d <fct>, hv244 <fct>, hv245 <fct>, hv246 <fct>, hv246a <fct>,
#> #   hv246b <int>, hv246c <int>, hv246d <int>, hv246e <int>, hv246f <int>,
#> #   hv246g <int>, hv246h <int>, hv246i <fct>, hv246j <fct>, hv246k <fct>,
#> #   hv247 <fct>, hv248 <int>, hv249 <fct>, hv250 <fct>, hv251 <int>,
#> #   hv270 <fct>, hv271 <int>, hml1 <int>, …