Pakistan Demographic and Health Survey 2006-07 Children's Recode Data

PKKR52 data of Children's Recode from Pakistan Demographic and Health Survey 2006-07.

caseid

case identification

midx

notappl

v000

country code and phase

v001

cluster number

v002

household number

v003

respondent's line number

v004

ultimate area unit

v005

sample weight

v006

month of interview

v007

year of interview

v008

date of interview (cmc)

v009

respondent's month of birth

v010

respondent's year of birth

v011

date of birth (cmc)

v012

current age - respondent

v013

age 5-year groups

v014

completeness of information

v015

result of individual interview

v016

day of interview

v017

cmc start of calendar

v018

row of month of interview

v019

length of calendar

v019a

number of calendar columns

v020

ever-married sample

v021

primary sampling unit

v022

sample stratum number

v023

sample domain

v024

region

v025

type of place of residence

v026

de facto place of residence

v027

number of visits

v028

interviewer identification

v029

keyer identification

v030

field supervisor

v031

field editor

v032

office editor

v033

ultimate area selection prob. - na

v034

line number of husband -na

v040

cluster altitude in meters -na

v042

household selection for hemoglobin

v043

selection for women's statu-na

v044

selection for domestic viol-na

v101

region

v102

type of place of residence

v103

childhood place of residenc-na

v104

years lived in place of res-na

v105

type of place of previous r-na

v106

highest educational level

v107

highest year of education

v113

source of drinking water

v115

time to get to water source

v116

type of toilet facility

v119

has electricity

v120

has radio

v121

has television

v122

has refrigerator

v123

has bicycle

v124

has motorcycle/scooter

v125

has car/truck

v127

main floor material

v128

main wall material

v129

main roof material

v130

religion -na

v131

ethnicity - mother tongue

v133

education in single years

v134

de facto place of residence

v135

usual resident or visitor

v136

number of household members

v137

number of children 5 and under

v138

number of eligible women in hh

v139

de jure region of residence

v140

de jure type of place of res.

v141

de jure place of residence

v149

educational attainment

v150

relationship to household head

v151

sex of household head

v152

age of household head

v153

has telephone

awfactt

all woman factor - total

awfactu

all woman factor - urban/rural

awfactr

all woman factor - regional

awfacte

all woman factor - educational

awfactw

all woman factor - wealth index

v155

literacy

v156

ever participated in a lite-na

v157

frequency of reading newpap-na

v158

frequency of listening to r-na

v159

frequency of watching telev-na

v160

toilet facilities shared

v161

type of cooking fuel

v166

results of salt iodine test-na

v167

number of trips in last 12 -na

v168

away for more than one mont-na

v190

wealth index

v191

wealth index factor score (5 decimals)

ml101

type of bednet(s) slept under last night

v201

total children ever born

v202

sons at home

v203

daughters at home

v204

sons elsewhere

v205

daughters elsewhere

v206

sons who have died

v207

daughters who have died

v208

births in last five years

v209

births in past year

v210

births in month of interview

v211

date of first birth (cmc)

v212

age of respondent at 1st birth

v213

currently pregnant

v214

duration of current pregnancy

v215

time since last menstrual perd

v216

menstruated in last six weeks

v217

knowledge of ovulatory cycl-na

v218

number of living children

v219

living children + current preg

v220

living children + curr preg 6+

v221

marriage to first birth int.

v222

last birth to interview

v223

completeness of curr preg. inf

v224

entries in birth history

v225

current pregnancy wanted

v226

time since last period (comp)

v227

flag for last period

v228

ever had a terminated preg.

v229

month pregnancy ended

v230

year pregnancy ended

v231

cmc pregnancy ended

v232

completeness of last term. inf

v233

months when pregnancy ended

v234

other such pregnancies

v235

index last child prior to cal.

v237

birth between last & interview

v238

births in last three years

v239

pregnancies terminated befo-na

v240

month of last termination p-na

v241

year of last termination pr-na

v242

cmc termination ended prior-na

v243

completeness of last termin-na

v310

living children at first us-na

v311

children at first use (grpd)-na

v312

current contraceptive method

v313

current use by method type

v315

month of start of use of method

v316

year of start of use of method

v317

date of start of use of method (cmc)

v318

completeness of information

v319

years since sterilization

v320

age at sterilization

v321

marital duration at sterilization

v322

parity at sterilization

v323

brand of pill used

v323a

brand of condom used

v325a

cost of current method

v326

last source for current users

v327

last source for users by type

v337

months of use of current method

v359

last method disc. last 5 ye-na

v360

reason of last discontinuat-na

v361

pattern of use

v362

intention to use

v363

preferred future method -na

v364

contraceptive use & intention

v367

wanted last child

v372

shown pill package

v372a

shown condom package

v375a

main reason not using a met-na

v376

main reason not to use a meth.

v376a

would ever use method if ma-na

v379

source known for any method-na

v380

source known for any method-na

v384a

heard fp on radio last months

v384b

heard fp on tv last months

v384c

heard fp newspaper last mon-na

v393

visited by fp worker last 12m

v394

visited health facil. last -na

v395

at health facility, told of-na

v3a00a

source of fp for non-users: government hosp./rhsc

v3a00b

source of fp for non-users: rural government health cntr/mc

v3a00c

source of fp for non-users:-na

v3a00d

source of fp for non-users: mobile service camp

v3a00e

source of fp for non-users: lady health worker/visitor * see

v3a00f

source of fp for non-users: family welfare center

v3a00g

source of fp for non-users: basic health unit

v3a00h

source of fp for non-users: male mobilizer

v3a00i

source of fp for non-users: oth public sector

v3a00j

source of fp for non-users: private hosp/clin

v3a00k

source of fp for non-users: private pharmacy

v3a00l

source of fp for non-users: private doctor

v3a00m

source of fp for non-users:-na

v3a00n

source of fp for non-users:-na

v3a00o

source of fp for non-users:-homeopath

v3a00p

source of fp for non-users: dispenser/compounder

v3a00q

source of fp for non-users:-na

v3a00r

source of fp for non-users: oth med.priv sect

v3a00s

source of fp for non-users: shop

v3a00t

source of fp for non-users:-na

v3a00u

source of fp for non-users: friend, relative

v3a00v

source of fp for non-users: hakim

v3a00w

source of fp for non-users: dai, tba

v3a00x

source of fp for non-users: other

v3a00y

source of fp for non-users: no source

v3a00z

source of fp for non-users: any source

v3a01

told sterilization would me-na

v3a02

told about side effects

v3a03

told about side effects by -na

v3a04

told how to deal with side effects

v3a05

told about other fp methods

v3a06

told about other fp methods-na

v3a07

first source for current method (sterilization only)

v3a08a

reason not using: not marri-na

v3a08b

reason not using: not having sex

v3a08c

reason not using: infrequen-na

v3a08d

reason not using: menopausal/hysterectomy

v3a08e

reason not using: subfecund/infecund

v3a08f

reason not using: postpartum amenorrheic

v3a08g

reason not using: breastfeeding

v3a08h

reason not using: fatalistic

v3a08i

reason not using: respondent opposed

v3a08j

reason not using: husband/partner opposed

v3a08k

reason not using: others opposed

v3a08l

reason not using: religious prohibition

v3a08m

reason not using: knows no method

v3a08n

reason not using: knows no source

v3a08o

reason not using: health concerns

v3a08p

reason not using: fear of side effects

v3a08q

reason not using: lack of access/too far

v3a08r

reason not using: costs too much

v3a08s

reason not using: inconvenient to use

v3a08t

reason not using: interferes with bodys processes

v3a08u

reason not using: wants to use in the future

v3a08v

reason not using: country s-na

v3a08w

reason not using: country s-na

v3a08x

reason not using: other

v3a08z

reason not using: don't know

v3a09a

how many pill cycles did get last time

v3a09b

how many condoms did get last time

v401

last birth ceasarean section

v404

currently breastfeeding

v405

currently amenorrheic

v406

currently abstaining

v407

times breastfed during night

v408

times breastfed during day

v409

gave child plain water -na

v409a

gave child sugar water -na

v410

gave child juice -na

v410a

gave child tea or coffe -na

v411

gave child tinned/powder or-na

v411a

gave child baby formula -na

v412

gave child fresh milk -na

v412a

gave child baby cereal -na

v412b

gave child other porridge/g-na

v413

gave child other liquid -na

v413a

gave child cs liquid -na

v413b

gave child cs liquid -na

v413c

gave child cs liquid -na

v413d

gave child cs liquid -na

v414a

gave child cs foods -na

v414b

gave child cs foods -na

v414c

gave child cs foods -na

v414d

gave child cs foods -na

v414e

gave child bread, noddles, -na

v414f

gave child potatoes, cassav-na

v414g

gave child eggs -na

v414h

gave child meat (beef, pork)-na

v414i

gave child pumpink, carrots-na

v414j

gave child any dark green l-na

v414k

gave child mangoes, papayas-na

v414l

gave child any other fruits-na

v414m

gave child liver, heart, ot-na

v414n

gave child fish or shellfis-na

v414o

gave child food made from b-na

v414p

gave child cheese, yogurt ,-na

v414q

gave child oil, fats, butte-na

v414r

gave child chocolates, swee-na

v414s

gave child other solid-semi-na

v414t

gave child cs foods -na

v414u

gave child cs foods -na

v415

drank from bottle with nipple

v416

heard of oral rehydration

v417

entries in maternity table

v418

entries in health table

v419

entries in height/weight table

v420

measurer's code -na

v421

assistant measurer's code -na

v426

when child put to breast

v437

respondent's weight (kilos)--na

v438

respondent's height (cms-1d)-na

v439

ht/a percentile (resp.) -na

v440

ht/a standard deviations (r)-na

v441

ht/a percent ref. median (r)-na

v442

wt/ht percent ref. median ()-na

v443

wt/ht percent ref. median ()-na

v444

wt/ht percent ref. median ()-na

v444a

wt/ht std deviations(resp) -na

v445

body mass index for respond-na

v446

rohrer's index for responde-na

v447

result of measurement of re-na

v447a

women's age in years from h-na

v452a

under age 18 (hh report) -na

v452b

line no. of parent/caretake-na

v452c

read consent statement -na

v453

hemoglobin level (g/dl - 1) -na

v454

currently pregnant (househo)-na

v455

result of measuring (hemogl)-na

v456

hemoglobin level adjusted b-na

v457

anemia level -na

v458

agrees to referral -na

v459

have bednet for sleeping (household report)

v460

children under 5 slept under bednet last night (hh report)

v461

respondent slept under bednet

v462

washed hands before prepari-na

v463a

smokes cigarettes -na

v463b

smokes pipe -na

v463c

chewing tobacco -na

v463d

smokes snuff -na

v463e

smokes country specific -na

v463f

smokes country specific -na

v463g

smokes country specific -na

v463x

smokes other -na

v463z

smokes nothing -na

v464

number of cigarettes in las-na

v465

disposal of youngest child'-na

v466

when child is seriously ill-na

v467a

getting medical help for se-na

v467b

getting medical help for se-na

v467c

getting medical help for se-na

v467d

getting medical help for se-na

v467e

getting medical help for se-na

v467f

getting medical help for se-na

v467g

getting medical help for se-na

v467h

getting medical help for se-na

v467i

getting medical help for se-na

v467j

getting medical help for se-na

v467k

getting medical help for se-na

v467l

getting medical help for se-na

v467m

getting medical help for se-na

v468

columns used for last birth only variables

v471a

mother had tinned, powdered-na

v471b

mother had tea or coffe -na

v471c

mother had any other liquid-na

v471d

mother had cs liquid -na

v471e

mother had cs liquid -na

v471f

mother had cs liquid -na

v471g

mother had cs liquid -na

v472a

mother had cs foods -na

v472b

mother had cs foods -na

v472c

mother had cs foods -na

v472d

mother had cs foods -na

v472e

mother had bread, noddles, -na

v472f

mother had potatoes, cassav-na

v472g

mother had eggs -na

v472h

mother had meat (beef, pork)-na

v472i

mother had pumpink, carrots-na

v472j

mother had any dark green l-na

v472k

mother had mangoes, papayas-na

v472l

mother had any other fuits -na

v472m

mother had liver, heart, ot-na

v472n

mother had fish or shellfis-na

v472o

mother had food made from b-na

v472p

mother had cheese, yogurt ,-na

v472q

mother had oil, fats, butte-na

v472r

mother had chocolates, swee-na

v472s

mother had other solid-semi-na

v472t

mother had cs foods -na

v472u

mother had cs foods -na

v473a

read consent statement (hiv)-na

v473b

result of measuring (hiv) -na

v474

heard of tuberculosis or tb

v474a

tuberculosis spread by: air when coughing or sneezing

v474b

tuberculosis spread by: sharing utensils

v474c

tuberculosis spread by: touching a person with tb

v474d

tuberculosis spread through food

v474e

tuberculosis spread by: sexual contact

v474f

tuberculosis spread by: mosquito bites

v474g

tuberculosis spread by: cs -na

v474h

tuberculosis spread by: cs -na

v474i

tuberculosis spread by: cs -na

v474j

tuberculosis spread by: cs -na

v474x

tuberculosis spread by: other

v474z

tuberculosis spread by: don't know

v475

can tuberculosis be cured

v476

keep secret when family mem-na

v477

number of injections in last 12 months

v478

injections administered by a health worker

v479

source for injection by health workder

v480

syringe and needle from new, unopened package

v481

covered by health insurance-na

v481a

health insurance type: mutu-na

v481b

health insurance type: prov-na

v481c

health insurance type: soci-na

v481d

health insurance type: priv-na

v481e

health insurance type: cs -na

v481f

health insurance type: cs -na

v481g

health insurance type: cs -na

v481h

health insurance type: cs -na

v481x

health insurance type: othe-na

v482a

arrange care of biological -na

v482b

primary caregiver of childr-na

v482c

arrange care of non-biologi-na

v501

current marital status

v502

currently/formerly/never marr.

v503

number of unions

v504

husband lives in house

v505

number of other wives

v506

wife rank number -na

v507

month of first marriage

v508

year of first marriage

v509

date of first marriage (cmc)

v510

completeness of date inform.

v511

age at first marriage

v512

years since first marriage

v513

marital duration (grouped)

v525

age at first intercourse -na

v527

time since last intercourse-na

v528

time since last intercourse-na

v529

time since last intercourse-na

v530

flag for v529 -na

v531

age at first intercourse (i)-na

v532

flag for v531 -na

v535

have ever been married -na

v536

recent sexual activity -na

v537

months of abstinence -na

v538

how previous marriage or un-na

v539

who received most of late h-na

v540

respondent received any of -na

v541

intends to postpone interco-na

v602

fertility preference

v603

preferred waiting time

v604

preferred waiting time (grpd)

v605

desire for more children

v613

ideal number of children

v614

ideal number of children (grp)

v616

time for future birth

v621

husbands desire for children

v623

exposure

v624

unmet need

v625

exposure (definition 2)

v626

unmet need (definition 2)

v627

ideal number of boys

v628

ideal number of girls

v629

ideal number of either sex

v631

problem if became pregnant -na

v632

decision maker for using co-na

v633a

reason for not having sex: -na

v633b

reason for not having sex: -na

v633c

reason for not having sex: -na

v633d

reason for not having sex: -na

v633e

reason for not having sex: -na

v633f

reason for not having sex: -na

v633g

reason for not having sex: -na

v634

husband knows that responde-na

v701

partner's education level

v702

highest year of education

v704

partner's occupation

v705

partner's occupation

v714

respondent currently working

v714a

have a job from which respondent was absent

v715

husbands education-single yrs

v716

respondent's occupation

v717

respondent's occupation

v719

work for family, others, se-na

v721

works at home or away -na

v729

partner's educational attainm.

v730

partners age

v731

worked in last 12 months

v732

employment all year/seasona-na

v739

who decides how to spend mo-na

v740

type of land where resp. wo-na

v741

type of earnings for work

v743a

final say on own health car-na

v743b

final say on making large h-na

v743c

final say on making househo-na

v743d

final say on visits to fami-na

v743e

final say on food to be coo-na

v743f

final say on deciding what -na

v744a

wife beating justified if s-na

v744b

wife beating justified if s-na

v744c

wife beating justified if s-na

v744d

wife beating justified if s-na

v744e

wife beating justified if s-na

v746

earns more than partner -na

bidx

birth column number

bord

birth order number

b0

child is twin

b1

month of birth

b2

year of birth

b3

date of birth (cmc)

b4

sex of child

b5

child is alive

b6

age at death

b7

age at death (months-imputed)

b8

current age of child

b9

child lives with whom

b10

completeness of information

b11

preceding birth interval

b12

succeeding birth interval

b13

flag for age at death

b15

live birth between births

b16

child's line number in household

m1

tetanus injections bef. birth

m1a

tetanus injections before pregnancy

m1b

month last tetanus injection before pregnancy

m1c

year last tetanus injection before pregnancy

m1d

years ago received last tetanus injection before pregnancy

m1e

last tetanus injection before pregnancy (cmc)

m2a

prenatal: doctor

m2b

prenatal: nurse/midwife/lady health visitor

m2c

prenatal: auxiliary midwife-na

m2d

prenatal: cs health profess-na

m2e

prenatal: cs health profess-na

m2f

prenatal: trained birth att-na

m2g

prenatal: trad.birth attendant/dai

m2h

prenatal: relative -na

m2i

prenatal: lady health worker

m2j

prenatal: homeopath

m2k

prenatal: other resp (uncoded)

m2l

prenatal: hakim

m2m

prenatal: dispenser/compounder

m2n

prenatal: no one

m3a

assistance: doctor

m3b

assistance: nurse/midwife/lady health visitor

m3c

assistance: auxiliary midwi-na

m3d

assistance: cs health profe-na

m3e

assistance: cs health profe-na

m3f

assistance: trained birth a-na

m3g

assistance: trad.birth attend/dai

m3h

assistance: relative, friend

m3i

assistance: lady health worker

m3j

assistance: homeopath

m3k

assistance: other resp (uncod)

m3l

assistance: hakim

m3m

assistance: cs other -na

m3n

assistance: no one

m4

duration of breastfeeding

m5

months of breastfeeding

m6

duration of amenorrhea

m7

months of amenorrhea

m8

duration of abstinence

m9

months of abstinence

m10

time wanted pregnancy

m11

time would have waited

m13

timing of 1st antenatal check

m14

antenatal visits for pregnancy

m15

place of delivery

m17

delivery by caesarian section

m18

size of child at birth

m19

birth weight (kilos - 3 dec.)

m19a

weight at birth recall

m27

flag for breastfeeding

m28

flag for amenorrhea

m29

flag for abstinence

m34

when child put to breast

m35

times breastfed during night

m36

times breastfed during day

m38

drank from bottle with nipple

m39

times ate solid, semi-solid-na

m42a

during pregnancy - weighed

m42b

during pregnancy - height m-na

m42c

during pregnancy - blood pressure taken

m42d

during pregnancy - urine sample taken

m42e

during pregnancy - blood sample taken

m43

told about pregnancy complications

m44

told where to go for pregnancy complications

m45

during pregancy, given or bought iron tablets/syrup

m46

days tablets or syrup taken

m47

during pregnancy, had difficulty with daylight vision

m48

during pregnancy, had difficulty with night blindness

m49a

during pregnancy - took fan-na

m49b

during pregnancy - took chl-na

m49c

during pregnancy - took unk-na

m49d

during pregnancy - took cou-na

m49e

during pregnancy - took cou-na

m49f

during pregnancy - took cou-na

m49g

during pregnancy - took cou-na

m49x

during pregnancy - took oth-na

m49z

during pregnancy - took no -na

m51a

respondent checkup after deliver timing

m54

received vitamin a dose in first 2 months after delivery

m55a

first 3 days, given milk other than breast milk

m55b

first 3 days, given plain water

m55c

first 3 days, given sugar/glucose water

m55d

first 3 days, given gripe w-na

m55e

first 3 days, given sugar/s-na

m55f

first 3 days, given fruit juice

m55g

first 3 days, given infant formula

m55h

first 3 days, given tea/infusions

m55i

first 3 days, given honey -na

m55j

first 3 days, given ghee, butter

m55k

first 3 days, given ghutee

m55l

first 3 days, given cs herbs/traditional medicines

m55m

first 3 days, given country-na

m55n

first 3 days, given country-na

m55x

first 3 days, given other

m55z

first 3 days, given nothing

m57a

antenatal care: your home

m57b

antenatal care: other home

m57c

antenatal care: cs home -na

m57d

antenatal care: cs home -na

m57e

antenatal care: govt. hospital

m57f

antenatal care: govt. health center

m57g

antenatal care: govt. healt-na

m57h

antenatal care: public mobi-na

m57i

antenatal care: cs public bhu/fwc

m57j

antenatal care: cs public -na

m57k

antenatal care: cs public -na

m57l

antenatal care: other public

m57m

antenatal care: pvt. hospital/clinic

m57n

antenatal care: pvt. mobile-na

m57o

antenatal care: cs pvt. private doctor

m57p

antenatal care: cs pvt.homeopath

m57q

antenatal care: cs pvt.dispenser/compounder

m57r

antenatal care: other private

m57s

antenatal care: cs other hakim

m57t

antenatal care: cs other -na

m57u

antenatal care: cs other -na

m57v

antenatal care: cs other -na

m57x

antenatal care: other

m60

drugs for intestinal parasi-na

m61

time spent at place of delivery

m62

anyone checked respondent health before discharge

m63

how long before discharging resondent health check took plac

m64

who checked respondent health before discharge

m65a

reason didn't deliver at health facility: cost too much

m65b

reason didn't deliver at health facility: facility not open

m65c

reason didn't deliver at health facility: too far/no transpo

m65d

reason didn't deliver at health facility: don't trust facili

m65e

reason didn't deliver at health facility: no female provider

m65f

reason didn't deliver at health facility: husband/family did

m65g

reason didn't deliver at health facility: not necessary

m65h

reason didn't deliver at health facility: not customary

m65i

reason didn't deliver at health facility: cs no time/baby ca

m65j

reason didn't deliver at he-na

m65k

reason didn't deliver at he-na

m65l

reason didn't deliver at he-na

m65x

reason didn't deliver at health facility: other

m66

after discharge/delivery at home anyone checked respondent h

m67

how long after discharge/delivery at home respondent health

m68

who checked respondent health after discharge/delivery at ho

m69

where respondent check after discharge/delivery at home took

m70

baby postnatal check within 2 months

m71

how long after delivery postnatal check took place

m72

who performed postnatal checkup

m73

where was the baby checked for the first time

hidx

index to birth history

h1

has health card

h2

received bcg

h2d

bcg day

h2m

bcg month

h2y

bcg year

h3

received dpt 1

h3d

dpt 1 day

h3m

dpt 1 month

h3y

dpt 1 year

h4

received polio 1

h4d

polio 1 day

h4m

polio 1 month

h4y

polio 1 year

h5

received dpt 2

h5d

dpt 2 day

h5m

dpt 2 month

h5y

dpt 2 year

h6

received polio 2

h6d

polio 2 day

h6m

polio 2 month

h6y

polio 2 year

h7

received dpt 3

h7d

dpt 3 day

h7m

dpt 3 month

h7y

dpt 3 year

h8

received polio 3

h8d

polio 3 day

h8m

polio 3 month

h8y

polio 3 year

h9

received measles

h9d

measles day

h9m

measles month

h9y

measles year

h0

received polio 0

h0d

polio 0 day

h0m

polio 0 month

h0y

polio 0 year

h10

ever had vaccination

h11

had diarrhea recently

h11b

blood in the stools

h12a

diarrhea: government hosp.

h12b

diarrhea: govt health center rhc/mch

h12c

diarrhea: govt health post -na

h12d

diarrhea: mobile clinic -na

h12e

diarrhea: comm.health worker

h12f

diarrhea: cs public sector bhu/fwc

h12g

diarrhea: cs public sector -na

h12h

diarrhea: cs public sector -na

h12i

diarrhea: other public sector

h12j

diarrhea: private hosp/clin.

h12k

diarrhea: private pharmacy

h12l

diarrhea: private doctor

h12m

diarrhea: private mobile cl-na

h12n

diarrhea: comm.health worke-na

h12o

diarrhea: cs med.priv sector homeopath

h12p

diarrhea: cs med.priv sector dispenser/compounder

h12q

diarrhea: cs med.priv secto-na

h12r

diarrhea: other med.priv sect.

h12s

diarrhea: shop

h12t

diarrhea: traditional pract-na

h12u

diarrhea: cs oth.priv sector hakim

h12v

diarrhea: cs oth.priv sector dai/tba

h12w

diarrhea: cs oth.priv secto-na

h12x

diarrhea: other

h12y

diarrhea: no treatment

h12z

diarrhea: medical treatment

h13

given oral rehydration

h13b

given a pre-packaged ors li-na

h14

given recommend. home solution

h15

given antibiotic pills or s-na

h15a

given antimotility -na

h15b

given an antibiotic injecti-na

h15c

given an intravenous (iv)

h15d

given home remedy, herbal med.

h15e

given zinc -na

h15f

given other (not antibiotic)-na

h15g

given unknown pill or syrup

h15h

given not antibiotic inject-na

h15i

given unknown injection

h15j

given cs -na

h15k

given cs -na

h15l

given cs -na

h15m

given cs -na

h20

given other treatment

h21a

given no treatment

h21

received any treatment

h22

had fever in last two weeks

h31

had cough in last two weeks

h31b

short, rapid breaths

h31c

problem in the chest or blocked or running nose

h31d

had fever/cough in last 2 weeks: amount offered to drink

h31e

had fever/cough in last 2 weeks: amount offered to eat

h32a

fever/cough: government hosp.

h32b

fever/cough: govt health cntr rhc/mch

h32c

fever/cough: govt health po-na

h32d

fever/cough: mobile clinic -na

h32e

fever/cough: comm.health wrkr (lady health worker)

h32f

fever/cough: cs public sector bhu/fwc

h32g

fever/cough: cs public sect-na

h32h

fever/cough: cs public sect-na

h32i

fever/cough: oth public sector

h32j

fever/cough: private hosp/clin

h32k

fever/cough: private pharmacy

h32l

fever/cough: private doctor

h32m

fever/cough: private mobile-na

h32n

fever/cough: comm.health wr-na

h32o

fever/cough: cs med.priv sect homeopath

h32p

fever/cough: cs med.priv sect dispenser/compounder

h32q

fever/cough: cs med.priv se-na

h32r

fever/cough: oth med.priv sect

h32s

fever/cough: shop

h32t

fever/cough: traditional pr-na

h32u

fever/cough: cs oth.priv sect hakim

h32v

fever/cough: cs oth.priv sect dai/tba

h32w

fever/cough: cs oth.priv se-na

h32x

fever/cough: other

h32y

fever/cough: no treatment

h32z

fever/cough: medical treatment

h33

received vitamin a1 (most r)-na

h33d

vitamin a1 day -na

h33m

vitamin a1 month -na

h33y

vitamin a1 year -na

h34

vitamin a in last 6 months

h35

any vaccinations in last 2 years part of campaign

h36a

vaccinated during campaign -na

h36b

vaccinated during campaign -na

h36c

vaccinated during campaign -na

h36d

vaccinated during campaign -na

h36e

vaccinated during campaign -na

h36f

vaccinated during campaign -na

h37a

fansidar taken for fever/cough

h37b

chloroquine taken for fever/cough

h37c

amodiaquine taken for fever-na

h37d

quinine taken for fever/cough

h37e

combination with artemisini-na

h37f

cs antimalarial taken for f-na

h37g

cs antimalarial taken for f-na

h37h

other antimalarial taken for fever/cough

h37i

pills/syrup taken for fever/cough

h37j

injection taken for fever/cough

h37k

aspirin taken for fever/cough

h37l

acetaminophen taken for fever/cough

h37m

ibuprofen taken for fever/c-na

h37n

cs brufen taken for fever/cough

h37o

cs pill/syrup for cough for fever/cough

h37p

cs other for fever/cough -na

h37x

other taken for fever/convulsion

h37y

nothing taken for fever/convulsion

h37z

don't know if or what was taken for fever/convulsion

h38

had diarrhea in last 2 weeks: amount offered to drink

h39

had diarrhea in last 2 weeks: amount offered to eat

h40

received vitamin a2 (2nd mo)-na

h40d

vitamin a2 day -na

h40m

vitamin a2 month -na

h40y

vitamin a2 year -na

h41a

received vitamin a after mo-na

h41b

ever received vitamin a dose

h42

taking iron pills, sprinkle-na

h43

drugs for intestinal parasi-na

h44a

place first sought treatment for diarrhea

h44b

days after diarrhea sought advice or treatment

h44c

still has diarrhea

h45

times zinc was given -na

h46a

place first sought treatment for fever

h46b

days after fever sought advice or treatment

hwidx

index to birth history

hw1

age in months -na

hw2

weight in kilograms (1 dec.)-na

hw3

height in centimeters (1 de)-na

hw4

ht/a percentile -na

hw5

ht/a standard deviations -na

hw6

ht/a percent of ref. median-na

hw7

wt/a percentile -na

hw8

wt/a standard deviations -na

hw9

wt/a percent of ref. median-na

hw10

wt/ht percentile -na

hw11

wt/ht standard deviations -na

hw12

wt/ht percent of ref. media-na

hw13

reason not measured

hw15

height: lying or standing -na

hw16

day of birth of child -na

hw17

date measured (day) -na

hw18

date measured (month) -na

hw19

date measured (year) -na

hw51

line no. of parent/caretake-na

hw52

read consent statement -na

hw53

hemoglobin level (g/dl - 1) -na

hw55

result of measuring (hemogl)-na

hw56

hemoglobin level adjusted b-na

hw57

anemia level -na

hw58

agrees to referral -na

hw70

ht/a standard deviations (a)-na

hw71

wt/a standard deviations (a)-na

hw72

wt/ht standard deviations ()-na

hw73

bmi standard deviations (ac)-na

idxml

index to birth history

ml0

type of bednet(s) child slept under last night

ml1

times took fansidar during -na

ml2

type of visit at source for-na

ml11

child has fever now

ml12

child has had convulsions i-na

ml13a

fansidar taken for fever/cough

ml13b

chloroquine taken for fever/cough

ml13c

amodiaquine taken for fever-na

ml13d

quinine taken for fever/cough

ml13e

combination with artemisini-na

ml13f

cs antimalarial taken for f-na

ml13g

cs antimalarial taken for f-na

ml13h

other antimalarial taken for fever/cough

ml13i

pills/syrup taken for fever/cough

ml13j

injection taken for fever/cough

ml13k

aspirin taken for fever/cough

ml13l

acetaminophen taken for fever/cough

ml13m

ibuprofen taken for fever/c-na

ml13n

cs brufen other taken for fever/cough

ml13o

cs pill/syrup for cough other for fever/cough

ml13p

cs other for fever/cough -na

ml13x

other taken for fever/convulsion

ml13y

nothing taken for fever/convulsion

ml13z

don't know if or what was taken for fever/convulsion

ml14a

injection for fever/convuls-na

ml14b

suppository for fever/convu-na

ml14y

no suppository or injection-na

ml14z

don't know if suppository o-na

ml15a

when started fansidar

ml15b

days child took fansidar

ml15c

first source for fansidar -na

ml16a

when started chloroquine

ml16b

days child took chloroquine

ml16c

first source for chloroquin-na

ml17a

when started amodiaquine -na

ml17b

days child took amodiaquine-na

ml17c

first source for amodiaquin-na

ml18a

when started quinine

ml18b

days child took quinine

ml18c

first source for quinine -na

ml19a

for fever/conv: consulted t-na

ml19b

for fever/conv: gave tepid -na

ml19c

for fever/conv: gave herbs -na

ml19d

for fever/conv: cs -na

ml19e

for fever/conv: cs -na

ml19f

for fever/conv: cs -na

ml19x

for fever/conv: other -na

ml19y

for fever/conv: gave nothin-na

ml19z

for fever/conv: don't know -na

ml20a

when started combination wi-na

ml20b

days child took combination-na

ml20c

first source for combinatio-na

ml21a

when started cs antimalaria-na

ml21b

days child took cs antimala-na

ml21c

first source for cs antimal-na

ml22a

when started cs antimalaria-na

ml22b

days child took cs antimala-na

ml22c

first source for cs antimal-na

ml23a

when started other anti-malarial

ml23b

days child took other anti-malarial

ml23c

first source for other anti-na

ml24c

first source for antibiotic-na

sdist

district

slangq

language of questionaire

slangi

language of interview

slangw

language woman speaks

awfactu2

awfactor (major urban, other urban, rural)

awfacte2

awfactor education (educ2- 0,princ,pricmp,secyinc,seccomp,hi)

awfacte3

awfactor education (educ2- 0,prim,middle,secy,high)

s107fh

room cooler, air conditioner owns

s107gh

washing machine owns

s107hh

water pump owns

s107ih

bed owns

s107jh

chairs owns

s107kh

almirah/cabinet owns

s107lh

clock owns

s107mh

sofa owns

s107nh

sewing machine owns

s107oh

camera owns

s107ph

personal computer owns

s113h

status of house

s128h

currently have any medicines for treating malaria in the hou

s108

type of blood relationship with husband

s113

educational level

s233a

number of terminated pregnancy since jan 2001 (miscarriages)

s233b

number of terminated pregnancy since jan 2001 (abortions)

s233c

number of terminated pregnancy since jan 2001 (stillbirths)

s236a

complications/problems have heard about it (1)

s236b

complications/problems have heard about it (2)

s236c

complications/problems have heard about it (3)

s236d

complications/problems have heard about it (4)

s312

social marketing brand

s314

name of injection using

s315

number of weeks one injection is effective

s323e

source of fp for non-user (lady health worker)

s323f

source of fp for non-user (lady health visitor)

s326a

type of help received:information

s326b

type of help received:contraceptive supplies

s326c

type of help received:referred to health/fp facility

s326d

type of help received:treatment of side effects

s326x

type of help received:other

s326y

type of help received:no help

s329a

fp message heard:limiting the family

s329b

fp message heard:higher age at marriage

s329c

fp message heard:spacing of children

s329d

fp message heard:use of contraceptives

s329e

fp message heard:welfare of family

s329f

fp message heard:maternal/child health

s329g

fp message heard:fewer children means prosperous life

s329h

fp message heard:more children means poverty and starvation

s329i

fp message heard:importance of breastfeeding

s329x

fp message heard:other

s329z

fp message heard:don't know/don't remember

s330

fp message heard was effective

smetneed

met need for contraceptives

sunmetnd

unmet need for contraceptives

stotdemd

total demand

s617a

presence of others section 6: children < 10

s617b

presence of others section 6: husband

s617c

presence of others section 6: mother-in-law

s617d

presence of others section 6: other male(s)

s617e

presence of others section 6: other female(s)

s704

husband's educational level

s710

would like to work if could find suitable job

s713

ever worked before first married

s714

ever worked after first married

s715

work status (current, before marriage, after marriage)

s811

ever talked with husband about ways to prevent getting hiv/a

s815a

sign of sti infection:wound without pain

s815b

sign of sti infection:wound with pain

s815c

sign of sti infection:wound, pain, with many pimples

s815d

sign of sti infection:pus-like discharge

s815e

sign of sti infection:dark pus-like discharge

s815f

sign of sti infection:sour-milk-like discharge

s815g

sign of sti infection:sponge-like discharge

s815h

sign of sti infection:discharge with bad odor/dirty water

s815i

sign of sti infection:itching/backache

s815j

sign of sti infection:hepatitis

s815k

sign of sti infection:tb

s815x

sign of sti infection:other

s904

current duration of tb treatment in months

s905

been told by doctor/nurse that respondent has tb

s907

ever experienced urine/stool problem (fistula)

s908

still have urine problem

s909

how fistula problem began

s910

happened to baby

s915a

can protect from hepatitis/hiv/aids by using new syringe/nee

s915b

can protect from hepatitis/hiv/aids by testing blood before

idx92

index to birth history

s221

place where died

idx94

index to birth history

s409

had a problem or routine checkup in the first antenatal visi

s411

fee paid for first antenatal visit

s413e

antenatal care: ultrasound

s414

knows blood group

s417a

no antenatal checkup reason:not necessary

s417b

no antenatal checkup reason:costs too much

s417c

no antenatal checkup reason:too far

s417d

no antenatal checkup reason:no transport

s417e

no antenatal checkup reason:no one to go with

s417f

no antenatal checkup reason:service not good

s417g

no antenatal checkup reason:no time to go

s417h

no antenatal checkup reason:did not know where to go

s417i

no antenatal checkup reason:did not want to see a male docto

s417j

no antenatal checkup reason:long waiting time

s417k

no antenatal checkup reason:not allowed to go

s417x

no antenatal checkup reason:other

s418

was told how to have a safe delivery

s428

given/took calcium tablets

s429

number of days took calcium tablets

s432

sufferred from malaria during pregnancy

s433

received treatment for the malaria during pregnancy

s434

place where recieved treatment for malaria

s435a

problems during pregnancy: serious headaches

s435b

problems during pregnancy: blurred vision

s435c

probelms during pregnancy: swelling of hands

s435d

problems during pregnancy: swelling of face

s435e

problems during pregnancy: vaginal bleeding/spotting

s435f

problems during pregnancy: fits or convulsion

s435g

problems during pregnancy: epigastric pains

s437

problems so severe that were afraid might die

s438a

seek advice/treatment for problems from: doctor

s438b

seek advice/treatment for problems from:nurse/midwife/lhv

s438c

seek advice/treatment for problems from:dai-tba

s438d

seek advice/treatment for problems from:lady health worker

s438e

seek advice/treatment for problems from:homeopath

s438f

seek advice/treatment for problems from:hakim

s438g

seek advice/treatment for problems from:dispenser/compounder

s438x

seek advice/treatment for problems from:other

s438y

seek advice/treatment for problems from:no one

s439a

place for treatment of problem:respondent's home

s439b

place for treatment of problem:other home

s439c

place for treatment of problem:govt. hospital

s439d

place for treatment of problem:rhc/mch

s439e

place for treatment of problem:bhu/fwc

s439f

place for treatment of problem:other public

s439h

place for treatment of problem:pvt. hospital/clinic

s439i

place for treatment of problem:pvt. doctor

s439j

place for treatment of problem:homeopath

s439k

place for treatment of problem:dispenser/compounder

s439l

place for treatment of problem:hakim

s439m

place for treatment of problem:other private medical

s439x

place for treatment of problem:other

s440

interval between problem and seeking advice

s441a

reason no treatment for problem sought:not necessary

s441b

reason no treatment for problem sought:costs too much

s441c

reason no treatment for problem sought:too far

s441d

reason no treatment for problem sought:no transport

s441e

reason no treatment for problem sought:no one to go with

s441f

reason no treatment for problem sought:service not good

s441g

reason no treatment for problem sought:no time to go

s441h

reason no treatment for problem sought:did not know where to

s441i

reason no treatment for problem sought:did not want to see m

s441j

reason no treatment for problem sought:long waiting time

s441k

reason no treatment for problem sought:not allowed to go

s441x

reason no treatment for problem sought:other

s442

ever discussed where would deliver during pregnancy

s443

set aside any money in case of emergency during pregnancy

s448

given an injection to induce labor

s450a

reason for delivering at health facility (1)

s450b

reason for delivering at health facility (2)

s450c

reason for delivering at health facility (3)

s453

fee paid for the delivery (inclusive of everything)

s459

total fee paid for the delivery

s460

safe delivery kit used

s461

used to tie the umbilical cord

s462

used to cut the umbilical cord

s463

instrument boiled before using it

s472

duration before baby was bathed

s473a

during delivery or after birth, experienced severe headaches

s473b

during delivery or after birth, experienced blurred vision

s473c

during delivery or after birth, experienced swelling on hand

s473d

during delivery or after birth, experienced swelling on face

s473e

during delivery or after birth, experienced high fever

s473f

during delivery or after birth, experienced fits or convulsi

s473g

during delivery or after birth, experienced labour for more

s473h

during delivery or after birth, experienced baby's feet came

s473i

during delivery or after birth, experienced placenta came fi

s473j

during delivery or after birth, experienced continous dribbl

s473k

during delivery or after birth, experienced bad-smelling vag

s473l

during delivery or after birth, experienced inability to con

s473m

during delivery or after birth, experienced heavy vaginal bl

s473ma

if experienced heavy vaginal bleeding, when experienced

s475

these problems were so severe that afraid might die

s476a

seek advice/treatment for the problem from:doctor

s476b

seek advice/treatment for the problem from:nurse/midwife/lhv

s476c

seek advice/treatment for the problem from:dai/tba

s476d

seek advice/treatment for the problem from:lady health worke

s476e

seek advice/treatment for the problem from:homeopath

s476f

seek advice/treatment for the problem from:hakim

s476x

seek advice/treatment for the problem from:other

s476y

seek advice/treatment for the problem from:no one

s477a

place sought advice/treatment:respondent's home

s477b

place sought advice/treatment:other home

s477c

place sought advice/treatment:govt. hospital

s477d

place sought advice/treatment:rhc/mch

s477e

place sought advice/treatment:bhu/fwc

s477f

place sought advice/treatment:other public

s477h

place sought advice/treatment:pvt. hospital/clinic

s477i

place sought advice/treatment:pvt. doctor

s477j

place sought advice/treatment:homeopath

s477k

place sought advice/treatment:dispenser/compounder

s477l

place sought advice/treatment:hakim

s477m

place sought advice/treatment:other private medical

s477x

place sought advice/treatment:other

s478

duration between start of problem and seeking advice/treatme

s479a

reason did not seek advice/treatment:not necessary

s479b

reason did not seek advice/treatment:costs too much

s479c

reason did not seek advice/treatment:facility too far away

s479d

reason did not seek advice/treatment:no transport

s479e

reason did not seek advice/treatment:no one to go with

s479f

reason did not seek advice/treatment:service not good

s479g

reason did not seek advice/treatment:no time to go

s479h

reason did not seek advice/treatment:did not know where to g

s479i

reason did not seek advice/treatment:did not want to see mal

s479j

reason did not seek advice/treatment:long waiting time

s479k

reason did not seek advice/treatment:not allowed to go

s479x

reason did not seek advice/treatment:other

s489

gave colostrum to child

s498a

gave child water yesterday or last night

s498b

gave child milk yesterday or last night

s498c

gave child juice/soda yesterday or last night

s498d

gave child food yesterday or last night

idx95

index to birth history

shbv1

received hbv 1

shbv1d

hbv 1 day

shbv1m

hbv 1 month

shbv1y

hbv 1 year

shbv2

received hbv 2

shbv2d

hbv 2 day

shbv2m

hbv 2 month

shbv2y

hbv 2 year

shbv3

received hbv 3

shbv3d

hbv 3 day

shbv3m

hbv 3 month

shbv3y

hbv 3 year

s512

months ago child received a vitamin a dose

s516

number of times stools passed in past 24 hours

s538

medicines for fever/cough prescribed by medical, health or p

References

National Institute of Population Studies (NIPS), Pakistan (http://nips.org.pk/Home.htm).

Examples

data("PKKR52")
PKKR52

#> # A tibble: 9,177 x 1,139
#>    caseid  midx v000   v001  v002  v003  v004   v005  v006  v007  v008  v009
#>    <chr>  <int> <chr> <int> <int> <int> <int>  <int> <int> <int> <int> <int>
#>  1 "    …     1 PK5       1    37     2     1 1.20e6     2  2007  1286    10
#>  2 "    …     1 PK5       1    47     2     1 1.20e6     2  2007  1286     3
#>  3 "    …     1 PK5       1    67     4     1 1.20e6     2  2007  1286     9
#>  4 "    …     1 PK5       1    67     6     1 1.20e6     2  2007  1286    11
#>  5 "    …     1 PK5       2     7     5     2 1.25e6     2  2007  1286     2
#>  6 "    …     1 PK5       2    17     2     2 1.25e6     2  2007  1286    10
#>  7 "    …     1 PK5       2    57     2     2 1.25e6     2  2007  1286    10
#>  8 "    …     1 PK5       2    77     5     2 1.25e6     2  2007  1286     8
#>  9 "    …     1 PK5       3     7     2     3 1.65e6     2  2007  1286    12
#> 10 "    …     2 PK5       3     7     2     3 1.65e6     2  2007  1286    12
#> # … with 9,167 more rows, and 1,127 more variables: v010 <int>, v011 <int>,
#> #   v012 <int>, v013 <fct>, v014 <fct>, v015 <fct>, v016 <int>, v017 <int>,
#> #   v018 <fct>, v019 <fct>, v019a <int>, v020 <fct>, v021 <int>, v022 <int>,
#> #   v023 <fct>, v024 <fct>, v025 <fct>, v026 <fct>, v027 <int>, v028 <int>,
#> #   v029 <int>, v030 <int>, v031 <int>, v032 <int>, v033 <int>, v034 <fct>,
#> #   v040 <int>, v042 <fct>, v043 <fct>, v044 <fct>, v101 <fct>, v102 <fct>,
#> #   v103 <fct>, v104 <fct>, v105 <fct>, v106 <fct>, v107 <int>, v113 <int>,
#> #   v115 <int>, v116 <int>, v119 <int>, v120 <int>, v121 <int>, v122 <int>,
#> #   v123 <int>, v124 <int>, v125 <int>, v127 <int>, v128 <int>, v129 <int>,
#> #   v130 <fct>, v131 <int>, v133 <int>, v134 <fct>, v135 <fct>, v136 <int>,
#> #   v137 <int>, v138 <int>, v139 <fct>, v140 <fct>, v141 <fct>, v149 <fct>,
#> #   v150 <int>, v151 <fct>, v152 <int>, v153 <int>, awfactt <int>,
#> #   awfactu <int>, awfactr <int>, awfacte <int>, awfactw <int>, v155 <int>,
#> #   v156 <fct>, v157 <fct>, v158 <fct>, v159 <fct>, v160 <int>, v161 <int>,
#> #   v166 <fct>, v167 <fct>, v168 <fct>, v190 <fct>, v191 <int>, ml101 <fct>,
#> #   v201 <int>, v202 <int>, v203 <int>, v204 <int>, v205 <int>, v206 <int>,
#> #   v207 <int>, v208 <int>, v209 <int>, v210 <int>, v211 <int>, v212 <int>,
#> #   v213 <fct>, v214 <int>, v215 <int>, v216 <fct>, …