Pakistan Demographic and Health Survey 2012-13 Children's Recode Data

PKKR61 data of Children's Recode from Pakistan Demographic and Health Survey 2012-13.

caseid

Case Identification

midx

Index to Birth History

v000

Country code and phase

v001

Cluster number

v002

Household number

v003

Respondent's line number

v004

Ultimate area unit

v005

Women's individual sample weight (6 decimals)

v006

Month of interview

v007

Year of interview

v008

Date of interview (CMC)

v009

Respondent's month of birth

v010

Respondent's year of birth

v011

Date of birth (CMC)

v012

Respondent's current age

v013

Age in 5-year groups

v014

Completeness of age information

v015

Result of individual interview

v016

Day of interview

v017

CMC start of calendar

v018

Row of month of interview

v019

Length of calendar

v019a

Number of calendar columns

v020

Ever-married sample

v021

Primary sampling unit

v022

Sample strata for sampling errors

v023

Stratification used in sample design

v024

Region

v025

Type of place of residence

v026

De facto place of residence

v027

Number of visits

v028

Interviewer identification

v029

Keyer identification

v030

Field supervisor

v031

Field editor

v032

NA - Office editor

v034

Line number of husband

v040

NA - Cluster altitude in meters

v042

NA - Household selected for hemoglobin

v044

Selected for Domestic Violence module

v101

Region

v102

Type of place of residence

v103

NA - Childhood place of residence

v104

NA - Years lived in place of residence

v105

NA - Type of place of previous residence

v106

Highest educational level

v107

Highest year of education

v113

Source of drinking water

v115

Time to get to water source

v116

Type of toilet facility

v119

Household has: electricity

v120

Household has: radio

v121

Household has: television

v122

Household has: refrigerator

v123

Household has: bicycle

v124

Household has: motorcycle/scooter

v125

Household has: car/truck

v127

Main floor material

v128

Main wall material

v129

Main roof material

v130

NA - Religion

v131

Ethnicity

v133

Education in single years

v134

De facto place of residence

v135

Usual resident or visitor

v136

Number of household members (listed)

v137

Number of children 5 and under in household (de jure)

v138

Number of eligible women in household (de facto)

v139

De jure region of residence

v140

De jure type of place of residence

v141

De jure place of residence

v149

Educational attainment

v150

Relationship to household head

v151

Sex of household head

v152

Age of household head

v153

Household has: telephone (land-line)

awfactt

All woman factor - total

awfactu

All woman factor - urban/rural

awfactr

All woman factor - regional

awfacte

All woman factor - educational

awfactw

All woman factor - wealth index

v155

Literacy

v156

NA - Ever participated in a literacy program outside of primary

v157

Frequency of reading newspaper or magazine

v158

Frequency of listening to radio

v159

Frequency of watching television

v160

Toilet facilities shared with other households

v161

Type of cooking fuel

v166

NA - Result of salt test for iodine

v167

NA - Number of trips in last 12 months

v168

NA - Away for more than one month in last 12 months

v190

Wealth index

v191

Wealth index factor score (5 decimals)

ml101

NA - Type of mosquito bed net(s) slept under last night

v201

Total children ever born

v202

Sons at home

v203

Daughters at home

v204

Sons elsewhere

v205

Daughters elsewhere

v206

Sons who have died

v207

Daughters who have died

v208

Births in last five years

v209

Births in past year

v210

Births in month of interview

v211

Date of first birth (CMC)

v212

Age of respondent at 1st birth

v213

Currently pregnant

v214

Duration of current pregnancy

v215

Time since last menstrual period

v216

Menstruated in last six weeks

v217

Knowledge of ovulatory cycle

v218

Number of living children

v219

Living children + current pregnancy

v220

Living children + current pregnancy (grouped)

v221

Marriage to first birth interval (months)

v222

Last birth to interview (months)

v223

Completeness of current pregnancy information

v224

Entries in birth history

v225

Current pregnancy wanted

v226

Time since last period (comp) (months)

v227

Flag for last period

v228

Ever had a terminated pregnancy

v229

Month pregnancy ended

v230

Year pregnancy ended

v231

CMC pregnancy ended

v232

Completeness of last termination information

v233

Months when pregnancy ended

v234

Other such pregnancies

v235

Index last child prior to maternity-health (calendar)

v237

Birth between last and interview

v238

Births in last three years

v239

Pregnancies terminated before calendar beginning

v240

Month of last termination prior to calendar

v241

Year of last termination prior to calendar

v242

CMC termination ended prior to calendar

v243

Completeness of last termination date prior to calendar

v310

NA - Living children at first use

v311

NA - Children at first use (grouped)

v312

Current contraceptive method

v313

Current use by method type

v315

Month of start of use of method

v316

Year of start of use of method

v317

Date of start of use of method (CMC)

v318

Completeness of information

v319

Years since sterilization

v320

Age at sterilization

v321

Marital duration at sterilization

v322

Parity at sterilization

v323

Brand of pill used

v323a

Brand of condom used

v325a

NA - Cost of current method (in CURRENCY)

v326

Last source for current users

v327

Last source for users by type

v337

Months of use of current method

v359

Last method discontinued in last 5 years

v360

Reason of last discontinuation

v361

Pattern of use

v362

Intention to use

v363

NA - Preferred future method

v364

Contraceptive use and intention

v367

Wanted last child

v372

NA - Shown pill package

v372a

NA - Shown condom package

v375a

NA - Main reason not using a method

v376

NA - Main reason not to use a method

v376a

NA - Would ever use method if married

v379

NA - Source known for any method

v380

NA - Source known for any method

v384a

Heard family planning on radio last few months

v384b

Heard family planning on TV last few months

v384c

Heard family planning in newspaper/magazine last few months

v393

Visited by family planning worker last 12 months

v394

Visited health facility last 12 months

v395

At health facility, told of family planning

v3a00a

Source of family planning for non-users: government hospital

v3a00b

Source of family planning for non-users: rural health centre(RHC)

v3a00c

Source of family planning for non-users: family welfare centre

v3a00d

Source of family planning for non-users: Mother child health centre(MCH)

v3a00e

Source of family planning for non-users: mobile service camp

v3a00f

Source of family planning for non-users: lady health worker

v3a00g

Source of family planning for non-users: Lady health visitor

v3a00h

Source of family planning for non-users: basic health unit

v3a00i

Source of family planning for non-users: other public

v3a00j

Source of family planning for non-users: private hospital/clinic

v3a00k

Source of family planning for non-users: pharmacy

v3a00l

Source of family planning for non-users: private doctor

v3a00m

Source of family planning for non-users: homeopath

v3a00n

Source of family planning for non-users: dispenser/compounder

v3a00o

Source of family planning for non-users: other private sector

v3a00p

NA - Source of family planning for non-users: CS private sector

v3a00q

NA - Source of family planning for non-users: CS private sector

v3a00r

NA - Source of family planning for non-users: CS private sector

v3a00s

Source of family planning for non-users: shop

v3a00t

Source of family planning for non-users: firend/relative

v3a00u

Source of family planning for non-users: hakim

v3a00v

Source of family planning for non-users: DAI trad birth attendant

v3a00w

NA - Source of family planning for non-users: CS other

v3a00x

Source of family planning for non-users: other

v3a00y

Source of family planning for non-users: no source

v3a00z

Source of family planning for non-users: any source

v3a01

NA - Told sterilization would mean no more children

v3a02

Told about side effects

v3a03

Told about side effects by health or family planning worker

v3a04

Told how to deal with side effects

v3a05

Told about other family planning methods

v3a06

Told about other family planning methods by health or family planning worker

v3a07

First source for current method

v3a08a

NA - Reason not using: not married

v3a08b

Reason not using: not having sex

v3a08c

Reason not using: infrequent sex

v3a08d

Reason not using: menopausal/hysterectomy

v3a08e

Reason not using: subfecund/infecund

v3a08f

Reason not using: postpartum amenorrheic

v3a08g

Reason not using: breastfeeding

v3a08h

Reason not using: fatalistic

v3a08i

Reason not using: respondent opposed

v3a08j

Reason not using: husband/partner opposed

v3a08k

Reason not using: others opposed

v3a08l

Reason not using: religious prohibition

v3a08m

Reason not using: knows no method

v3a08n

Reason not using: knows no source

v3a08o

NA - Reason not using: health concerns

v3a08p

Reason not using: fear of side effects/health concerns

v3a08q

Reason not using: lack of access/too far

v3a08r

Reason not using: costs too much

v3a08s

Reason not using: inconvenient to use

v3a08t

Reason not using: interferes with body?s processes

v3a08u

Reason not using: preferred method not available

v3a08v

Reason not using: no method available

v3a08w

NA - Reason not using: CS

v3a08aa

NA - Reason not using: CS

v3a08ab

NA - Reason not using: CS

v3a08ac

NA - Reason not using: CS

v3a08ad

NA - Reason not using: CS

v3a08x

Reason not using: other

v3a08z

Reason not using: don't know

v3a09a

NA - How many pill cycles did get last time

v3a09b

NA - How many condoms did get last time

v401

Last birth a caesarean section

v404

Currently breastfeeding

v405

Currently amenorrheic

v406

Currently abstaining

v407

NA - Number of times breastfed during night

v408

NA - Number of times breastfed during day

v409

Gave child plain water

v409a

NA - Gave child sugar water

v410

Gave child juice

v410a

NA - Gave child tea or coffee

v411

Gave child tinned, powdered or fresh milk

v411a

Gave child baby formula

v412

NA - Gave child fresh milk

v412a

Gave child fortified baby food (cerelac, etc)

v412b

NA - Gave child other porridge/gruel

v412c

Gave child soup/clear broth

v413

Gave child other liquid

v413a

NA - Gave child CS liquid

v413b

NA - Gave child CS liquid

v413c

NA - Gave child CS liquid

v413d

NA - Gave child CS liquid

v414a

NA - Gave child CS foods

v414b

NA - Gave child CS foods

v414c

NA - Gave child CS foods

v414d

NA - Gave child CS foods

v414e

Gave child bread, noodles, other made from grains

v414f

Gave child potatoes, cassava, or other tubers

v414g

Gave child eggs

v414h

Gave child meat (beef, pork, lamb, chicken, etc)

v414i

Gave child pumpkin, carrots, squash (yellow or orange inside)

v414j

Gave child any dark green leafy vegetables

v414k

Gave child mangoes, papayas, other vitamin A fruits

v414l

Gave child any other fruits

v414m

Gave child liver, heart, other organs

v414n

Gave child fish or shellfish

v414o

Gave child food made from beans, peas, lentils, nuts

v414p

Gave child cheese, yogurt, other milk products

v414q

NA - Gave child oil, fats, butter, products made of them

v414r

NA - Gave child chocolates, sweets, candies, pastries, etc

v414s

Gave child other solid-semisolid food

v414t

NA - Gave child CS foods

v414u

NA - Gave child CS foods

v414v

Gave child yogurt

v414w

NA - Gave child CS foods

v415

Drank from bottle with nipple yesterday/last night

v416

NA - Heard of oral dehydration

v417

Entries in maternity table

v418

Entries in health table

v419

Entries in height/weight table

v420

NA - Measurer's code

v421

NA - Assistant measurer's code

v426

When child put to breast

v437

Respondent's weight in kilograms (1 decimal)

v438

Respondent's height in centimeters (1 decimal)

v439

Height/Age percentile

v440

Height/Age standard deviation

v441

Height/Age percent ref. median

v442

Weight/Height percent ref. median (DHS)

v443

Weight/Height percent ref. median (Fog)

v444

Weight/Height percent ref. median (WHO)

v444a

Weight/Height standard deviation (DHS)

v445

Body Mass Index

v446

Rohrer's index

v447

Result of measurement - height/weight

v447a

Women's age in years (from household questionnaire)

v452a

NA - Under age 18 (from household questionnaire)

v452b

NA - Line number of parent/caretaker

v452c

NA - Read consent statement - hemoglobin

v453

NA - Hemoglobin level (g/dl - 1 decimal)

v454

NA - Currently pregnant (from household questionnaire)

v455

NA - Result of measurement - hemoglobin

v456

NA - Hemoglobin level adjusted for altitude and smoking (g/dl - 1 decimal)

v457

NA - Anemia level

v458

NA - Agrees to referral - anemia

v459

Have mosquito bed net for sleeping (from household questionnaire)

v460

NA - Children under 5 slept under mosquito bed net last night (household questio

v461

NA - Respondent slept under mosquito bed net

v462

NA - Washed hands before preparing last meal

v463a

Smokes cigarettes

v463b

Smokes pipe

v463c

Uses chewing tobacco

v463d

Uses snuff

v463e

Smokes huka/sheesha

v463f

NA - Smokes country specific

v463g

NA - Smokes country specific

v463x

Smokes other

v463z

Does not use tobacco

v464

Number of cigarettes in last 24 hours

v465

NA - Disposal of youngest child's stools when not using toilet

v466

NA - When child is seriously ill, can decide whether medical treatment sought

v467a

NA - Getting medical help for self: know where to go

v467b

Getting medical help for self: getting permission to go

v467c

Getting medical help for self: getting money needed for treatment

v467d

Getting medical help for self: distance to health facility

v467e

Getting medical help for self: having to take transport

v467f

Getting medical help for self: not wanting to go alone

v467g

NA - Getting medical help for self: concern no female health provider

v467h

NA - Getting medical help for self: concern no provider

v467i

NA - Getting medical help for self: concern no drugs available

v467j

NA - Getting medical help for self: CS

v467k

NA - Getting medical help for self: CS

v467l

NA - Getting medical help for self: CS

v467m

NA - Getting medical help for self: CS

v468

Record for Last Birth

v469e

Times gave child powdered/tinned/fresh milk

v469f

Times gave child infant formula

v469x

Times gave child yogurt

v471a

NA - Mother had tinned, powdered or fresh milk

v471b

NA - Mother had tea or coffee

v471c

NA - Mother had any other liquid

v471d

NA - Mother had CS liquid

v471e

NA - Mother had CS liquid

v471f

NA - Mother had CS liquid

v471g

NA - Mother had CS liquid

v472a

NA - Mother had CS foods

v472b

NA - Mother had CS foods

v472c

NA - Mother had CS foods

v472d

NA - Mother had CS foods

v472e

NA - Mother had bread, noodles, other made from grains

v472f

NA - Mother had potatoes, cassava, or other tubers

v472g

NA - Mother had eggs

v472h

NA - Mother had meat (beef, pork, lamb, chicken, etc)

v472i

NA - Mother had pumpkin, carrots, squash (yellow or orange inside)

v472j

NA - Mother had any dark green leafy vegetables

v472k

NA - Mother had mangoes, papayas, other vitamin A fruits

v472l

NA - Mother had any other fruits

v472m

NA - Mother had liver, heart, other organs

v472n

NA - Mother had fish or shellfish

v472o

NA - Mother had food made from beans, peas, lentils, nuts

v472p

NA - Mother had cheese, yogurt, other milk products

v472q

NA - Mother had oil, fats, butter, products made of them

v472r

NA - Mother had chocolates, sweets, candies, pastries, etc

v472s

NA - Mother had other solid-semisolid food

v472t

NA - Mother had CS foods

v472u

NA - Mother had CS foods

v473a

NA - Read consent statement - HIV

v473b

NA - Result of measurement - HIV

v474

Heard of tuberculosis or TB

v474a

Tuberculosis spread by: air when coughing or sneezing

v474b

Tuberculosis spread by: sharing utensils

v474c

Tuberculosis spread by: touching a person with TB

v474d

Tuberculosis spread by: food

v474e

Tuberculosis spread by: sexual contact

v474f

Tuberculosis spread by: mosquito bites

v474g

NA - Tuberculosis spread by: CS

v474h

NA - Tuberculosis spread by: CS

v474i

NA - Tuberculosis spread by: CS

v474j

NA - Tuberculosis spread by: CS

v474x

Tuberculosis spread by: other

v474z

Tuberculosis spread by: don't know

v475

Tuberculosis can be cured

v476

NA - Keep secret when family member gets TB

v477

Number of injections in last 12 months

v478

Number of injections administered by a health worker

v479

NA - Place of injection by health worker

v480

Syringe and needle from new, unopened package

v481

NA - Covered by health insurance

v481a

NA - Health insurance type: mutual/community organization

v481b

NA - Health insurance type: provided by employer

v481c

NA - Health insurance type: social security

v481d

NA - Health insurance type: private/commercially purchased

v481e

NA - Health insurance type: CS

v481f

NA - Health insurance type: CS

v481g

NA - Health insurance type: CS

v481h

NA - Health insurance type: CS

v481x

NA - Health insurance type: other

v482a

NA - Arrange care of biological children under age 18

v482b

NA - Primary caregiver of children under age 18

v482c

NA - Arrange care of non-biological children < 18

v501

Current marital status

v502

Currently/formerly/never in union

v503

Number of unions

v504

Currently residing with husband/partner

v505

Number of other wives

v506

Respondent's rank among wives

v507

Month of first cohabitation

v508

Year of first cohabitation

v509

Date of first cohabitation (CMC)

v510

Completeness of date information in V509

v511

Age at first cohabitation

v512

Years since first cohabitation

v513

Cohabitation duration (grouped)

v525

NA - Age at first sex

v527

Time since last sex

v528

Time since last sex (in days)

v529

Time since last sex (in months)

v530

Flag for V529

v531

NA - Age at first sex (imputed)

v532

NA - Flag for V531

v535

NA - Ever been married or in union

v536

Recent sexual activity

v537

Months of abstinence

v538

NA - How previous marriage or union ended

v539

NA - Person who received most of late husband's property

v540

NA - Respondent received any of late husband's assets or valuables

v541

NA - Intends to wait until marriage to have sex

v602

Fertility preference

v603

Preferred waiting time for birth of a/another child

v604

Preferred waiting time for birth of a/another child (grouped)

v605

Desire for more children

v613

Ideal number of children

v614

Ideal number of children (grouped)

v616

Time for future birth

v621

Husband's desire for children

v623

Exposure

v624

Unmet need

v625

Exposure (definition 2)

v626

Unmet need (definition 2)

v625a

Exposure to need for contraception (definition 3)

v626a

Unmet need for contraception (definition 3)

v627

Ideal number of boys

v628

Ideal number of girls

v629

Ideal number of either sex

v631

NA - Problem if became pregnant

v632

Decision maker for using contraception

v633a

NA - Reason for not having sex: husband has STI

v633b

NA - Reason for not having sex: husband has other women

v633c

NA - Reason for not having sex: recent birth

v633d

NA - Reason for not having sex: tired, not in mood

v633e

NA - Reason for not having sex: CS

v633f

NA - Reason for not having sex: CS

v633g

NA - Reason for not having sex: CS

v634

NA - Husband knows that respondent is using contraception

v701

Husband/partner's education level

v702

Husband/partner's highest year of education (at level in V701)

v704

Husband/partner's occupation

v705

Husband/partner's occupation (grouped)

v714

Respondent currently working

v714a

Respondent has a job, but currently absent

v715

Husband/partner's total number of years of education

v716

Respondent's occupation

v717

Respondent's occupation (grouped)

v719

Respondent works for family, others, self

v721

Respondent works at home or away

v729

Husband/partner's educational attainment

v730

Husband/partner's age

v731

Respondent worked in last 12 months

v732

Respondent employed all year/seasonal

v739

Person who usually decides how to spend respondent's earnings

v740

NA - Type of land where respondent works

v741

Type of earnings from respondent's work

v743a

Person who usually decides on respondent's health care

v743b

Person who usually decides on large household purchases

v743c

NA - Person who usually decides on household purchases for daily needs

v743d

Person who usually decides on visits to family or relatives

v743e

NA - Person who usually decides on food to be cooked each day

v743f

Person who usually decides what to do with money husband earns

v744a

Beating justified if wife goes out without telling husband

v744b

Beating justified if wife neglects the children

v744c

Beating justified if wife argues with husband

v744d

Beating justified if wife refuses to have sex with husband

v744e

Beating justified if wife burns the food

v745a

Owns a house alone or jointly

v745b

Owns land alone or jointly

v746

Respondent earns more than husband/partner

bidx

Birth column number

bord

Birth order number

b0

Child is twin

b1

Month of birth

b2

Year of birth

b3

Date of birth (CMC)

b4

Sex of child

b5

Child is alive

b6

Age at death

b7

Age at death (months, imputed)

b8

Current age of child

b9

Child lives with whom

b10

Completeness of information

b11

Preceding birth interval (months)

b12

Succeeding birth interval (months)

b13

Flag for age at death

b15

Live birth between births

b16

Child's line number in household

m1

Number of tetanus injections before birth

m1a

Number of tetanus injections before pregnancy

m1b

NA - Month of last tetanus injection before pregnancy

m1c

NA - Year of last tetanus injection before pregnancy

m1d

Years ago received last tetanus injection before pregnancy

m1e

NA - Last tetanus injection before pregnancy (CMC)

m2a

Prenatal: doctor

m2b

Prenatal: Nurse/midwife/LHV

m2c

NA - Prenatal: CS health professional

m2d

NA - Prenatal: CS health professional

m2e

NA - Prenatal: CS health professional

m2f

NA - Prenatal: CS health professional

m2g

Prenatal: DAI-TBA

m2h

Prenatal: lady health worker

m2i

Prenatal: homeopath

m2j

Prenatal: hakim

m2k

Prenatal: other

m2l

Prenatal: dispenser/compounder

m2m

NA - Prenatal: CS other

m2n

Prenatal: no one

m3a

Assistance: doctor

m3b

Assistance: nurse/midwife/LHV

m3c

NA - Assistance: CS health professional

m3d

NA - Assistance: CS health professional

m3e

NA - Assistance: CS health professional

m3f

NA - Assistance: CS health professional

m3g

Assistance: DAI-traditional birth attendant

m3h

Assistance: Family welfare center (FWW)

m3i

Assistance: lady health worker

m3j

Assistance: homeopath

m3k

Assistance: other

m3l

Assistance: hakim

m3m

Assistance: relative friend

m3n

Assistance: no one

m4

Duration of breastfeeding

m5

Months of breastfeeding

m6

Duration of amenorrhea

m7

Months of amenorrhea

m8

Duration of abstinence

m9

Months of abstinence

m10

Wanted pregnancy when became pregnant

m11

Desired time would have waited

m13

Timing of 1st antenatal check (months)

m14

Number of antenatal visits during pregnancy

m15

Place of delivery

m17

Delivery by caesarean section

m18

Size of child at birth

m19

Birth weight in kilograms (3 decimals)

m19a

Weight at birth/recall

m27

NA - Flag for breastfeeding

m28

Flag for amenorrhea

m29

Flag for abstinence

m34

When child put to breast

m35

NA - Number of times breastfed during previous night

m36

NA - Number of times breastfed during previous day

m38

Drank from bottle with nipple yesterday/last night

m39a

Did eat any solid, semi-solid or soft foods yesterday

m39

Number of times ate solid, semi-solid or soft food yesterday

m42a

During pregnancy: weighed

m42b

NA - During pregnancy: height measured

m42c

During pregnancy: blood pressure taken

m42d

During pregnancy: urine sample taken

m42e

During pregnancy: blood sample taken

m43

Told about pregnancy complications

m44

NA - Told where to go for pregnancy complications

m45

During pregnancy, given or bought iron tablets/syrup

m46

Days tablets or syrup taken

m47

NA - During pregnancy, had difficulty with daylight vision

m48

NA - During pregnancy, had difficulty with night blindness

m49a

NA - During pregnancy took: SP/fansidar for malaria

m49b

NA - During pregnancy took: chloroquine for malaria

m49c

NA - During pregnancy took: CS drug for malaria

m49d

NA - During pregnancy took: CS drug for malaria

m49e

NA - During pregnancy took: CS drug for malaria

m49f

NA - During pregnancy took: CS drug for malaria

m49g

NA - During pregnancy took: CS drug for malaria

m49x

NA - During pregnancy took: other drug for malaria

m49z

NA - During pregnancy took: don't know

m49y

NA - During pregnancy took: no drug for malaria

m50

Respondent's check up after delivery

m51

Respondent's checkup after deliver timing

m52

Respondent's health professional checked up after delivery

m54

Received Vitamin A dose in first 2 months after delivery

m55a

First 3 days, given milk (other than breast milk)

m55b

First 3 days, given plain water

m55c

First 3 days, given honey/sugar water

m55d

First 3 days, given gripe water

m55e

NA - First 3 days, given sugar/salt solution

m55f

First 3 days, given fruit juice

m55g

First 3 days, given infant formula

m55h

First 3 days, given green tea

m55i

NA - First 3 days, given honey

m55j

NA - First 3 days, given coffee

m55k

First 3 days, ghee butter

m55l

First 3 days, ghutee

m55m

First 3 days, given rose water

m55n

NA - First 3 days, given country specific

m55o

NA - First 3 days, given country specific

m55x

First 3 days, given other

m55z

First 3 days, given nothing

m57a

Antenatal care: respondent's home

m57b

Antenatal care: other home

m57c

NA - Antenatal care: CS home

m57d

NA - Antenatal care: CS home

m57e

Antenatal care: government hospital

m57f

Antenatal care: Rural health centre(RHC)/Mother child health centre(MCH)

m57g

Antenatal care: BHU/FWC

m57h

Antenatal care: other public health

m57i

NA - Antenatal care: CS public health

m57j

NA - Antenatal care: CS public health

m57k

NA - Antenatal care: CS public health

m57l

NA - Antenatal care: CS public health

m57m

Antenatal care: private hospital/clinic

m57n

Antenatal care: private doctor

m57o

Antenatal care: homeopath

m57p

Antenatal care: dispenser/compounder

m57q

Antenatal care: other private

m57r

NA - Antenatal care: CS private medical

m57s

Antenatal care: hakim

m57t

NA - Antenatal care: CS other

m57u

NA - Antenatal care: CS other

m57v

NA - Antenatal care: CS other

m57x

Antenatal care: other

m60

Drugs for intestinal parasites during pregnancy

m61

Time spent at place of delivery

m62

Respondent's health checked before discharge

m65a

NA - Reason didn't deliver at health facility: cost too much

m65b

NA - Reason didn't deliver at health facility: facility not open

m65c

NA - Reason didn't deliver at health facility: too far/no transport

m65d

NA - Reason didn't deliver at health facility: don't trust facility/poor service

m65e

NA - Reason didn't deliver at health facility: no female provider

m65f

NA - Reason didn't deliver at health facility: husband/family didn't allow

m65g

NA - Reason didn't deliver at health facility: not necessary

m65h

NA - Reason didn't deliver at health facility: not customary

m65i

NA - Reason didn't deliver at health facility: CS

m65j

NA - Reason didn't deliver at health facility: CS

m65k

NA - Reason didn't deliver at health facility: CS

m65l

NA - Reason didn't deliver at health facility: CS

m65x

NA - Reason didn't deliver at health facility: other

m66

Respondent's health checked after discharge/delivery at home

m70

Baby postnatal check within 2 months

m71

Time after delivery postnatal check took place

m72

Person who performed postnatal checkup

m73

Place baby was first checked

hidx

Index to birth history

h1

Has health card

h2

Received BCG

h2d

BCG day

h2m

BCG month

h2y

BCG year

h3

Received DPT 1

h3d

DPT 1 day

h3m

DPT 1 month

h3y

DPT 1 year

h4

Received POLIO 1

h4d

POLIO 1 day

h4m

POLIO 1 month

h4y

POLIO 1 year

h5

Received DPT 2

h5d

DPT 2 day

h5m

DPT 2 month

h5y

DPT 2 year

h6

Received POLIO 2

h6d

POLIO 2 day

h6m

POLIO 2 month

h6y

POLIO 2 year

h7

Received DPT 3

h7d

DPT 3 day

h7m

DPT 3 month

h7y

DPT 3 year

h8

Received POLIO 3

h8d

POLIO 3 day

h8m

POLIO 3 month

h8y

POLIO 3 year

h9

Received MEASLES

h9d

MEASLES day

h9m

MEASLES month

h9y

MEASLES year

h0

Received POLIO 0

h0d

POLIO 0 day

h0m

POLIO 0 month

h0y

POLIO 0 year

h10

Ever had vaccination

h11

Had diarrhea recently

h11b

Blood in the stools

h12a

Diarrhea: government hospital

h12b

Diarrhea: Rural health centre(RHC)/Mother child health centre(MCH)

h12c

Diarrhea: BHU/FWC

h12d

Diarrhea: lady health worker

h12e

Diarrhea: other public sector

h12f

NA - Diarrhea: CS public sector

h12g

NA - Diarrhea: CS public sector

h12h

NA - Diarrhea: CS public sector

h12i

NA - Diarrhea: CS public sector

h12j

Diarrhea: private hospital/clinic

h12k

Diarrhea: private pharmacy

h12l

Diarrhea: private doctor

h12m

Diarrhea: homeopath

h12n

Diarrhea: dispenser/compounder

h12o

Diarrhea: Other private sector

h12p

NA - Diarrhea: CS private medical

h12q

NA - Diarrhea: CS private medical

h12r

NA - Diarrhea: CS private medical

h12s

Diarrhea: shop

h12t

Diarrhea: hakim

h12u

Diarrhea: DAI-TBA

h12v

NA - Diarrhea: CS other sector

h12w

NA - Diarrhea: CS other sector

h12x

Diarrhea: Other

h12y

Diarrhea: no treatment

h12z

Diarrhea: medical treatment

h13

Given oral dehydration

h13b

NA - Given pre-packaged ORS liquid

h14

Given recommended home solution

h15

Given antibiotic pills or syrups

h15a

Given antimotility

h15b

Given antibiotic injection

h15c

Given intravenous (IV)

h15d

NA - Given home remedy, herbal medicine

h15e

Given zinc

h15f

Given other (not antibiotic, antimotility, zinc)

h15g

Given unknown pill or syrup

h15h

Given non-antibiotic injection

h15i

Given unknown injection

h15j

Given rice starch

h15k

Given mint extract

h15l

Given herbal remedy

h15m

NA - Given CS

h20

Given other treatment

h21a

Given no treatment

h21

Received any treatment

h22

Had fever in last two weeks

h31

Had cough in last two weeks

h31b

Short, rapid breaths

h31c

Problem in the chest or blocked or running nose

h31d

Had fever/cough in last 2 weeks: amount offered to drink

h31e

Had fever/cough in last 2 weeks: amount offered to eat

h32a

Fever/cough: government hospital

h32b

Fever/cough: Rural health centre(RHC)/Mother child health centre(MCH)

h32c

Fever/cough: BHU/FWC

h32d

Fever/cough: lady health worker

h32e

Fever/cough: other public sector

h32f

NA - Fever/cough: CS public sector

h32g

NA - Fever/cough: CS public sector

h32h

NA - Fever/cough: CS public sector

h32i

NA - Fever/cough: CS public sector

h32j

Fever/cough: private hospital/ clinic

h32k

Fever/cough: private pharmacy

h32l

Fever/cough: private doctor

h32m

Fever/cough: homeopath

h32n

Fever/cough: dispenser/compounder

h32o

Fever/cough: other private medical

h32p

NA - Fever/cough: CS private medical

h32q

NA - Fever/cough: CS private medical

h32r

NA - Fever/cough: CS private medical

h32s

Fever/cough: shop

h32t

Fever/cough: hakim

h32u

Fever/cough: DAI - TBA

h32v

NA - Fever/cough: CS other sector

h32w

NA - Fever/cough: CS other sector

h32x

Fever/cough: other

h32y

Fever/cough: no treatment

h32z

Fever/cough: medical treatment

h33

Received Vitamin A1 (most recent)

h33d

Vitamin A1 day

h33m

Vitamin A1 month

h33y

Vitamin A1 year

h34

Vitamin A in last 6 months

h35

NA - Any vaccinations in last 2 years part of campaign

h36a

NA - Vaccinated during Campaign A

h36b

NA - Vaccinated during Campaign B

h36c

NA - Vaccinated during Campaign C

h36d

NA - Vaccinated during Campaign D

h36e

NA - Vaccinated during Campaign E

h36f

NA - Vaccinated during Campaign F

h37a

SP/Fansidar taken for fever/cough

h37b

Chloroquine taken for fever/cough

h37c

Amodiaquine taken for fever/cough

h37d

Quinine taken for fever/cough

h37e

Combination with artemisinin taken for fever/cough

h37f

NA - CS antimalarial taken for fever/cough

h37g

NA - CS antimalarial taken for fever/cough

h37h

Other antimalarial taken for fever/cough

h37i

Antibiotic pill/syrup taken for fever/cough

h37j

Antibiotic injection taken for fever/cough

h37k

NA - Aspirin taken for fever/cough

h37l

Acetaminophen/paracetamol/panadol taken for fever/cough

h37m

Ibuprofen taken for fever/cough

h37n

Ponstan taken for fever/cough

h37o

Cough syrup for fever/cough

h37p

NA - CS other for fever/cough

h37x

Other taken for fever/cough

h37y

Nothing taken for fever/cough

h37z

Don't know if or what was taken for fever/cough

h38

Diarrhea in last 2 weeks: amount offered to drink

h39

Diarrhea in last 2 weeks: amount offered to eat

h40

NA - Received Vitamin A2 (2nd most recent)

h40d

NA - Vitamin A2 day

h40m

NA - Vitamin A2 month

h40y

NA - Vitamin A2 year

h41a

NA - Received vitamin A after most recent in card

h41b

NA - Ever received Vitamin A dose

h42

Taking iron pills, sprinkles or syrup

h43

Drugs for intestinal parasites in last 6 months

h44a

NA - Place first sought treatment for diarrhea

h44b

NA - Days after diarrhea sought advice or treatment

h44c

NA - Still has diarrhea

h45

NA - Times zinc was given

h46a

NA - Place first sought treatment for fever

h46b

NA - Days after fever sought advice or treatment

h47

Blood taken from child's finger/heel for testing

hwidx

Index to Birth History

hw1

Child's age in months

hw2

Child's weight in kilograms (1 decimal)

hw3

Child's height in centimeters (1 decimal)

hw4

Height/Age percentile

hw5

Height/Age standard deviation

hw6

Height/Age percent of ref. median

hw7

Weight/Age percentile

hw8

Weight/Age standard deviation

hw9

Weight/Age percent of ref. median

hw10

Weight/Height percentile

hw11

Weight/Height standard deviation

hw12

Weight/Height percent of ref. median

hw13

Result of measurement - height/weight

hw15

Height: lying or standing

hw16

Day of birth of child

hw17

Date measured (day)

hw18

Date measured (month)

hw19

Date measured (year)

hw51

NA - Line number of parent/caretaker

hw52

NA - Read consent statement - hemoglobin

hw53

NA - Hemoglobin level (g/dl - 1 decimal)

hw55

NA - Result of measurement - hemoglobin

hw56

NA - Hemoglobin level adjusted for altitude (g/dl - 1 decimal)

hw57

NA - Anemia level

hw58

NA - Agrees to referral - anemia

hw70

Height/Age standard deviation (new WHO)

hw71

Weight/Age standard deviation (new WHO)

hw72

Weight/Height standard deviation (new WHO)

hw73

BMI standard deviation (new WHO)

sdist

District

slangq

Language of questionnaire

slangi

Language of interview

slangr

Language respondent speaks a home

seduc

Woman's educational level for report

swfeduc1

All woman factor - education as will be used in the report

swfregur

All woman factor - region combined with urban/rural

s110

Blood relation with husband

s111

Type of relationship

s113

Woman have say in choosing husband

s117

Highest class completed

s209

Number of pregnancy losses

s210

Total pregnancy outcomes

s318

Source for current method when started

s323

Ever experienced side effects with current FP method

s324a

Side effects for current method: obesity/weight gain

s324b

Side effects for current method: headache

s324c

Side effects for current method: nausea/dizziness

s324d

Side effects for current method: excessive bleeding

s324e

Side effects for current method: spotting

s324f

Side effects for current method: irregular/no menses

s324g

Side effects for current method: depression

s324x

Side effects for current method: other

s325

Seek medical advice for side effects with current FP method

s326a

Seek advice for side effects from: doctor

s326b

Seek advice for side effects from: nurse/midwife

s326c

Seek advice for side effects from: DAI-TBA

s326d

Seek advice for side effects from: lady worker

s326e

Seek advice for side effects from: homeopath

s326f

Seek advice for side effects from: hakim

s326g

Seek advice for side effects from: dispenser

s326x

Seek advice for side effects from: other

s326y

Seek advice for side effects from: no one

s327a

Place where advice was sought: govt. hospital

s327b

Place where advice was sought: rural health centre

s327c

Place where advice was sought: family welfare centre

s327d

Place where advice was sought: Mother child health centre(MCH)

s327e

Place where advice was sought: dispenser

s327f

Place where advice was sought: mobile service camp

s327g

Place where advice was sought: lady health worker

s327h

Place where advice was sought: Lady health visitor

s327i

Place where advice was sought: basic health unit

s327j

Place where advice was sought: male mobilizer

s327k

Place where advice was sought: Family welfare assistant(FWA)

s327l

Place where advice was sought: other public

s327m

Place where advice was sought: private/NGO hospital/clinic

s327n

Place where advice was sought: pharmacy

s327o

Place where advice was sought: private doctor

s327p

Place where advice was sought: homeopath

s327q

Place where advice was sought: dispenser/compounder

s327r

Place where advice was sought: other private medical

s327s

Place where advice was sought: shop

s327t

Place where advice was sought: friend/relative

s327u

Place where advice was sought: hakim

s327v

Place where advice was sought: DAI-TBA

s327x

Place where advice was sought: other

s327z

Place where advice was sought: Don't know

s328a

Reason didn't seek advice: not necessary

s328b

Reason didn't seek advice: cost too much

s328c

Reason didn't seek advice: too far

s328d

Reason didn't seek advice: no transport

s328e

Reason didn't seek advice: no one to go with

s328f

Reason didn't seek advice: service not good

s328g

Reason didn't seek advice: no time to go

s328h

Reason didn't seek advice: didn't know where to go

s328i

Reason didn't seek advice: lady doctor wasn't available

s328j

Reason didn't seek advice: long waiting time

s328k

Reason didn't seek advice: not allowed to go

s328x

Reason didn't seek advice: other

s332

Last source of FP

s335

Knows that lady health worker is present in the area

s337

Received any care from lady health worker

s338a

Help received on: info on mother and child health

s338b

Help received on: contraceptive supplies

s338c

Help received on: referred to health facility

s338d

Help received on: treatment of side effects

s338e

Help received on: vaccination

s338f

Help received on: treatment of minor ailment

s338g

Help received on: treatment for mother and child

s338x

Help received on: other

s341

Any service outlet that provide FP services

s342

Ever visited any service outlet that provide FP services

s343

Kind of service outlet visit

s344a

Satisfied with: Provision of contraception

s344b

Satisfied with: Follow-up care

s344c

Satisfied with: Infection prevention

s344d

Satisfied with: Counseling services

s344e

Satisfied with: Timely treatment

s344f

Satisfied with: Attitude of staff

s344g

Satisfied with: Punctuality maintained

s344h

Satisfied with: Timely referring

s344i

Satisfied with: Cooperative

s344j

Satisfied with: Handle complications promptly

s344x

Satisfied with: Others

s345a

Reason for not visiting FP services: services at door steps

s345b

Reason for not visiting FP services: preferred other providers

s345c

Reason for not visiting FP services: undesirable location

s345d

Reason for not visiting FP services: not necessary to visit

s345e

Reason for not visiting FP services: unaware about the service

s345f

Reason for not visiting FP services: wanted more children

s345x

Reason for not visiting FP services: other

s617a

Family planning message conveyed: limiting the family

s617b

Family planning message conveyed: higher age at marriage

s617c

Family planning message conveyed: spacing of children

s617d

Family planning message conveyed: use of contraceptives

s617e

Family planning message conveyed: welfare of family

s617f

Family planning message conveyed: maternal and child health

s617g

Family planning message conveyed: less children, prosperous life

s617h

Family planning message conveyed: more children, poverty and starvation

s617i

Family planning message conveyed: importance of breastfeeding

s617x

Family planning message conveyed: other

s617z

Family planning message conveyed: DK/don't remember

s618

Was family planning message effective

s705a

Women work cash before got married

s705b

Women work cash after got married

s705c

Year when start working after marriage

s724

Have the autonomy to sell the house own

s726

Have the autonomy to sell the land own

s728e

Justifies domestic violence: Neglected In-laws

s904

Duration of treatment of TB nowadays

s905

Ever been told by a doctor/nurse that had TB

s906

Ever heard about Hepatitis

s907

Person can avoid Hepatitis

s908a

Avoid getting Hepatitis by: safe sex

s908b

Avoid getting Hepatitis by: safe blood transfer

s908c

Avoid getting Hepatitis by: disposable syringe

s908d

Avoid getting Hepatitis by: avoid contaminated food/water

s908e

Avoid getting Hepatitis by: avoid contact with infected person

s908f

Avoid getting Hepatitis by: dentist instruments sterilized

s908x

Avoid getting Hepatitis by: other

s908y

Avoid getting Hepatitis by: no response

s908z

Avoid getting Hepatitis by: don't know

idx94

Index to Birth History

s413e

During pregnancy: have ultrasound exam

idx95

Index to Birth History

s506

Type of vaccination administered Combo/Penta/DPT

shb1

Received hepatitis B-1

shb1d

Hepatitis B-1 day

shb1m

Hepatitis B-1 month

shb1y

Hepatitis B-1 year

shb2

Received hepatitis B-2

shb2d

Hepatitis B-2 day

shb2m

Hepatitis B-2 month

shb2y

Hepatitis B-2 year

shb3

Received hepatitis B-3

shb3d

Hepatitis B-3 day

shb3m

Hepatitis B-3 month

shb3y

Hepatitis B-3 year

References

National Institute of Population Studies (NIPS), Pakistan (http://nips.org.pk/Home.htm).

Examples

data("PKKR61")
PKKR61

#> # A tibble: 11,763 x 1,000
#>    caseid  midx v000   v001  v002  v003  v004   v005  v006  v007  v008  v009
#>    <chr>  <int> <chr> <int> <int> <int> <int>  <int> <int> <int> <int> <int>
#>  1 1 2  2     1 PK6       1     2     2     1 2.15e6    11  2012  1355     2
#>  2 1 2  2     2 PK6       1     2     2     1 2.15e6    11  2012  1355     2
#>  3 1 2  2     3 PK6       1     2     2     1 2.15e6    11  2012  1355     2
#>  4 1 2  2     4 PK6       1     2     2     1 2.15e6    11  2012  1355     2
#>  5 1 5  2     1 PK6       1     5     2     1 2.15e6    11  2012  1355    10
#>  6 1 5  2     2 PK6       1     5     2     1 2.15e6    11  2012  1355    10
#>  7 1 7  4     1 PK6       1     7     4     1 2.15e6    11  2012  1355     1
#>  8 1 9  6     1 PK6       1     9     6     1 2.15e6    11  2012  1355     4
#>  9 1 9  6     2 PK6       1     9     6     1 2.15e6    11  2012  1355     4
#> 10 1 9  6     3 PK6       1     9     6     1 2.15e6    11  2012  1355     4
#> # … with 11,753 more rows, and 988 more variables: v010 <int>, v011 <int>,
#> #   v012 <int>, v013 <fct>, v014 <fct>, v015 <fct>, v016 <int>, v017 <int>,
#> #   v018 <int>, v019 <int>, v019a <int>, v020 <fct>, v021 <int>, v022 <int>,
#> #   v023 <fct>, v024 <fct>, v025 <fct>, v026 <fct>, v027 <int>, v028 <int>,
#> #   v029 <int>, v030 <int>, v031 <int>, v032 <int>, v034 <int>, v040 <int>,
#> #   v042 <fct>, v044 <fct>, v101 <fct>, v102 <fct>, v103 <fct>, v104 <fct>,
#> #   v105 <fct>, v106 <fct>, v107 <int>, v113 <int>, v115 <int>, v116 <int>,
#> #   v119 <fct>, v120 <int>, v121 <int>, v122 <int>, v123 <int>, v124 <int>,
#> #   v125 <int>, v127 <int>, v128 <int>, v129 <int>, v130 <fct>, v131 <int>,
#> #   v133 <int>, v134 <fct>, v135 <int>, v136 <int>, v137 <int>, v138 <int>,
#> #   v139 <fct>, v140 <fct>, v141 <fct>, v149 <fct>, v150 <fct>, v151 <fct>,
#> #   v152 <int>, v153 <int>, awfactt <int>, awfactu <int>, awfactr <int>,
#> #   awfacte <int>, awfactw <int>, v155 <int>, v156 <fct>, v157 <int>,
#> #   v158 <int>, v159 <int>, v160 <int>, v161 <int>, v166 <fct>, v167 <fct>,
#> #   v168 <fct>, v190 <fct>, v191 <int>, ml101 <fct>, v201 <int>, v202 <int>,
#> #   v203 <int>, v204 <int>, v205 <int>, v206 <int>, v207 <int>, v208 <int>,
#> #   v209 <int>, v210 <int>, v211 <int>, v212 <int>, v213 <fct>, v214 <int>,
#> #   v215 <int>, v216 <fct>, v217 <int>, v218 <int>, …