Pakistan Demographic and Health Survey 2012-13 Men's Recode Data

PKMR61 data of Men's Recode from Pakistan Demographic and Health Survey 2012-13.

mcaseid

Case Identification

mv000

Country code and phase

mv001

Cluster number

mv002

Household number

mv003

Respondent's line number

mv004

Ultimate area unit

mv005

Men's sample weight (6 decimals)

mv006

Month of interview

mv007

Year of interview

mv008

Date of interview (CMC)

mv009

Month of birth

mv010

Year of birth

mv011

Date of birth (CMC)

mv012

Current age

mv013

Age in 5-year groups

mv014

Completeness of age information

mv015

Result of interview

mv016

Day of interview

mv021

Primary sampling unit

mv022

Sample stratum number

mv023

Sample domain

mv024

Region

mv025

Type of place of residence

mv026

De facto place of residence

mv027

Number of visits

mv028

Interviewer identification

mv029

Keyer identification

mv030

Field supervisor

mv031

Field editor

mv032

NA - Office editor

mv034.1

NA - Line number of wife/partner

mv034.2

NA - Line number of wife/partner

mv034.3

NA - Line number of wife/partner

mv034.4

NA - Line number of wife/partner

mv034.5

NA - Line number of wife/partner

mv034.6

NA - Line number of wife/partner

mv034.7

NA - Line number of wife/partner

mv034.8

NA - Line number of wife/partner

mv034a.1

NA - Wife or partner***

mv034a.2

NA - Wife or partner***

mv034a.3

NA - Wife or partner***

mv034a.4

NA - Wife or partner***

mv034a.5

NA - Wife or partner***

mv034a.6

NA - Wife or partner***

mv034a.7

NA - Wife or partner***

mv034a.8

NA - Wife or partner***

mv034b.1

NA - Age of wife/partner

mv034b.2

NA - Age of wife/partner

mv034b.3

NA - Age of wife/partner

mv034b.4

NA - Age of wife/partner

mv034b.5

NA - Age of wife/partner

mv034b.6

NA - Age of wife/partner

mv034b.7

NA - Age of wife/partner

mv034b.8

NA - Age of wife/partner

mv035

Number of wives/partners

mv801

Time interview started (hhmm - 24 hour clock)

mv802

Time interview ended (hhmm - 24 hour clock)

mv803

Length of interview in minutes

mv101

Region

mv102

Type of place of residence

mv103

NA - Childhood place of residence***

mv104

NA - Years lived in place of residence

mv105

NA - Type of place of previous residence

mv106

Educational level

mv107

Highest year of education (at level in MV106)

mv130

NA - Religion

mv131

Ethnicity

mv133

Total number of years of education

mv134

De facto place of residence

mv135

Usual resident or visitor

mv136

Number of household members (total listed)

mv138

Number of eligible men in household (de facto)

mv149

Educational attainment

mv150

Relationship to household head

mv151

Sex of household head

mv152

Age of household head

mv155

Literacy

mv156

NA - Ever participated in a literacy program (excluding primary school)

mv157

Frequency of reading newspaper or magazine

mv158

Frequency of listening to radio

mv159

Frequency of watching television

mv167

NA - Times away from home in last 12 months

mv168

NA - Away for more than one month in last 12 months

mv190

Wealth index

mv191

Wealth index factor score (5 decimals)

mv201

Total children ever born

mv202

Sons at home

mv203

Daughters at home

mv204

Sons elsewhere

mv205

Daughters elsewhere

mv206

Sons who have died

mv207

Daughters who have died

mv212

Age of respondent at 1st birth

mv213

Partner currently pregnant

mv217

Knowledge of ovulatory cycle

mv218

Number of living children

mv225

NA - Most recent pregnancy wanted

mv245

Number of women fathered children with

mv246

NA - Married to mother when first child was born

mv247

Age of most recent child

mv248

Antenatal check-ups for the mother of most recent child

mv249

Respondent present during check-ups for most recent child

mv250

Place of birth of most recent child

mv251

NA - Reason for not delivering most recent child in health facility

mv252

Amount a child should drink, when that child has diarrhea

mv301

Knowledge of any contraceptive method

mv302

NA - Ever use of any contraceptive method

m304a.01

Type of contraceptive method

m304a.02

Type of contraceptive method

m304a.03

Type of contraceptive method

m304a.04

Type of contraceptive method

m304a.05

Type of contraceptive method

m304a.06

Type of contraceptive method

m304a.07

Type of contraceptive method

m304a.08

Type of contraceptive method

m304a.09

Type of contraceptive method

m304a.10

Type of contraceptive method

m304a.11

Type of contraceptive method

m304a.12

Type of contraceptive method

m304a.13

Type of contraceptive method

m304a.14

Type of contraceptive method

m304a.15

Type of contraceptive method

m304a.16

Type of contraceptive method

m304a.17

Type of contraceptive method

m304a.18

Type of contraceptive method

m304a.19

Type of contraceptive method

m304a.20

Type of contraceptive method

mv304.01

Knows contraceptive method

mv304.02

Knows contraceptive method

mv304.03

Knows contraceptive method

mv304.04

Knows contraceptive method

mv304.05

Knows contraceptive method

mv304.06

Knows contraceptive method

mv304.07

Knows contraceptive method

mv304.08

Knows contraceptive method

mv304.09

Knows contraceptive method

mv304.10

Knows contraceptive method

mv304.11

Knows contraceptive method

mv304.12

Knows contraceptive method

mv304.13

Knows contraceptive method

mv304.14

Knows contraceptive method

mv304.15

Knows contraceptive method

mv304.16

Knows contraceptive method

mv304.17

Knows contraceptive method

mv304.18

Knows contraceptive method

mv304.19

Knows contraceptive method

mv304.20

Knows contraceptive method

mv305.01

Ever used contraceptive method

mv305.02

Ever used contraceptive method

mv305.03

Ever used contraceptive method

mv305.04

Ever used contraceptive method

mv305.05

Ever used contraceptive method

mv305.06

Ever used contraceptive method

mv305.07

Ever used contraceptive method

mv305.08

Ever used contraceptive method

mv305.09

Ever used contraceptive method

mv305.10

Ever used contraceptive method

mv305.11

Ever used contraceptive method

mv305.12

Ever used contraceptive method

mv305.13

Ever used contraceptive method

mv305.14

Ever used contraceptive method

mv305.15

Ever used contraceptive method

mv305.16

Ever used contraceptive method

mv305.17

Ever used contraceptive method

mv305.18

Ever used contraceptive method

mv305.19

Ever used contraceptive method

mv305.20

Ever used contraceptive method

mv307.01

NA - Contraceptive method currently used

mv307.02

NA - Contraceptive method currently used

mv307.03

NA - Contraceptive method currently used

mv307.04

NA - Contraceptive method currently used

mv307.05

NA - Contraceptive method currently used

mv307.06

NA - Contraceptive method currently used

mv307.07

NA - Contraceptive method currently used

mv307.08

NA - Contraceptive method currently used

mv307.09

NA - Contraceptive method currently used

mv307.10

NA - Contraceptive method currently used

mv307.11

NA - Contraceptive method currently used

mv307.12

NA - Contraceptive method currently used

mv307.13

NA - Contraceptive method currently used

mv307.14

NA - Contraceptive method currently used

mv307.15

NA - Contraceptive method currently used

mv307.16

NA - Contraceptive method currently used

mv307.17

NA - Contraceptive method currently used

mv307.18

NA - Contraceptive method currently used

mv307.19

NA - Contraceptive method currently used

mv307.20

NA - Contraceptive method currently used

mv312

NA - Current contraceptive method

mv313

NA - Current contraceptive by method type

mv323a

NA - Brand of condom used

mv325b

NA - Cost of condoms (in CURRENCY) last time obtained

mv372a

NA - Interviewer shown condom package

mv384a

On radio heard about Family Planning in the last few months

mv384b

On TV saw about Family Planning in the last few months

mv384c

In newspaper read about Family Planning in the last few months

mv395

Discussed Family Planning with health worker in last few months

mv396

NA - Possible for a breastfeeding woman to get pregnant

mv3a09b

NA - Number of condoms did get last time

mv3b17

NA - Source of condoms

mv3b25a

Contraception is woman's business, man should not worry

mv3b25b

Women who use contraception become promiscuous

mv463a

Smokes cigarettes

mv463b

Smokes pipe

mv463c

Uses chewing tobacco

mv463d

Uses snuff

mv463e

Smokes hukaa/sheesha

mv463f

NA - Smokes CS

mv463g

NA - Smokes CS

mv463x

Smokes other

mv463z

Smokes nothing

mv464

Number of cigarettes in last 24 hours

mv474

Heard of tuberculosis or TB

mv474a

Tuberculosis spread by: air through coughing or sneezing

mv474b

Tuberculosis spread by: sharing utensils

mv474c

Tuberculosis spread by: touching a person with TB

mv474d

Tuberculosis spread by: food

mv474e

Tuberculosis spread by: sexual contact

mv474f

Tuberculosis spread by: mosquito bites

mv474g

NA - Tuberculosis spread by: CS

mv474h

NA - Tuberculosis spread by: CS

mv474i

NA - Tuberculosis spread by: CS

mv474j

NA - Tuberculosis spread by: CS

mv474x

Tuberculosis spread by: other

mv474z

Don't know how tuberculosis is spread

mv475

Tuberculosis can be cured

mv476

NA - Would want tuberculosis infection in family to remain secret

mv477

Number of injections in last 12 months

mv478

Number of injections administered by a health worker

mv479

NA - Place where given injection by health worker

mv480

Syringe and needle from new, unopened package

mv481

NA - Covered by health insurance

mv481a

NA - Health insurance type: mutual/community organization

mv481b

NA - Health insurance type: provided by employer

mv481c

NA - Health insurance type: social security

mv481d

NA - Health insurance type: private/commercially purchased

mv481e

NA - Health insurance type: CS

mv481f

NA - Health insurance type: CS

mv481g

NA - Health insurance type: CS

mv481h

NA - Health insurance type: CS

mv481x

NA - Health insurance type: other

mv482a

NA - Arrange care of biological children under age 18***

mv482b

NA - Primary caregiver of children under age 18***

mv482c

NA - Arrange care of non-biological children < 18***

mv483

NA - Respondent circumcised

mv483a

NA - Age at circumcision

mv483b

NA - Who performed the circumcision

mv483c

NA - Place where circumcision was done

mv501

Current marital status

mv502

Currently/formerly/never in union

mv503

NA - Number of unions

mv504

Currently residing with wife/partner

mv505

Number of wives/partners

mv507

Month of first cohabitation

mv508

Year of first cohabitation

mv509

Date of first cohabitation (CMC)

mv510

Completeness of date information in MV509

mv511

Age at first cohabitation

mv512

Years since first cohabitation

mv513

Cohabitation duration (grouped)

mv525

NA - Age at first sex

mv527

NA - Time since last sex

mv528

NA - Time since last sex (in days)

mv529

NA - Time since last sex (in months)

mv531

NA - Age at first sex (imputed)

mv532

NA - Flag for MV531

mv535

Ever been married or in union

mv536

NA - Recent sexual activity

mv541

NA - Intends to wait until marriage to have sex

mv602

Fertility preference

mv603

Preferred waiting time for birth of a/another child

mv604

Preferred waiting time for birth of a/another child (grouped)

mv605

Desire for more children

mv613

Ideal number of children

mv614

Ideal number of children (grouped)

mv616

Desired time for future birth

mv621

NA - Wife's desire for children***

mv627

Ideal number of boys

mv628

Ideal number of girls

mv629

Ideal number of either sex

mv631

NA - Problem if wife became pregnant***

mv633a

NA - Wife justified refusing sex: husband has STI

mv633b

NA - Wife justified refusing sex: husband has other women

mv633c

NA - Wife justified refusing sex: recent birth

mv633d

NA - Wife justified refusing sex: tired, not in mood

mv633e

NA - Wife justified refusing sex: CS reason

mv633f

NA - Wife justified refusing sex: CS reason

mv633g

NA - Wife justified refusing sex: CS reason

mv634a

NA - Wife refuses to have sex, husband has right to: get angry

mv634b

NA - Wife refuses to have sex, husband has right to: refuse financial support

mv634c

NA - Wife refuses to have sex, husband has right to: use force for sex

mv634d

NA - Wife refuses to have sex, husband has right to: have sex with another women

mv714

Currently working

mv714a

Have a job, but currently absent

mv716

Occupation

mv717

Occupation (grouped)

mv719

NA - Works for family, others, self

mv721

NA - Works at home or away

mv731

Worked in last 12 months

mv732

Employment all year/seasonal

mv739

Person who usually decides how to spend respondent's earnings

mv740

NA - Type of land where respondent works

mv741

Type of earnings from respondent's work

mv743a

Person who should have greater say: respondent's health care

mv743b

Person who should have greater say: large household purchases

mv743c

NA - Person who should have greater say: daily household purchases

mv743d

NA - Person who should have greater say: visits to family or relatives

mv743e

NA - Person who should have greater say: food to be cooked each day

mv743f

NA - Person who should have greater say: deciding what to do with money wife ear

mv743g

NA - Person who should have greater say: deciding how many children to have

mv744a

Beating justified: wife goes out without telling husband

mv744b

Beating justified: wife neglects the children

mv744c

Beating justified: wife argues with husband

mv744d

Beating justified: wife refuses to have sex with husband

mv744e

Beating justified: wife burns food

mv745a

Owns a house alone or jointly

mv745b

Owns land alone or jointly

mv747a

NA - Childbearing is a woman's concern

mv747b

NA - Important for mother/child health to get assistance at delivery

mv750

Ever heard of a Sexually Transmitted Infection (STI)

mv751

Ever heard of AIDS

mv754bp

NA - Reduce risk of getting HIV: do not have sex at all

mv754cp

Reduce risk of getting HIV: always use condoms during sex

mv754dp

Reduce risk of getting HIV: have 1 sex partner only, who has no other partners

mv754jp

Can get HIV from mosquito bites

mv754wp

Can get HIV by sharing food with person who has AIDS

mv756

A healthy looking person can have HIV

mv761

NA - Condom used during last sex with most recent partner

mv761b

NA - Condom used during last sex with 2nd to most recent partner

mv761c

NA - Condom used during last sex with 3rd to most recent partner

mv762aa

Source for condoms: government hospital

mv762ab

Source for condoms: rural health center(RHC)

mv762ac

Source for condoms: family welfare center

mv762ad

Source for condoms: Mother child health centre(MCH)

mv762ae

Source for condoms: dispenser

mv762af

Source for condoms: lady health worker

mv762ag

Source for condoms: Lady health visitor

mv762ah

Source for condoms: male mobilizer

mv762ai

Source for condoms: other public sector

mv762aj

Source for condoms: private hospital/clinic

mv762ak

Source for condoms: pharmacy

mv762al

Source for condoms: private doctor

mv762am

Source for condoms: homeopath

mv762an

Source for condoms: dispenser/compounder

mv762ao

Source for condoms: other private medical

mv762ap

NA - Source for condoms: CS private

mv762aq

NA - Source for condoms: CS private

mv762ar

NA - Source for condoms: CS private

mv762as

Source for condoms: shop

mv762at

Source for condoms: friend/relative

mv762au

Source for condoms: hakim

mv762av

Source for condoms: DAI traditional birth attendant

mv762aw

NA - Source for condoms: CS other

mv762ax

Source for condoms: other

mv762az

Don't know any source for condoms

mv762ba

NA - Source for female condoms: government hospital

mv762bb

NA - Source for female condoms: CS public

mv762bc

NA - Source for female condoms: CS public

mv762bd

NA - Source for female condoms: CS public

mv762be

NA - Source for female condoms: CS public

mv762bf

NA - Source for female condoms: CS public

mv762bg

NA - Source for female condoms: CS public

mv762bh

NA - Source for female condoms: CS public

mv762bi

NA - Source for female condoms: CS public

mv762bj

NA - Source for female condoms: private hospital/clinic

mv762bk

NA - Source for female condoms: pharmacy

mv762bl

NA - Source for female condoms: private doctor

mv762bm

NA - Source for female condoms: CS private

mv762bn

NA - Source for female condoms: CS private

mv762bo

NA - Source for female condoms: CS private

mv762bp

NA - Source for female condoms: CS private

mv762bq

NA - Source for female condoms: CS private

mv762br

NA - Source for female condoms: CS private

mv762bs

NA - Source for female condoms: CS other

mv762bt

NA - Source for female condoms: CS other

mv762bu

NA - Source for female condoms: CS other

mv762bv

NA - Source for female condoms: CS other

mv762bw

NA - Source for female condoms: CS other

mv762bx

NA - Source for female condoms: other

mv762bz

NA - Don't know any source for female condoms

mv763a

Had any STI in last 12 months

mv763b

NA - Had genital sore/ulcer in last 12 months

mv763c

NA - Had genital discharge in last 12 months

mv763d

NA - Had CS STI in last 12 months

mv763e

NA - Had CS STI in last 12 months

mv763f

NA - Had CS STI in last 12 months

mv763g

NA - Had CS STI in last 12 months

mv766a

NA - Number of sex partners, excluding spouse, in last 12 months

mv766b

NA - Number of sex partners, including spouse, in last 12 months

mv767a

NA - Relationship with most recent sex partner

mv767b

NA - Relationship with 2nd to most recent sex partner

mv767c

NA - Relationship with 3rd to most recent sex partner

mv768a

NA - Length of time had sex relations with most recent partner

mv768b

NA - Length of time had sex relations with 2nd to most recent partner

mv768c

NA - Length of time had sex relations with 3rd to most recent partner

mv769

Can get a condom

mv769a

NA - Can get a female condom

mv770

Sought advice/treatment for last STI infection

mv770a

Sought STI advice/treatment from: government hospital

mv770b

Sought STI advice/treatment from: government health center

mv770c

Sought STI advice/treatment from: stand-alone VCT center

mv770d

Sought STI advice/treatment from: family planning clinic

mv770e

Sought STI advice/treatment from: mobile clinic

mv770f

Sought STI advice/treatment from: fieldworker

mv770g

Sought STI advice/treatment from: other public sector

mv770h

NA - Sought STI advice/treatment from: CS public

mv770i

NA - Sought STI advice/treatment from: CS public

mv770j

NA - Sought STI advice/treatment from: CS public

mv770k

Sought STI advice/treatment from: private hospital/clinic/doctor

mv770l

Sought STI advice/treatment from: stand-alone VCT center

mv770m

Sought STI advice/treatment from: pharmacy

mv770n

Sought STI advice/treatment from: mobile clinic

mv770o

Sought STI advice/treatment from: fieldworker

mv770p

Sought STI advice/treatment from: other private medical

mv770q

NA - Sought STI advice/treatment from: CS private

mv770r

NA - Sought STI advice/treatment from: CS private

mv770s

NA - Sought STI advice/treatment from: CS private

mv770t

Sought STI advice/treatment from: shop

mv770u

NA - Sought STI advice/treatment from: CS other

mv770v

NA - Sought STI advice/treatment from: CS other

mv770w

NA - Sought STI advice/treatment from: CS other

mv770x

Sought STI advice/treatment from: other

mv774a

HIV transmitted during pregnancy

mv774b

HIV transmitted during delivery

mv774c

HIV transmitted by breastfeeding

mv775

NA - Knows someone who has, or is suspected of having, HIV

mv777

Would want HIV infection in family to remain secret

mv778

Willing to care for relative with AIDS

mv779

A female teacher infected with HIV, but is not sick, should be allowed to contin

mv780

NA - Children should be taught about condoms to avoid AIDS

mv781

NA - Ever been tested for HIV

mv783

Know a place to get HIV test

mv784a

Place for HIV test: government hospital

mv784b

Place for HIV test: government health center

mv784c

Place for HIV test: stand-alone VCT center

mv784d

Place for HIV test: other public sector

mv784e

NA - Place for HIV test: CS public

mv784f

NA - Place for HIV test: CS public

mv784g

NA - Place for HIV test: CS public

mv784h

NA - Place for HIV test: CS public

mv784i

NA - Place for HIV test: CS public

mv784j

NA - Place for HIV test: CS public

mv784k

Place for HIV test: private hospital/clinic/doctor

mv784l

Place for HIV test: stand-alone VCT center

mv784m

Place for HIV test: other private medical

mv784n

NA - Place for HIV test: CS private

mv784o

NA - Place for HIV test: CS private

mv784p

NA - Place for HIV test: CS private

mv784q

NA - Place for HIV test: CS private

mv784r

NA - Place for HIV test: CS private

mv784s

NA - Place for HIV test: CS other

mv784t

NA - Place for HIV test: CS other

mv784u

NA - Place for HIV test: CS other

mv784v

NA - Place for HIV test: CS other

mv784x

Place for HIV test: other

mv785

Heard about other STIs

mv791

NA - Have ever paid anyone in exchange for sex

mv793

NA - Paid for sex in last 12 months

mv793a

NA - Condom used last time paid for sex in last 12 months

mv793b

NA - Condom used every time paid for sex in last 12 months

mv820

NA - Condom used at first sex

mv822

NA - Wife justified asking husband to use condom if he has STI

mv823

Can get HIV by witchcraft or supernatural means

mv824

Drugs to avoid HIV transmission to baby during pregnancy

mv825

Would buy vegetables from vendor with HIV

mv826

NA - Last time tested for HIV

mv826a

NA - Months ago most recent HIV test

mv827

NA - Last HIV test: on your own, offered or required

mv828

NA - Received result from last HIV test

mv829

NA - Place where last HIV test was taken

mv832b

NA - Time since last sex with 2nd to most recent partner

mv832c

NA - Time since last sex with 3rd to most recent partner

mv833a

NA - Used condom every time had sex with most recent partner in last 12 months

mv833b

NA - Used condom every time had sex with 2nd to most recent partner in last 12 m

mv833c

NA - Used condom every time had sex with 3rd to most recent partner in last 12 m

mv834a

NA - Age of most recent partner

mv834b

NA - Age of 2nd to most recent partner

mv834c

NA - Age of 3rd to most recent partner

mv835a

NA - Alcohol consumption at last sex with most recent partner

mv835b

NA - Alcohol consumption at last sex with 2nd to most recent partner

mv835c

NA - Alcohol consumption at last sex with 3rd to most recent partner

mv836

NA - Total lifetime number of sex partners

mv837

NA - Heard of drugs to help HIV infected people live longer

mv844

NA - Knows someone denied health services for having, or suspected of having, HI

mv845

NA - Knows someone denied social event for having, or suspected of having, HIV;

mv846

NA - Knows someone verbally abused for having, or suspected of having, HIV; in l

mv847

NA - People with HIV should be ashamed of themselves

mv848

NA - People with HIV should be blamed for bringing disease to community

mv849

NA - Children 12-14 should be taught to wait for sex until marriage

mv851a

NA - Young men should wait for sex until marriage

mv851b

NA - Most young men wait for sex until marriage

mv851c

NA - Unmarried sexually active men should have only one partner

mv851d

NA - Most unmarried sexually active men have only one partner

mv851e

NA - Married men should only have sex with their wives

mv851f

NA - Most married men only have sex with their wives

mv851g

NA - Young women should wait for sex until marriage

mv851h

NA - Most young women wait for sex until marriage

mv851i

NA - Unmarried sexually active women should have only one partner

mv851j

NA - Most unmarried sexually active women have only one partner

mv851k

NA - Married women should only have sex with their husbands

mv851l

NA - Most married women only have sex with their husbands

mv852a

NA - How long ago first had sex with most recent partner

mv852b

NA - How long ago first had sex with 2nd most recent partner

mv852c

NA - How long ago first had sex with 3rd most recent partner

mv853a

NA - Times in last 12 months had sex with most recent partner

mv853b

NA - Times in last 12 months had sex with 2nd most recent partner

mv853c

NA - Times in last 12 months had sex with 3rd most recent partner

mv854a

NA - Concurrent sexual partners

mv854b

NA - Cumulative concurrent sexual partners

smdist

District

smlangq

Language of questionnaire

smlangi

Language of interview

smlangr

Language respondent speaks a home

smeduc

Man's educational level for report

smfactt

All men factor - total

smfactu

All men factor - urban/rural

smfactr

All men factor - regional

smfacte

All men factor - educational

smfactw

All men factor - wealth index

smfeduc1

All men factor - education as will be used in the report

smfregur

All woman factor - region combined with urban/rural

sm111

Highest class completed

sm216

Line number of youngest child

sm304a

Family planning message conveyed: limiting the family

sm304b

Family planning message conveyed: higher age at marriage

sm304c

Family planning message conveyed: spacing of children

sm304d

Family planning message conveyed: use of contraceptives

sm304e

Family planning message conveyed: welfare of family

sm304f

Family planning message conveyed: maternal and child health

sm304g

Family planning message conveyed: less children, prosperous life

sm304h

Family planning message conveyed: more children, poverty and starvation

sm304i

Family planning message conveyed: importance of breastfeeding

sm304x

Family planning message conveyed: other

sm304z

Family planning message conveyed: DK/don't remember

sm304b1

Was family planning message effective

sm405

Desire for future child

sm514e

Beating justified: wife neglects in-laws

sm704

Duration of treatment of TB nowadays

sm705

Ever been told by a doctor/nurse that had TB

sm706

Ever heard about Hepatitis

sm707

Person can avoid Hepatitis

sm708a

Avoid getting Hepatitis by: safe sex

sm708b

Avoid getting Hepatitis by: safe blood transfer

sm708c

Avoid getting Hepatitis by: disposable syringe

sm708d

Avoid getting Hepatitis by: avoid contaminated food/water

sm708e

Avoid getting Hepatitis by: avoid contact with infected person

sm708f

Avoid getting Hepatitis by: dentist instruments sterilized

sm708x

Avoid getting Hepatitis by: other

sm708y

Avoid getting Hepatitis by: no response

sm708z

Avoid getting Hepatitis by: don't know

mcaseid

Case Identification

mv000

Country code and phase

mv001

Cluster number

mv002

Household number

mv003

Respondent's line number

mv004

Ultimate area unit

mv005

Men's sample weight (6 decimals)

mv006

Month of interview

mv007

Year of interview

mv008

Date of interview (CMC)

mv009

Month of birth

mv010

Year of birth

mv011

Date of birth (CMC)

mv012

Current age

mv013

Age in 5-year groups

mv014

Completeness of age information

mv015

Result of interview

mv016

Day of interview

mv021

Primary sampling unit

mv022

Sample stratum number

mv023

Sample domain

mv024

Region

mv025

Type of place of residence

mv026

De facto place of residence

mv027

Number of visits

mv028

Interviewer identification

mv029

Keyer identification

mv030

Field supervisor

mv031

Field editor

mv032

NA - Office editor

mv034.1

NA - Line number of wife/partner

mv034.2

NA - Line number of wife/partner

mv034.3

NA - Line number of wife/partner

mv034.4

NA - Line number of wife/partner

mv034.5

NA - Line number of wife/partner

mv034.6

NA - Line number of wife/partner

mv034.7

NA - Line number of wife/partner

mv034.8

NA - Line number of wife/partner

mv034a.1

NA - Wife or partner***

mv034a.2

NA - Wife or partner***

mv034a.3

NA - Wife or partner***

mv034a.4

NA - Wife or partner***

mv034a.5

NA - Wife or partner***

mv034a.6

NA - Wife or partner***

mv034a.7

NA - Wife or partner***

mv034a.8

NA - Wife or partner***

mv034b.1

NA - Age of wife/partner

mv034b.2

NA - Age of wife/partner

mv034b.3

NA - Age of wife/partner

mv034b.4

NA - Age of wife/partner

mv034b.5

NA - Age of wife/partner

mv034b.6

NA - Age of wife/partner

mv034b.7

NA - Age of wife/partner

mv034b.8

NA - Age of wife/partner

mv035

Number of wives/partners

mv801

Time interview started (hhmm - 24 hour clock)

mv802

Time interview ended (hhmm - 24 hour clock)

mv803

Length of interview in minutes

mv101

Region

mv102

Type of place of residence

mv103

NA - Childhood place of residence***

mv104

NA - Years lived in place of residence

mv105

NA - Type of place of previous residence

mv106

Educational level

mv107

Highest year of education (at level in MV106)

mv130

NA - Religion

mv131

Ethnicity

mv133

Total number of years of education

mv134

De facto place of residence

mv135

Usual resident or visitor

mv136

Number of household members (total listed)

mv138

Number of eligible men in household (de facto)

mv149

Educational attainment

mv150

Relationship to household head

mv151

Sex of household head

mv152

Age of household head

mv155

Literacy

mv156

NA - Ever participated in a literacy program (excluding primary school)

mv157

Frequency of reading newspaper or magazine

mv158

Frequency of listening to radio

mv159

Frequency of watching television

mv167

NA - Times away from home in last 12 months

mv168

NA - Away for more than one month in last 12 months

mv190

Wealth index

mv191

Wealth index factor score (5 decimals)

mv201

Total children ever born

mv202

Sons at home

mv203

Daughters at home

mv204

Sons elsewhere

mv205

Daughters elsewhere

mv206

Sons who have died

mv207

Daughters who have died

mv212

Age of respondent at 1st birth

mv213

Partner currently pregnant

mv217

Knowledge of ovulatory cycle

mv218

Number of living children

mv225

NA - Most recent pregnancy wanted

mv245

Number of women fathered children with

mv246

NA - Married to mother when first child was born

mv247

Age of most recent child

mv248

Antenatal check-ups for the mother of most recent child

mv249

Respondent present during check-ups for most recent child

mv250

Place of birth of most recent child

mv251

NA - Reason for not delivering most recent child in health facility

mv252

Amount a child should drink, when that child has diarrhea

mv301

Knowledge of any contraceptive method

mv302

NA - Ever use of any contraceptive method

m304a.01

Type of contraceptive method

m304a.02

Type of contraceptive method

m304a.03

Type of contraceptive method

m304a.04

Type of contraceptive method

m304a.05

Type of contraceptive method

m304a.06

Type of contraceptive method

m304a.07

Type of contraceptive method

m304a.08

Type of contraceptive method

m304a.09

Type of contraceptive method

m304a.10

Type of contraceptive method

m304a.11

Type of contraceptive method

m304a.12

Type of contraceptive method

m304a.13

Type of contraceptive method

m304a.14

Type of contraceptive method

m304a.15

Type of contraceptive method

m304a.16

Type of contraceptive method

m304a.17

Type of contraceptive method

m304a.18

Type of contraceptive method

m304a.19

Type of contraceptive method

m304a.20

Type of contraceptive method

mv304.01

Knows contraceptive method

mv304.02

Knows contraceptive method

mv304.03

Knows contraceptive method

mv304.04

Knows contraceptive method

mv304.05

Knows contraceptive method

mv304.06

Knows contraceptive method

mv304.07

Knows contraceptive method

mv304.08

Knows contraceptive method

mv304.09

Knows contraceptive method

mv304.10

Knows contraceptive method

mv304.11

Knows contraceptive method

mv304.12

Knows contraceptive method

mv304.13

Knows contraceptive method

mv304.14

Knows contraceptive method

mv304.15

Knows contraceptive method

mv304.16

Knows contraceptive method

mv304.17

Knows contraceptive method

mv304.18

Knows contraceptive method

mv304.19

Knows contraceptive method

mv304.20

Knows contraceptive method

mv305.01

Ever used contraceptive method

mv305.02

Ever used contraceptive method

mv305.03

Ever used contraceptive method

mv305.04

Ever used contraceptive method

mv305.05

Ever used contraceptive method

mv305.06

Ever used contraceptive method

mv305.07

Ever used contraceptive method

mv305.08

Ever used contraceptive method

mv305.09

Ever used contraceptive method

mv305.10

Ever used contraceptive method

mv305.11

Ever used contraceptive method

mv305.12

Ever used contraceptive method

mv305.13

Ever used contraceptive method

mv305.14

Ever used contraceptive method

mv305.15

Ever used contraceptive method

mv305.16

Ever used contraceptive method

mv305.17

Ever used contraceptive method

mv305.18

Ever used contraceptive method

mv305.19

Ever used contraceptive method

mv305.20

Ever used contraceptive method

mv307.01

NA - Contraceptive method currently used

mv307.02

NA - Contraceptive method currently used

mv307.03

NA - Contraceptive method currently used

mv307.04

NA - Contraceptive method currently used

mv307.05

NA - Contraceptive method currently used

mv307.06

NA - Contraceptive method currently used

mv307.07

NA - Contraceptive method currently used

mv307.08

NA - Contraceptive method currently used

mv307.09

NA - Contraceptive method currently used

mv307.10

NA - Contraceptive method currently used

mv307.11

NA - Contraceptive method currently used

mv307.12

NA - Contraceptive method currently used

mv307.13

NA - Contraceptive method currently used

mv307.14

NA - Contraceptive method currently used

mv307.15

NA - Contraceptive method currently used

mv307.16

NA - Contraceptive method currently used

mv307.17

NA - Contraceptive method currently used

mv307.18

NA - Contraceptive method currently used

mv307.19

NA - Contraceptive method currently used

mv307.20

NA - Contraceptive method currently used

mv312

NA - Current contraceptive method

mv313

NA - Current contraceptive by method type

mv323a

NA - Brand of condom used

mv325b

NA - Cost of condoms (in CURRENCY) last time obtained

mv372a

NA - Interviewer shown condom package

mv384a

On radio heard about Family Planning in the last few months

mv384b

On TV saw about Family Planning in the last few months

mv384c

In newspaper read about Family Planning in the last few months

mv395

Discussed Family Planning with health worker in last few months

mv396

NA - Possible for a breastfeeding woman to get pregnant

mv3a09b

NA - Number of condoms did get last time

mv3b17

NA - Source of condoms

mv3b25a

Contraception is woman's business, man should not worry

mv3b25b

Women who use contraception become promiscuous

mv463a

Smokes cigarettes

mv463b

Smokes pipe

mv463c

Uses chewing tobacco

mv463d

Uses snuff

mv463e

Smokes hukaa/sheesha

mv463f

NA - Smokes CS

mv463g

NA - Smokes CS

mv463x

Smokes other

mv463z

Smokes nothing

mv464

Number of cigarettes in last 24 hours

mv474

Heard of tuberculosis or TB

mv474a

Tuberculosis spread by: air through coughing or sneezing

mv474b

Tuberculosis spread by: sharing utensils

mv474c

Tuberculosis spread by: touching a person with TB

mv474d

Tuberculosis spread by: food

mv474e

Tuberculosis spread by: sexual contact

mv474f

Tuberculosis spread by: mosquito bites

mv474g

NA - Tuberculosis spread by: CS

mv474h

NA - Tuberculosis spread by: CS

mv474i

NA - Tuberculosis spread by: CS

mv474j

NA - Tuberculosis spread by: CS

mv474x

Tuberculosis spread by: other

mv474z

Don't know how tuberculosis is spread

mv475

Tuberculosis can be cured

mv476

NA - Would want tuberculosis infection in family to remain secret

mv477

Number of injections in last 12 months

mv478

Number of injections administered by a health worker

mv479

NA - Place where given injection by health worker

mv480

Syringe and needle from new, unopened package

mv481

NA - Covered by health insurance

mv481a

NA - Health insurance type: mutual/community organization

mv481b

NA - Health insurance type: provided by employer

mv481c

NA - Health insurance type: social security

mv481d

NA - Health insurance type: private/commercially purchased

mv481e

NA - Health insurance type: CS

mv481f

NA - Health insurance type: CS

mv481g

NA - Health insurance type: CS

mv481h

NA - Health insurance type: CS

mv481x

NA - Health insurance type: other

mv482a

NA - Arrange care of biological children under age 18***

mv482b

NA - Primary caregiver of children under age 18***

mv482c

NA - Arrange care of non-biological children < 18***

mv483

NA - Respondent circumcised

mv483a

NA - Age at circumcision

mv483b

NA - Who performed the circumcision

mv483c

NA - Place where circumcision was done

mv501

Current marital status

mv502

Currently/formerly/never in union

mv503

NA - Number of unions

mv504

Currently residing with wife/partner

mv505

Number of wives/partners

mv507

Month of first cohabitation

mv508

Year of first cohabitation

mv509

Date of first cohabitation (CMC)

mv510

Completeness of date information in MV509

mv511

Age at first cohabitation

mv512

Years since first cohabitation

mv513

Cohabitation duration (grouped)

mv525

NA - Age at first sex

mv527

NA - Time since last sex

mv528

NA - Time since last sex (in days)

mv529

NA - Time since last sex (in months)

mv531

NA - Age at first sex (imputed)

mv532

NA - Flag for MV531

mv535

Ever been married or in union

mv536

NA - Recent sexual activity

mv541

NA - Intends to wait until marriage to have sex

mv602

Fertility preference

mv603

Preferred waiting time for birth of a/another child

mv604

Preferred waiting time for birth of a/another child (grouped)

mv605

Desire for more children

mv613

Ideal number of children

mv614

Ideal number of children (grouped)

mv616

Desired time for future birth

mv621

NA - Wife's desire for children***

mv627

Ideal number of boys

mv628

Ideal number of girls

mv629

Ideal number of either sex

mv631

NA - Problem if wife became pregnant***

mv633a

NA - Wife justified refusing sex: husband has STI

mv633b

NA - Wife justified refusing sex: husband has other women

mv633c

NA - Wife justified refusing sex: recent birth

mv633d

NA - Wife justified refusing sex: tired, not in mood

mv633e

NA - Wife justified refusing sex: CS reason

mv633f

NA - Wife justified refusing sex: CS reason

mv633g

NA - Wife justified refusing sex: CS reason

mv634a

NA - Wife refuses to have sex, husband has right to: get angry

mv634b

NA - Wife refuses to have sex, husband has right to: refuse financial support

mv634c

NA - Wife refuses to have sex, husband has right to: use force for sex

mv634d

NA - Wife refuses to have sex, husband has right to: have sex with another women

mv714

Currently working

mv714a

Have a job, but currently absent

mv716

Occupation

mv717

Occupation (grouped)

mv719

NA - Works for family, others, self

mv721

NA - Works at home or away

mv731

Worked in last 12 months

mv732

Employment all year/seasonal

mv739

Person who usually decides how to spend respondent's earnings

mv740

NA - Type of land where respondent works

mv741

Type of earnings from respondent's work

mv743a

Person who should have greater say: respondent's health care

mv743b

Person who should have greater say: large household purchases

mv743c

NA - Person who should have greater say: daily household purchases

mv743d

NA - Person who should have greater say: visits to family or relatives

mv743e

NA - Person who should have greater say: food to be cooked each day

mv743f

NA - Person who should have greater say: deciding what to do with money wife ear

mv743g

NA - Person who should have greater say: deciding how many children to have

mv744a

Beating justified: wife goes out without telling husband

mv744b

Beating justified: wife neglects the children

mv744c

Beating justified: wife argues with husband

mv744d

Beating justified: wife refuses to have sex with husband

mv744e

Beating justified: wife burns food

mv745a

Owns a house alone or jointly

mv745b

Owns land alone or jointly

mv747a

NA - Childbearing is a woman's concern

mv747b

NA - Important for mother/child health to get assistance at delivery

mv750

Ever heard of a Sexually Transmitted Infection (STI)

mv751

Ever heard of AIDS

mv754bp

NA - Reduce risk of getting HIV: do not have sex at all

mv754cp

Reduce risk of getting HIV: always use condoms during sex

mv754dp

Reduce risk of getting HIV: have 1 sex partner only, who has no other partners

mv754jp

Can get HIV from mosquito bites

mv754wp

Can get HIV by sharing food with person who has AIDS

mv756

A healthy looking person can have HIV

mv761

NA - Condom used during last sex with most recent partner

mv761b

NA - Condom used during last sex with 2nd to most recent partner

mv761c

NA - Condom used during last sex with 3rd to most recent partner

mv762aa

Source for condoms: government hospital

mv762ab

Source for condoms: rural health center(RHC)

mv762ac

Source for condoms: family welfare center

mv762ad

Source for condoms: Mother child health centre(MCH)

mv762ae

Source for condoms: dispenser

mv762af

Source for condoms: lady health worker

mv762ag

Source for condoms: Lady health visitor

mv762ah

Source for condoms: male mobilizer

mv762ai

Source for condoms: other public sector

mv762aj

Source for condoms: private hospital/clinic

mv762ak

Source for condoms: pharmacy

mv762al

Source for condoms: private doctor

mv762am

Source for condoms: homeopath

mv762an

Source for condoms: dispenser/compounder

mv762ao

Source for condoms: other private medical

mv762ap

NA - Source for condoms: CS private

mv762aq

NA - Source for condoms: CS private

mv762ar

NA - Source for condoms: CS private

mv762as

Source for condoms: shop

mv762at

Source for condoms: friend/relative

mv762au

Source for condoms: hakim

mv762av

Source for condoms: DAI traditional birth attendant

mv762aw

NA - Source for condoms: CS other

mv762ax

Source for condoms: other

mv762az

Don't know any source for condoms

mv762ba

NA - Source for female condoms: government hospital

mv762bb

NA - Source for female condoms: CS public

mv762bc

NA - Source for female condoms: CS public

mv762bd

NA - Source for female condoms: CS public

mv762be

NA - Source for female condoms: CS public

mv762bf

NA - Source for female condoms: CS public

mv762bg

NA - Source for female condoms: CS public

mv762bh

NA - Source for female condoms: CS public

mv762bi

NA - Source for female condoms: CS public

mv762bj

NA - Source for female condoms: private hospital/clinic

mv762bk

NA - Source for female condoms: pharmacy

mv762bl

NA - Source for female condoms: private doctor

mv762bm

NA - Source for female condoms: CS private

mv762bn

NA - Source for female condoms: CS private

mv762bo

NA - Source for female condoms: CS private

mv762bp

NA - Source for female condoms: CS private

mv762bq

NA - Source for female condoms: CS private

mv762br

NA - Source for female condoms: CS private

mv762bs

NA - Source for female condoms: CS other

mv762bt

NA - Source for female condoms: CS other

mv762bu

NA - Source for female condoms: CS other

mv762bv

NA - Source for female condoms: CS other

mv762bw

NA - Source for female condoms: CS other

mv762bx

NA - Source for female condoms: other

mv762bz

NA - Don't know any source for female condoms

mv763a

Had any STI in last 12 months

mv763b

NA - Had genital sore/ulcer in last 12 months

mv763c

NA - Had genital discharge in last 12 months

mv763d

NA - Had CS STI in last 12 months

mv763e

NA - Had CS STI in last 12 months

mv763f

NA - Had CS STI in last 12 months

mv763g

NA - Had CS STI in last 12 months

mv766a

NA - Number of sex partners, excluding spouse, in last 12 months

mv766b

NA - Number of sex partners, including spouse, in last 12 months

mv767a

NA - Relationship with most recent sex partner

mv767b

NA - Relationship with 2nd to most recent sex partner

mv767c

NA - Relationship with 3rd to most recent sex partner

mv768a

NA - Length of time had sex relations with most recent partner

mv768b

NA - Length of time had sex relations with 2nd to most recent partner

mv768c

NA - Length of time had sex relations with 3rd to most recent partner

mv769

Can get a condom

mv769a

NA - Can get a female condom

mv770

Sought advice/treatment for last STI infection

mv770a

Sought STI advice/treatment from: government hospital

mv770b

Sought STI advice/treatment from: government health center

mv770c

Sought STI advice/treatment from: stand-alone VCT center

mv770d

Sought STI advice/treatment from: family planning clinic

mv770e

Sought STI advice/treatment from: mobile clinic

mv770f

Sought STI advice/treatment from: fieldworker

mv770g

Sought STI advice/treatment from: other public sector

mv770h

NA - Sought STI advice/treatment from: CS public

mv770i

NA - Sought STI advice/treatment from: CS public

mv770j

NA - Sought STI advice/treatment from: CS public

mv770k

Sought STI advice/treatment from: private hospital/clinic/doctor

mv770l

Sought STI advice/treatment from: stand-alone VCT center

mv770m

Sought STI advice/treatment from: pharmacy

mv770n

Sought STI advice/treatment from: mobile clinic

mv770o

Sought STI advice/treatment from: fieldworker

mv770p

Sought STI advice/treatment from: other private medical

mv770q

NA - Sought STI advice/treatment from: CS private

mv770r

NA - Sought STI advice/treatment from: CS private

mv770s

NA - Sought STI advice/treatment from: CS private

mv770t

Sought STI advice/treatment from: shop

mv770u

NA - Sought STI advice/treatment from: CS other

mv770v

NA - Sought STI advice/treatment from: CS other

mv770w

NA - Sought STI advice/treatment from: CS other

mv770x

Sought STI advice/treatment from: other

mv774a

HIV transmitted during pregnancy

mv774b

HIV transmitted during delivery

mv774c

HIV transmitted by breastfeeding

mv775

NA - Knows someone who has, or is suspected of having, HIV

mv777

Would want HIV infection in family to remain secret

mv778

Willing to care for relative with AIDS

mv779

A female teacher infected with HIV, but is not sick, should be allowed to contin

mv780

NA - Children should be taught about condoms to avoid AIDS

mv781

NA - Ever been tested for HIV

mv783

Know a place to get HIV test

mv784a

Place for HIV test: government hospital

mv784b

Place for HIV test: government health center

mv784c

Place for HIV test: stand-alone VCT center

mv784d

Place for HIV test: other public sector

mv784e

NA - Place for HIV test: CS public

mv784f

NA - Place for HIV test: CS public

mv784g

NA - Place for HIV test: CS public

mv784h

NA - Place for HIV test: CS public

mv784i

NA - Place for HIV test: CS public

mv784j

NA - Place for HIV test: CS public

mv784k

Place for HIV test: private hospital/clinic/doctor

mv784l

Place for HIV test: stand-alone VCT center

mv784m

Place for HIV test: other private medical

mv784n

NA - Place for HIV test: CS private

mv784o

NA - Place for HIV test: CS private

mv784p

NA - Place for HIV test: CS private

mv784q

NA - Place for HIV test: CS private

mv784r

NA - Place for HIV test: CS private

mv784s

NA - Place for HIV test: CS other

mv784t

NA - Place for HIV test: CS other

mv784u

NA - Place for HIV test: CS other

mv784v

NA - Place for HIV test: CS other

mv784x

Place for HIV test: other

mv785

Heard about other STIs

mv791

NA - Have ever paid anyone in exchange for sex

mv793

NA - Paid for sex in last 12 months

mv793a

NA - Condom used last time paid for sex in last 12 months

mv793b

NA - Condom used every time paid for sex in last 12 months

mv820

NA - Condom used at first sex

mv822

NA - Wife justified asking husband to use condom if he has STI

mv823

Can get HIV by witchcraft or supernatural means

mv824

Drugs to avoid HIV transmission to baby during pregnancy

mv825

Would buy vegetables from vendor with HIV

mv826

NA - Last time tested for HIV

mv826a

NA - Months ago most recent HIV test

mv827

NA - Last HIV test: on your own, offered or required

mv828

NA - Received result from last HIV test

mv829

NA - Place where last HIV test was taken

mv832b

NA - Time since last sex with 2nd to most recent partner

mv832c

NA - Time since last sex with 3rd to most recent partner

mv833a

NA - Used condom every time had sex with most recent partner in last 12 months

mv833b

NA - Used condom every time had sex with 2nd to most recent partner in last 12 m

mv833c

NA - Used condom every time had sex with 3rd to most recent partner in last 12 m

mv834a

NA - Age of most recent partner

mv834b

NA - Age of 2nd to most recent partner

mv834c

NA - Age of 3rd to most recent partner

mv835a

NA - Alcohol consumption at last sex with most recent partner

mv835b

NA - Alcohol consumption at last sex with 2nd to most recent partner

mv835c

NA - Alcohol consumption at last sex with 3rd to most recent partner

mv836

NA - Total lifetime number of sex partners

mv837

NA - Heard of drugs to help HIV infected people live longer

mv844

NA - Knows someone denied health services for having, or suspected of having, HI

mv845

NA - Knows someone denied social event for having, or suspected of having, HIV;

mv846

NA - Knows someone verbally abused for having, or suspected of having, HIV; in l

mv847

NA - People with HIV should be ashamed of themselves

mv848

NA - People with HIV should be blamed for bringing disease to community

mv849

NA - Children 12-14 should be taught to wait for sex until marriage

mv851a

NA - Young men should wait for sex until marriage

mv851b

NA - Most young men wait for sex until marriage

mv851c

NA - Unmarried sexually active men should have only one partner

mv851d

NA - Most unmarried sexually active men have only one partner

mv851e

NA - Married men should only have sex with their wives

mv851f

NA - Most married men only have sex with their wives

mv851g

NA - Young women should wait for sex until marriage

mv851h

NA - Most young women wait for sex until marriage

mv851i

NA - Unmarried sexually active women should have only one partner

mv851j

NA - Most unmarried sexually active women have only one partner

mv851k

NA - Married women should only have sex with their husbands

mv851l

NA - Most married women only have sex with their husbands

mv852a

NA - How long ago first had sex with most recent partner

mv852b

NA - How long ago first had sex with 2nd most recent partner

mv852c

NA - How long ago first had sex with 3rd most recent partner

mv853a

NA - Times in last 12 months had sex with most recent partner

mv853b

NA - Times in last 12 months had sex with 2nd most recent partner

mv853c

NA - Times in last 12 months had sex with 3rd most recent partner

mv854a

NA - Concurrent sexual partners

mv854b

NA - Cumulative concurrent sexual partners

smdist

District

smlangq

Language of questionnaire

smlangi

Language of interview

smlangr

Language respondent speaks a home

smeduc

Man's educational level for report

smfactt

All men factor - total

smfactu

All men factor - urban/rural

smfactr

All men factor - regional

smfacte

All men factor - educational

smfactw

All men factor - wealth index

smfeduc1

All men factor - education as will be used in the report

smfregur

All woman factor - region combined with urban/rural

sm111

Highest class completed

sm216

Line number of youngest child

sm304a

Family planning message conveyed: limiting the family

sm304b

Family planning message conveyed: higher age at marriage

sm304c

Family planning message conveyed: spacing of children

sm304d

Family planning message conveyed: use of contraceptives

sm304e

Family planning message conveyed: welfare of family

sm304f

Family planning message conveyed: maternal and child health

sm304g

Family planning message conveyed: less children, prosperous life

sm304h

Family planning message conveyed: more children, poverty and starvation

sm304i

Family planning message conveyed: importance of breastfeeding

sm304x

Family planning message conveyed: other

sm304z

Family planning message conveyed: DK/don't remember

sm304b1

Was family planning message effective

sm405

Desire for future child

sm514e

Beating justified: wife neglects in-laws

sm704

Duration of treatment of TB nowadays

sm705

Ever been told by a doctor/nurse that had TB

sm706

Ever heard about Hepatitis

sm707

Person can avoid Hepatitis

sm708a

Avoid getting Hepatitis by: safe sex

sm708b

Avoid getting Hepatitis by: safe blood transfer

sm708c

Avoid getting Hepatitis by: disposable syringe

sm708d

Avoid getting Hepatitis by: avoid contaminated food/water

sm708e

Avoid getting Hepatitis by: avoid contact with infected person

sm708f

Avoid getting Hepatitis by: dentist instruments sterilized

sm708x

Avoid getting Hepatitis by: other

sm708y

Avoid getting Hepatitis by: no response

sm708z

Avoid getting Hepatitis by: don't know

References

National Institute of Population Studies (NIPS), Pakistan (http://nips.org.pk/Home.htm).

Examples

load("PKMR61")

#> Warning: cannot open compressed file 'PKMR61', probable reason 'No such file or directory'

#> Error in readChar(con, 5L, useBytes = TRUE): cannot open the connection

PKMR61

#> # A tibble: 3,134 x 549
#>    mcaseid mv000 mv001 mv002 mv003 mv004  mv005 mv006 mv007 mv008 mv009 mv010
#>    <chr>   <chr> <int> <int> <int> <int>  <int> <int> <int> <int> <int> <int>
#>  1 1 2  1  PK6       1     2     1     1 2.25e6    11  2012  1355     7  1984
#>  2 1 2  9  PK6       1     2     9     1 2.25e6    11  2012  1355     1  1989
#>  3 1 2 11  PK6       1     2    11     1 2.25e6    11  2012  1355     9  1992
#>  4 1 5  1  PK6       1     5     1     1 2.25e6    11  2012  1355     5  1979
#>  5 1 8  4  PK6       1     8     4     1 2.25e6    11  2012  1355     1  1971
#>  6 1 8  7  PK6       1     8     7     1 2.25e6    11  2012  1355     3  1977
#>  7 110  1  PK6       1    10     1     1 2.25e6    11  2012  1355    10  1970
#>  8 119  1  PK6       1    19     1     1 2.25e6    11  2012  1355     2  1965
#>  9 2 6  1  PK6       2     6     1     2 2.05e6    11  2012  1355     1  1982
#> 10 2 9  1  PK6       2     9     1     2 2.05e6    11  2012  1355     4  1967
#> # … with 3,124 more rows, and 537 more variables: mv011 <int>, mv012 <int>,
#> #   mv013 <fct>, mv014 <fct>, mv015 <fct>, mv016 <int>, mv021 <int>,
#> #   mv022 <int>, mv023 <fct>, mv024 <fct>, mv025 <fct>, mv026 <fct>,
#> #   mv027 <int>, mv028 <int>, mv029 <int>, mv030 <int>, mv031 <int>,
#> #   mv032 <int>, mv034.1 <fct>, mv034.2 <fct>, mv034.3 <fct>, mv034.4 <fct>,
#> #   mv034.5 <fct>, mv034.6 <fct>, mv034.7 <fct>, mv034.8 <fct>, mv034a.1 <fct>,
#> #   mv034a.2 <fct>, mv034a.3 <fct>, mv034a.4 <fct>, mv034a.5 <fct>,
#> #   mv034a.6 <fct>, mv034a.7 <fct>, mv034a.8 <fct>, mv034b.1 <int>,
#> #   mv034b.2 <int>, mv034b.3 <int>, mv034b.4 <int>, mv034b.5 <int>,
#> #   mv034b.6 <int>, mv034b.7 <int>, mv034b.8 <int>, mv035 <int>, mv801 <int>,
#> #   mv802 <int>, mv803 <int>, mv101 <fct>, mv102 <fct>, mv103 <fct>,
#> #   mv104 <fct>, mv105 <fct>, mv106 <fct>, mv107 <int>, mv130 <fct>,
#> #   mv131 <int>, mv133 <int>, mv134 <fct>, mv135 <fct>, mv136 <int>,
#> #   mv138 <int>, mv149 <fct>, mv150 <fct>, mv151 <fct>, mv152 <int>,
#> #   mv155 <int>, mv156 <fct>, mv157 <int>, mv158 <int>, mv159 <int>,
#> #   mv167 <fct>, mv168 <fct>, mv190 <fct>, mv191 <int>, mv201 <int>,
#> #   mv202 <int>, mv203 <int>, mv204 <int>, mv205 <int>, mv206 <int>,
#> #   mv207 <int>, mv212 <int>, mv213 <int>, mv217 <int>, mv218 <int>,
#> #   mv225 <fct>, mv245 <int>, mv246 <fct>, mv247 <int>, mv248 <int>,
#> #   mv249 <fct>, mv250 <int>, mv251 <fct>, mv252 <int>, mv301 <fct>,
#> #   mv302 <fct>, m304a.01 <fct>, m304a.02 <fct>, m304a.03 <fct>,
#> #   m304a.04 <fct>, m304a.05 <fct>, …