Pakistan Demographic and Health Survey 2017-18 Men's Recode Data

PKMR71 data of Men's Recode from Pakistan Demographic and Health Survey 2017-18.

MCASEID

Case Identification

MV000

Country code and phase

MV001

Cluster number

MV002

Household number

MV003

Respondent's line number

MV004

Ultimate area unit

MV005

Men's sample weight (6 decimals)

MV006

Month of interview

MV007

Year of interview

MV008

Date of interview (CMC)

MV008A

Date of interview Century Day Code (CDC)

MV009

Month of birth

MV010

Year of birth

MV011

Date of birth (CMC)

MV012

Current age

MV013

Age in 5-year groups

MV014

Completeness of age information

MV015

Result of interview

MV016

Day of interview

MV021

Primary sampling unit

MV022

Sample stratum number

MV023

Sample domain

MV024

Region

MV025

Type of place of residence

MV026

NA - De facto place of residence

MV027

Number of visits

MV028

Interviewer identification

MV029

Keyer identification

MV030

Field supervisor

MV031

Field editor

MV032

NA - Office editor

MV034.1

Line number of wife/partner

MV034.2

Line number of wife/partner

MV034.3

Line number of wife/partner

MV034.4

Line number of wife/partner

MV034.5

Line number of wife/partner

MV034.6

Line number of wife/partner

MV034.7

Line number of wife/partner

MV034.8

Line number of wife/partner

MV034A.1

NA - Wife or partner

MV034A.2

NA - Wife or partner

MV034A.3

NA - Wife or partner

MV034A.4

NA - Wife or partner

MV034A.5

NA - Wife or partner

MV034A.6

NA - Wife or partner

MV034A.7

NA - Wife or partner

MV034A.8

NA - Wife or partner

MV034B.1

Age of wife/partner

MV034B.2

Age of wife/partner

MV034B.3

Age of wife/partner

MV034B.4

Age of wife/partner

MV034B.5

Age of wife/partner

MV034B.6

Age of wife/partner

MV034B.7

Age of wife/partner

MV034B.8

Age of wife/partner

MV035

Number of wives/partners

MV045A

Language of questionnaire

MV045B

Language of interview

MV045C

Native language of respondent

MV046

Translator used

MV801

Time interview started (hhmm - 24 hour clock)

MV802

Time interview ended (hhmm - 24 hour clock)

MV803

Length of interview in minutes

MV101

Region

MV102

Type of place of residence

MV103

NA - Childhood place of residence

MV104

NA - Years lived in place of residence

MV105

NA - Type of place of previous residence

MV105A

NA - Region of previous residence

MV106

Educational level

MV107

Highest year of education (at level in MV106)

MV130

NA - Religion

MV131

NA - Ethnicity

MV133

Total number of years of education

MV134

NA - De facto place of residence

MV135

Usual resident or visitor

MV136

Number of household members (total listed)

MV138

Number of eligible men in household (de facto)

MV149

Educational attainment

MV150

Relationship to household head

MV151

Sex of household head

MV152

Age of household head

MV155

Literacy

MV156

NA - Ever participated in a literacy program (excluding primary school)

MV157

Frequency of reading newspaper or magazine

MV158

Frequency of listening to radio

MV159

Frequency of watching television

MV167

NA - Times away from home in last 12 months

MV168

NA - Away for more than one month in last 12 months

MV169A

Owns a mobile telephone

MV169B

Use mobile telephone for financial transactions

MV170

Has an account in a bank or other financial institution

MV171A

Use of internet

MV171B

Frequency of using internet last month

MV190

Wealth index combined

MV191

Wealth index factor score combined (5 decimals)

MV190A

Wealth index for urban/rural

MV191A

Wealth index factor score for urban/rural (5 decimals)

MV201

Total children ever born

MV202

Sons at home

MV203

Daughters at home

MV204

Sons elsewhere

MV205

Daughters elsewhere

MV206

Sons who have died

MV207

Daughters who have died

MV212

Age of respondent at 1st birth

MV213

Partner currently pregnant

MV217

Knowledge of ovulatory cycle

MV218

Number of living children

MV225

NA - Most recent pregnancy wanted

MV245

Number of women fathered children with

MV246

NA - Married to mother when first child was born

MV247

Age of most recent child

MV248

Antenatal check-ups for the mother of most recent child

MV249

Respondent present during check-ups for most recent child

MV250

Place of birth of most recent child

MV251

NA - Reason for not delivering most recent child in health facility

MV252

Amount a child should drink, when that child has diarrhea

MV301

Knowledge of any contraceptive method

MV302

NA - Ever use of any contraceptive method

M304A.01

Type of contraceptive method

M304A.02

Type of contraceptive method

M304A.03

Type of contraceptive method

M304A.04

Type of contraceptive method

M304A.05

Type of contraceptive method

M304A.06

Type of contraceptive method

M304A.07

Type of contraceptive method

M304A.08

Type of contraceptive method

M304A.09

Type of contraceptive method

M304A.10

Type of contraceptive method

M304A.11

Type of contraceptive method

M304A.12

Type of contraceptive method

M304A.13

Type of contraceptive method

M304A.14

Type of contraceptive method

M304A.15

Type of contraceptive method

M304A.16

Type of contraceptive method

M304A.17

Type of contraceptive method

M304A.18

Type of contraceptive method

M304A.19

Type of contraceptive method

M304A.20

Type of contraceptive method

MV304.01

Knows contraceptive method

MV304.02

Knows contraceptive method

MV304.03

Knows contraceptive method

MV304.04

Knows contraceptive method

MV304.05

Knows contraceptive method

MV304.06

Knows contraceptive method

MV304.07

Knows contraceptive method

MV304.08

Knows contraceptive method

MV304.09

Knows contraceptive method

MV304.10

Knows contraceptive method

MV304.11

Knows contraceptive method

MV304.12

Knows contraceptive method

MV304.13

Knows contraceptive method

MV304.14

Knows contraceptive method

MV304.15

Knows contraceptive method

MV304.16

Knows contraceptive method

MV304.17

Knows contraceptive method

MV304.18

Knows contraceptive method

MV304.19

Knows contraceptive method

MV304.20

Knows contraceptive method

MV305.01

NA - Ever used contraceptive method

MV305.02

NA - Ever used contraceptive method

MV305.03

NA - Ever used contraceptive method

MV305.04

NA - Ever used contraceptive method

MV305.05

NA - Ever used contraceptive method

MV305.06

NA - Ever used contraceptive method

MV305.07

NA - Ever used contraceptive method

MV305.08

NA - Ever used contraceptive method

MV305.09

NA - Ever used contraceptive method

MV305.10

NA - Ever used contraceptive method

MV305.11

NA - Ever used contraceptive method

MV305.12

NA - Ever used contraceptive method

MV305.13

NA - Ever used contraceptive method

MV305.14

NA - Ever used contraceptive method

MV305.15

NA - Ever used contraceptive method

MV305.16

NA - Ever used contraceptive method

MV305.17

NA - Ever used contraceptive method

MV305.18

NA - Ever used contraceptive method

MV305.19

NA - Ever used contraceptive method

MV305.20

NA - Ever used contraceptive method

MV307.01

Contraceptive method currently used

MV307.02

Contraceptive method currently used

MV307.03

Contraceptive method currently used

MV307.04

Contraceptive method currently used

MV307.05

Contraceptive method currently used

MV307.06

Contraceptive method currently used

MV307.07

Contraceptive method currently used

MV307.08

Contraceptive method currently used

MV307.09

Contraceptive method currently used

MV307.10

Contraceptive method currently used

MV307.11

Contraceptive method currently used

MV307.12

Contraceptive method currently used

MV307.13

Contraceptive method currently used

MV307.14

Contraceptive method currently used

MV307.15

Contraceptive method currently used

MV307.16

Contraceptive method currently used

MV307.17

Contraceptive method currently used

MV307.18

Contraceptive method currently used

MV307.19

Contraceptive method currently used

MV307.20

Contraceptive method currently used

MV312

Current contraceptive method

MV313

Current contraceptive by method type

MV323A

NA - Brand of condom used

MV325B

NA - Cost of condoms (in CURRENCY) last time obtained

MV372A

NA - Interviewer shown condom package

MV384A

On radio heard about Family Planning in the last few months

MV384B

On TV saw about Family Planning in the last few months

MV384C

In newspaper read about Family Planning in the last few months

MV384D

Heard family planning by text messages on mobile phone

MV395

Discussed Family Planning with health worker in last few months

MV396

NA - Possible for a breastfeeding woman to get pregnant

MV396A

Can women get pregnant after birth of a child before period return

MV3A00Y

NA - Source of family planning for non-users: no source

MV3A00Z

NA - Source of family planning for non-users: any source

MV3A09B

NA - Number of condoms did get last time

MV3B17

NA - Source of condoms

MV3B25A

Contraception is woman's business, man should not worry

MV3B25B

Women who use contraception become promiscuous

MV463A

NA - Smokes cigarettes

MV463B

NA - Smokes pipe

MV463C

NA - Uses chewing tobacco

MV463D

NA - Uses snuff

MV463E

NA - Smokes/uses CS

MV463F

NA - Smokes/uses CS

MV463G

NA - Smokes/uses CS

MV463X

NA - Smokes/uses other

MV463Z

NA - Smokes nothing

MV463AA

Frequency currently smokes tobacco

MV463AB

Frequency currently uses smokeless tobacco

MV463AC

In the past smoked tobacco every day

MV463AD

Frequency in the past smoked tobacco

MV464

NA - Number of cigarettes in last 24 hours

MV464A

On average respondent smokes daily: manufactured cigarettes

MV464B

On average respondent smokes daily: hand roll cigarettes

MV464C

On average respondent smokes daily: kreteks

MV464D

On average respondent smokes daily: pipes full of tobacco

MV464E

On average respondent smokes daily: cigars, cheroots, cigarillos

MV464F

On average respondent smokes daily: water pipe sessions

MV464G

On average respondent smokes daily: others

MV464H

On average respondent uses daily: snuff by mouth

MV464I

On average respondent uses daily: snuff by nose

MV464J

On average respondent uses daily: chewing tobacco

MV464K

On average respondent uses daily: betel quid with tobacco

MV464L

On average respondent uses daily: any others

MV474

Heard of tuberculosis or TB

MV474A

Tuberculosis spread by: air through coughing or sneezing

MV474B

Tuberculosis spread by: sharing utensils

MV474C

Tuberculosis spread by: touching a person with TB

MV474D

Tuberculosis spread by: food

MV474E

Tuberculosis spread by: sexual contact

MV474F

Tuberculosis spread by: mosquito bites

MV474G

NA - Tuberculosis spread by: CS

MV474H

NA - Tuberculosis spread by: CS

MV474I

NA - Tuberculosis spread by: CS

MV474J

NA - Tuberculosis spread by: CS

MV474X

Tuberculosis spread by: other

MV474Z

Don't know how tuberculosis is spread

MV475

Tuberculosis can be cured

MV476

NA - Would want tuberculosis infection in family to remain secret

MV477

Number of injections in last 12 months

MV478

Number of injections administered by a health worker

MV479

NA - Place where given injection by health worker

MV480

Syringe and needle from new, unopened package

MV481

Covered by health insurance

MV481A

Health insurance type: mutual/community organization

MV481B

Health insurance type: provided by employer

MV481C

Health insurance type: social security (sehat sahulat)

MV481D

Health insurance type: private/commercially purchased

MV481E

NA - Health insurance type: CS

MV481F

NA - Health insurance type: CS

MV481G

NA - Health insurance type: CS

MV481H

NA - Health insurance type: CS

MV481X

Health insurance type: other

MV482A

NA - Arrange care of biological children under age 18

MV482B

NA - Primary caregiver of children under age 18

MV482C

NA - Arrange care of non-biological children < 18

MV483

NA - Respondent circumcised

MV483A

NA - Age at circumcision

MV483B

NA - Who performed the circumcision

MV483C

NA - Place where circumcision was done

MV484A

On average respondent smokes weekly: manufactured cigarettes

MV484B

On average respondent smokes weekly: hand roll cigarettes

MV484C

On average respondent smokes weekly: kreteks

MV484D

On average respondent smokes weekly: pipes full of tobacco

MV484E

On average respondent smokes weekly: cigars, cheroots, cigarillos

MV484F

On average respondent smokes weekly: water pipe sessions

MV484G

On average respondent smokes weekly: others

MV484H

On average respondent uses weekly: snuff by mouth

MV484I

On average respondent uses weekly: snuff by nose

MV484J

On average respondent uses weekly: chewing tobacco

MV484K

On average respondent uses weekly: betel quid with tobacco

MV484L

On average respondent uses weekly: any others

MV501

Current marital status

MV502

Currently/formerly/never in union

MV503

Number of unions

MV504

Currently residing with wife/partner

MV505

Number of wives/partners

MV507

Month of first cohabitation

MV508

Year of first cohabitation

MV509

Date of first cohabitation (CMC)

MV510

Completeness of date information in MV509

MV511

Age at first cohabitation

MV512

Years since first cohabitation

MV513

Cohabitation duration (grouped)

MV525

Age at first sex

MV527

Time since last sex

MV528

Time since last sex (in days)

MV529

Time since last sex (in months)

MV531

Age at first sex (imputed)

MV532

Flag for MV531

MV535

NA - Ever been married or in union

MV536

Recent sexual activity

MV541

NA - Intends to wait until marriage to have sex

MV602

Fertility preference

MV603

Preferred waiting time for birth of a/another child

MV604

Preferred waiting time for birth of a/another child (grouped)

MV605

Desire for more children

MV613

Ideal number of children

MV614

Ideal number of children (grouped)

MV616

Desired time for future birth

MV621

NA - Wife's desire for children

MV627

Ideal number of boys

MV628

Ideal number of girls

MV629

Ideal number of either sex

MV631

NA - Problem if wife became pregnant

MV633A

NA - Wife justified refusing sex: husband has STI

MV633B

Wife justified refusing sex: husband has other women

MV633C

NA - Wife justified refusing sex: recent birth

MV633D

NA - Wife justified refusing sex: tired, not in mood

MV633E

NA - Wife justified refusing sex: CS reason

MV633F

NA - Wife justified refusing sex: CS reason

MV633G

NA - Wife justified refusing sex: CS reason

MV634A

NA - Wife refuses to have sex, husband has right to: get angry

MV634B

NA - Wife refuses to have sex, husband has right to: refuse financial support

MV634C

NA - Wife refuses to have sex, husband has right to: use force for sex

MV634D

NA - Wife refuses to have sex, husband has right to: have sex with another women

MV714

Currently working

MV714A

Have a job, but currently absent

MV716

Occupation

MV717

Occupation (grouped)

MV719

NA - Works for family, others, self

MV721

NA - Works at home or away

MV731

Worked in last 12 months

MV732

Employment all year/seasonal

MV739

Person who usually decides how to spend respondent's earnings

MV740

NA - Type of land where respondent works

MV741

Type of earnings from respondent's work

MV743A

Person who should have greater say: respondent's health care

MV743B

Person who should have greater say: large household purchases

MV743C

NA - Person who should have greater say: daily household purchases

MV743D

NA - Person who should have greater say: visits to family or relatives

MV743E

NA - Person who should have greater say: food to be cooked each day

MV743F

NA - Person who should have greater say: deciding what to do with money wife earns

MV743G

NA - Person who should have greater say: deciding how many children to have

MV744A

Beating justified: wife goes out without telling husband

MV744B

Beating justified: wife neglects the children

MV744C

Beating justified: wife argues with husband

MV744D

Beating justified: wife refuses to have sex with husband

MV744E

Beating justified: wife burns food

MV745A

Owns a house alone or jointly

MV745B

Owns land alone or jointly

MV745C

Title deed on house own by respondent

MV745D

Title deed on land own by respondent

MV747A

NA - Childbearing is a woman's concern

MV747B

NA - Important for mother/child health to get assistance at delivery

MV750

Ever heard of a Sexually Transmitted Infection (STI)

MV751

Ever heard of AIDS

MV754BP

NA - Reduce risk of getting HIV: do not have sex at all

MV754CP

Reduce risk of getting HIV: always use condoms during sex

MV754DP

Reduce risk of getting HIV: have 1 sex partner only, who has no other partners

MV754JP

Can get HIV from mosquito bites

MV754WP

Can get HIV by sharing food with person who has AIDS

MV756

A healthy looking person can have HIV

MV761

NA - Condom used during last sex with most recent partner

MV761B

NA - Condom used during last sex with 2nd to most recent partner

MV761C

NA - Condom used during last sex with 3rd to most recent partner

MV762

NA - Source of condoms used for last sex

MV762A

NA - Brand of condom used for last sex

MV762AA

NA - Source for condoms: government hospital

MV762AB

NA - Source for condoms: CS public

MV762AC

NA - Source for condoms: CS public

MV762AD

NA - Source for condoms: CS public

MV762AE

NA - Source for condoms: CS public

MV762AF

NA - Source for condoms: CS public

MV762AG

NA - Source for condoms: CS public

MV762AH

NA - Source for condoms: CS public

MV762AI

NA - Source for condoms: CS public

MV762AJ

NA - Source for condoms: private hospital/clinic

MV762AK

NA - Source for condoms: pharmacy

MV762AL

NA - Source for condoms: private doctor

MV762AM

NA - Source for condoms: CS private

MV762AN

NA - Source for condoms: CS private

MV762AO

NA - Source for condoms: CS private

MV762AP

NA - Source for condoms: CS private

MV762AQ

NA - Source for condoms: CS private

MV762AR

NA - Source for condoms: CS private

MV762AS

NA - Source for condoms: CS other

MV762AT

NA - Source for condoms: CS other

MV762AU

NA - Source for condoms: CS other

MV762AV

NA - Source for condoms: CS other

MV762AW

NA - Source for condoms: CS other

MV762AX

NA - Source for condoms: other

MV762AZ

NA - Don't know any source for condoms

MV762BA

NA - Source for female condoms: government hospital

MV762BB

NA - Source for female condoms: CS public

MV762BC

NA - Source for female condoms: CS public

MV762BD

NA - Source for female condoms: CS public

MV762BE

NA - Source for female condoms: CS public

MV762BF

NA - Source for female condoms: CS public

MV762BG

NA - Source for female condoms: CS public

MV762BH

NA - Source for female condoms: CS public

MV762BI

NA - Source for female condoms: CS public

MV762BJ

NA - Source for female condoms: private hospital/clinic

MV762BK

NA - Source for female condoms: pharmacy

MV762BL

NA - Source for female condoms: private doctor

MV762BM

NA - Source for female condoms: CS private

MV762BN

NA - Source for female condoms: CS private

MV762BO

NA - Source for female condoms: CS private

MV762BP

NA - Source for female condoms: CS private

MV762BQ

NA - Source for female condoms: CS private

MV762BR

NA - Source for female condoms: CS private

MV762BS

NA - Source for female condoms: CS other

MV762BT

NA - Source for female condoms: CS other

MV762BU

NA - Source for female condoms: CS other

MV762BV

NA - Source for female condoms: CS other

MV762BW

NA - Source for female condoms: CS other

MV762BX

NA - Source for female condoms: other

MV762BZ

NA - Don't know any source for female condoms

MV763A

Had any STI in last 12 months

MV763B

Had genital sore/ulcer in last 12 months

MV763C

Had genital discharge in last 12 months

MV763D

NA - Had CS STI in last 12 months

MV763E

NA - Had CS STI in last 12 months

MV763F

NA - Had CS STI in last 12 months

MV763G

NA - Had CS STI in last 12 months

MV766A

NA - Number of sex partners, excluding spouse, in last 12 months

MV766B

NA - Number of sex partners, including spouse, in last 12 months

MV767A

NA - Relationship with most recent sex partner

MV767B

NA - Relationship with 2nd to most recent sex partner

MV767C

NA - Relationship with 3rd to most recent sex partner

MV768A

NA - Length of time had sex relations with most recent partner

MV768B

NA - Length of time had sex relations with 2nd to most recent partner

MV768C

NA - Length of time had sex relations with 3rd to most recent partner

MV769

NA - Can get a condom

MV769A

NA - Can get a female condom

MV770

Sought advice/treatment for last STI infection

MV770A

Sought STI advice/treatment from: government hospital

MV770B

Sought STI advice/treatment from: government health center

MV770C

Sought STI advice/treatment from: Stand-alone htc center

MV770D

Sought STI advice/treatment from: Family planning clinic

MV770E

Sought STI advice/treatment from: Mobile htc services

MV770F

Sought STI advice/treatment from: Other public sector

MV770G

NA - Sought STI advice/treatment from: CS public

MV770H

NA - Sought STI advice/treatment from: CS public

MV770I

NA - Sought STI advice/treatment from: CS public

MV770J

NA - Sought STI advice/treatment from: CS public

MV770K

Sought STI advice/treatment from: private hospital/clinic/doctor

MV770L

Sought STI advice/treatment from: Stand-alone htc center

MV770M

Sought STI advice/treatment from: Pharmacy

MV770N

Sought STI advice/treatment from: Other private medical sector

MV770O

NA - Sought STI advice/treatment from: CS private

MV770P

NA - Sought STI advice/treatment from: CS private

MV770Q

NA - Sought STI advice/treatment from: CS private

MV770R

NA - Sought STI advice/treatment from: CS private

MV770S

NA - Sought STI advice/treatment from: CS private

MV770T

Sought STI advice/treatment from: shop

MV770U

NA - Sought STI advice/treatment from: CS other

MV770V

NA - Sought STI advice/treatment from: CS other

MV770W

NA - Sought STI advice/treatment from: CS other

MV770X

Sought STI advice/treatment from: other

MV774A

HIV transmitted during pregnancy

MV774B

HIV transmitted during delivery

MV774C

HIV transmitted by breastfeeding

MV775

NA - Knows someone who has, or is suspected of having, HIV

MV777

NA - Would want HIV infection in family to remain secret

MV777A

Would be ashamed if someone in the family had HIV

MV778

NA - Willing to care for relative with AIDS

MV779

NA - A female teacher infected with HIV, but is not sick, should be allowed to continue teaching

MV780

NA - Children should be taught about condoms to avoid AIDS

MV781

Ever been tested for HIV

MV783

Know a place to get HIV test

MV784A

Place for HIV test: government hospital

MV784B

Place for HIV test: government health center

MV784C

Place for HIV test: stand-alone htc center

MV784D

Place for HIV test: other public sector

MV784E

NA - Place for HIV test: CS public

MV784F

NA - Place for HIV test: CS public

MV784G

NA - Place for HIV test: CS public

MV784H

NA - Place for HIV test: CS public

MV784I

NA - Place for HIV test: CS public

MV784J

NA - Place for HIV test: CS public

MV784K

Place for HIV test: private hospital/clinic/doctor

MV784L

Place for HIV test: stand-alone htc center

MV784M

Place for HIV test: pharmacy

MV784N

Place for HIV test: other private medical sector

MV784O

NA - Place for HIV test: CS private

MV784P

NA - Place for HIV test: CS private

MV784Q

NA - Place for HIV test: CS private

MV784R

NA - Place for HIV test: CS private

MV784S

NA - Place for HIV test: CS other

MV784T

NA - Place for HIV test: CS other

MV784U

NA - Place for HIV test: CS other

MV784V

NA - Place for HIV test: CS other

MV784X

Place for HIV test: other

MV785

Heard about other STIs

MV791

NA - Have ever paid anyone in exchange for sex

MV791A

NA - Last 12 months have given gifts or other goods to have sex

MV791B

NA - Have ever given gifts or other goods to have sex

MV793

NA - Paid for sex in last 12 months

MV793A

NA - Condom used last time paid for sex in last 12 months

MV793B

NA - Condom used every time paid for sex in last 12 months

MV820

NA - Condom used at first sex

MV822

Wife justified asking husband to use condom if he has STI

MV823

Can get HIV by witchcraft or supernatural means

MV824

Drugs to avoid HIV transmission to baby during pregnancy

MV825

Would buy vegetables from vendor with HIV

MV826

NA - Last time tested for HIV

MV826A

Months ago most recent HIV test

MV827

NA - Last HIV test: on your own, offered or required

MV828

Received result from last HIV test

MV829

Place where last HIV test was taken

MV832B

NA - Time since last sex with 2nd to most recent partner

MV832C

NA - Time since last sex with 3rd to most recent partner

MV833A

NA - Used condom every time had sex with most recent partner in last 12 months

MV833B

NA - Used condom every time had sex with 2nd to most recent partner in last 12 months

MV833C

NA - Used condom every time had sex with 3rd to most recent partner in last 12 months

MV834A

NA - Age of most recent partner

MV834B

NA - Age of 2nd to most recent partner

MV834C

NA - Age of 3rd to most recent partner

MV835A

NA - Alcohol consumption at last sex with most recent partner

MV835B

NA - Alcohol consumption at last sex with 2nd to most recent partner

MV835C

NA - Alcohol consumption at last sex with 3rd to most recent partner

MV836

NA - Total lifetime number of sex partners

MV837

NA - Heard of drugs to help HIV infected people live longer

MV844

NA - Knows someone denied health services for having, or suspected of having, HIV; in last 12 months

MV845

NA - Knows someone denied social event for having, or suspected of having, HIV; in last 12 months

MV846

NA - Knows someone verbally abused for having, or suspected of having, HIV; in last 12 months

MV847

NA - People with HIV should be ashamed of themselves

MV848

NA - People with HIV should be blamed for bringing disease to community

MV849

NA - Children 12-14 should be taught to wait for sex until marriage

MV851A

NA - Young men should wait for sex until marriage

MV851B

NA - Most young men wait for sex until marriage

MV851C

NA - Unmarried sexually active men should have only one partner

MV851D

NA - Most unmarried sexually active men have only one partner

MV851E

NA - Married men should only have sex with their wives

MV851F

NA - Most married men only have sex with their wives

MV851G

NA - Young women should wait for sex until marriage

MV851H

NA - Most young women wait for sex until marriage

MV851I

NA - Unmarried sexually active women should have only one partner

MV851J

NA - Most unmarried sexually active women have only one partner

MV851K

NA - Married women should only have sex with their husbands

MV851L

NA - Most married women only have sex with their husbands

MV852A

NA - How long ago first had sex with most recent partner

MV852B

NA - How long ago first had sex with 2nd most recent partner

MV852C

NA - How long ago first had sex with 3rd most recent partner

MV853A

NA - Times in last 12 months had sex with most recent partner

MV853B

NA - Times in last 12 months had sex with 2nd most recent partner

MV853C

NA - Times in last 12 months had sex with 3rd most recent partner

MV854A

NA - Concurrent sexual partners

MV854B

NA - Cumulative concurrent sexual partners

MV856

NA - Knowledge and use of HIV test kits

MV857A

Children with HIV should be allowed to attend school with children without HIV

MV857B

People hesitate to take HIV test because reaction of other people if positive

MV857C

People talk badly about people with or believed to have HIV

MV857D

People with or believed to have HIV lose respect from other people

MV858

Would be afraid to get HIV from contact with saliva from infected person

SMDIST

District

SMEDUC

Man's educational level for report

SMFACTT

All men factor - total

SMFACTU

All men factor - urban/rural

SMFACTR

All men factor - regional

SMFACTE

All men factor - educational

SMFACTW

All men factor - wealth index

SMFEDUC1

All men factor - education as will be used in the report

SMFREGUR

All men actor - region combined with urban/rural

SM109

Highest class completed

SM121A

Mother tongue

SM217

Respondent present during check-ups for most recent child

SM302BBA

Messages conveyed: Limiting the family size

SM302BBB

Messages conveyed: Disadvantages of getting married at young age

SM302BBC

Messages conveyed: Spacing births

SM302BBD

Messages conveyed: Use of contraception

SM302BBE

Messages conveyed: Family welfare

SM302BBF

Messages conveyed: Maternal and child health

SM302BBG

Messages conveyed: Less children means prosperous life

SM302BBH

Messages conveyed: More children means poverty and starvation

SM302BBI

Messages conveyed: Importance of breastfeeding

SM302BBJ

Messages conveyed: Importance of girls' education

SM302BBK

Messages conveyed: Reduction in maternal deaths

SM302BBX

Messages conveyed: Other

SM302BBZ

Messages conveyed: Don't know/do not remember

SM302CC

Message effectivity

SM307AA

Know a place to obtain a method of FP

SM614A

Have autonomy to sell the house

SM617A

Have autonomy to sell the land

SM618F

Beating justification: Neglected In-laws

SM717A

Think there is treatment for HIV

SM717B

Knows where HIV treatment can be received

SM724A

People with HIV get discriminatory treatment

SM812AA

Age started smoking

SM815AA

Currently use any type of drugs

SM815EE

Duration of treatment of TB now a days

SM815FF

Ever been told by a doctor/nurse/Lady health visitor that god forbid have TB

SM815GG

Ever heard of Hepatitis B or C

SM815HH

Can a person do something to avoid getting Hepatitis B or C

SM815IIA

Avoid getting hepatitis B or C: Practice safe sex

SM815IIB

Avoid getting hepatitis B or C: Safe blood transfer

SM815IIC

Avoid getting hepatitis B or C: Use disposable syringe

SM815IID

Avoid getting hepatitis B or C: Avoid contaminated food/water

SM815IIE

Avoid getting hepatitis B or C: Avoid contact with infected person

SM815IIF

Avoid getting hepatitis B or C: Ensure instruments of dentists are properly sterilized

SM815IIX

Avoid getting hepatitis B or C: Other

SM815IIZ

Avoid getting hepatitis B or C: Don't know

SM815JJ

Ever been tested for Hepatitis B or C

SM815KK

Months ago tested for Hepatitis B or C

SM817A

Receive any cash/kind benefit from Benazir Income Support Program

SMV005

Men's sample weight (AJK/GB, 6 decimals)

References

National Institute of Population Studies (NIPS), Pakistan (http://nips.org.pk/Home.htm).

Examples

data("PKMR71")
PKMR71

#> # A tibble: 3,691 x 617
#>    MCASEID MV000 MV001 MV002 MV003 MV004  MV005 MV006 MV007 MV008 MV008A MV009
#>    <fct>   <fct> <int> <int> <int> <int>  <int> <int> <int> <int>  <int> <int>
#>  1 "     … PK7       1     1     1     1 1.00e6     2  2018  1418  43135     9
#>  2 "     … PK7       1     7     1     1 1.00e6     2  2018  1418  43133     5
#>  3 "     … PK7       1    10     1     1 1.00e6     2  2018  1418  43134     6
#>  4 "     … PK7       1    21     3     1 1.00e6     2  2018  1418  43133     9
#>  5 "     … PK7       2     1     1     2 7.99e4     2  2018  1418  43136     3
#>  6 "     … PK7       2     4     1     2 7.99e4     2  2018  1418  43136     8
#>  7 "     … PK7       2    18     4     2 7.99e4     2  2018  1418  43137     4
#>  8 "     … PK7       2    21     1     2 7.99e4     2  2018  1418  43137     6
#>  9 "     … PK7       2    24     1     2 7.99e4     2  2018  1418  43137    11
#> 10 "     … PK7       3     7     1     3 9.14e5     1  2018  1417  43122     3
#> # … with 3,681 more rows, and 605 more variables: MV010 <int>, MV011 <int>,
#> #   MV012 <int>, MV013 <fct>, MV014 <fct>, MV015 <fct>, MV016 <int>,
#> #   MV021 <int>, MV022 <fct>, MV023 <fct>, MV024 <fct>, MV025 <fct>,
#> #   MV026 <fct>, MV027 <int>, MV028 <int>, MV029 <int>, MV030 <int>,
#> #   MV031 <int>, MV032 <int>, MV034.1 <fct>, MV034.2 <fct>, MV034.3 <fct>,
#> #   MV034.4 <fct>, MV034.5 <fct>, MV034.6 <fct>, MV034.7 <fct>, MV034.8 <fct>,
#> #   MV034A.1 <fct>, MV034A.2 <fct>, MV034A.3 <fct>, MV034A.4 <fct>,
#> #   MV034A.5 <fct>, MV034A.6 <fct>, MV034A.7 <fct>, MV034A.8 <fct>,
#> #   MV034B.1 <int>, MV034B.2 <int>, MV034B.3 <int>, MV034B.4 <int>,
#> #   MV034B.5 <int>, MV034B.6 <int>, MV034B.7 <int>, MV034B.8 <int>,
#> #   MV035 <fct>, MV045A <fct>, MV045B <fct>, MV045C <fct>, MV046 <fct>,
#> #   MV801 <int>, MV802 <int>, MV803 <fct>, MV101 <fct>, MV102 <fct>,
#> #   MV103 <fct>, MV104 <fct>, MV105 <fct>, MV105A <fct>, MV106 <fct>,
#> #   MV107 <fct>, MV130 <fct>, MV131 <int>, MV133 <fct>, MV134 <fct>,
#> #   MV135 <fct>, MV136 <int>, MV138 <int>, MV149 <fct>, MV150 <fct>,
#> #   MV151 <fct>, MV152 <fct>, MV155 <fct>, MV156 <fct>, MV157 <fct>,
#> #   MV158 <fct>, MV159 <fct>, MV167 <fct>, MV168 <fct>, MV169A <fct>,
#> #   MV169B <fct>, MV170 <fct>, MV171A <fct>, MV171B <fct>, MV190 <fct>,
#> #   MV191 <int>, MV190A <fct>, MV191A <int>, MV201 <int>, MV202 <int>,
#> #   MV203 <int>, MV204 <int>, MV205 <int>, MV206 <int>, MV207 <int>,
#> #   MV212 <int>, MV213 <fct>, MV217 <fct>, MV218 <int>, MV225 <fct>,
#> #   MV245 <fct>, MV246 <fct>, …