Pakistan Demographic and Health Survey 2006-07 Household Member Recode Data

PKPR53 data of Household Member Recode from Pakistan Demographic and Health Survey 2006-07.

hhid

case identification

hvidx

line number

hv000

country code and phase

hv001

cluster number

hv002

household number

hv003

respondent's line number

hv004

ultimate area unit

hv005

sample weight

hv006

month of interview

hv007

year of interview

hv008

date of interview (cmc)

hv009

number of household members

hv010

number of eligible women in hh

hv011

number of eligible men in hh

hv012

number of de jure members

hv013

number of de facto members

hv014

number of children 5 and under

hv015

result of household interview

hv016

day of interview

hv017

number of visits

hv018

interviewer identification

hv019

keyer identification

hv020

ever-married sample

hv021

primary sampling unit

hv022

sample stratum number

hv023

sample domain

hv024

region

hv025

type of place of residence

hv026

place of residence

hv027

selection for male/husb. int.

hv028

sample weight for male subsamp

hv030

field supervisor

hv031

field editor

hv032

office editor

hv033

ultimate area selection pr-na

hv035

number of eligible children for height & weight

hv040

cluster altitude in meters -na

hv041

total adults measured

hv042

selection for hemoglobin

hv043

selection for women's statu-na

hv044

selection for domestic viol-na

hv201

source of drinking water

hv202

source of non-drinking wate-na

hv204

time to get to water source

hv205

type of toilet facility

hv206

has electricity

hv207

has radio

hv208

has television

hv209

has refrigerator

hv210

has bicycle

hv211

has motorcycle/scooter

hv212

has car/truck

hv213

main floor material

hv214

main wall material

hv215

main roof material

hv216

rooms used for sleeping

hv217

relationship structure

hv218

line number of head of househ.

hv219

sex of head of household

hv220

age of head of household

hv221

has telephone

hv225

share toilet with other households

hv226

type of cooking fuel

hv227

have bednet for sleeping

hv228

children under 5 slept under bednet last night

hv234

test salt for iodine -na

hv235

location of source for wate-na

hv236

person fetching water -na

hv237

anything done to water to make safe to drink

hv237a

water usually treated by: boil

hv237b

water usually treated by: add bleach/chlorine

hv237c

water usually treated by: strain through a cloth

hv237d

water usually treated by: use water filter

hv237e

water usually treated by: solar disinfection

hv237f

water usually treated by: let it stand and settle

hv237g

water usually treated by: c-na

hv237h

water usually treated by: c-na

hv237i

water usually treated by: c-na

hv237j

water usually treated by: c-na

hv237k

water usually treated by: c-na

hv237x

water usually treated by: other

hv237z

water usually treated by: don't know

hv238

number of households sharin-na

hv239

food cooked on stove or ope-na

hv240

household has a chimney, ho-na

hv241

food cooked in the house / -na

hv242

household has separate room-na

hv243a

has a mobile telephone -na

hv243b

has a watch

hv243c

has an animal-drawn cart

hv243d

has a boat with a motor

hv244

own land usable for agriculture

hv245

hectares for agricultural l-na

hv246

livestock, herds or farm animals

hv246a

cattle own -na

hv246b

cows, bulls own

hv246c

horses, donkeys, mules own

hv246d

goats own

hv246e

sheep own

hv246f

chickens own

hv246g

buffalo own

hv246h

camels own

hv246i

cs own -na

hv246j

cs own -na

hv246k

cs own -na

hv247

owns a bank account -na

hv248

number of sick people 18-59-na

hv249

member of the hh died last -na

hv250

number of members who died -na

hv251

number of orphans and vulne-na

hv270

wealth index

hv271

wealth index factor score (5 decimals)

hml1

number of mosquito nets

hml1a

number of mosquito nets with specific information

hml2

number of children under bednet previous night

shdist

district

shtype

type of questionnaire

shv005

hh weight for all households

shv218

line number of head of household (all households).

shv219

sex of head of household (all households)

shv220

age of head of household (all households)

shbirths

total births in household since january 2003

shdeaths

total deaths in household since january 2003

shcdeath

total deaths < 5/sbs from q38

shfdeath

total female deaths age 12-49 from q39

sh107f

room cooler, air conditioner owns

sh107g

washing machine owns

sh107h

water pump owns

sh107i

bed owns

sh107j

chairs owns

sh107k

almirah/cabinet owns

sh107l

clock owns

sh107m

sofa owns

sh107n

sewing machine owns

sh107o

camera owns

sh107p

personal computer owns

sh113

status of house

sh127a

hh uses coil to avoid mosquitos

sh127b

hh uses mats to avoid mosquitos

sh127c

hh uses spray to avoid mosquitos

sh127d

hh uses electric spray repellant to avoid mosquitos

sh127e

hh uses insect repellant to avoid mosquitos

sh127f

hh uses smoke to avoid mosquitos

sh127g

hh uses fan to avoid mosquitos

sh127h

hh uses covered windows/screens to avoid mosquitos

sh127x

hh uses other to avoid mosquitos

sh127y

hh uses nothing to avoid mosquitos

sh128

currently have any medicines for treating malaria in the hou

hv101

relationship to head

hv102

usual resident

hv103

slept last night

hv104

sex of household member

hv105

age of household members

hv106

highest educational level

hv107

highest year of education

hv108

education in single years

hv109

educational attainment

hv110

member still in school -na

hv111

mother alive

hv112

mother's line number -na

hv113

father alive

hv114

father's line number -na

hv115

current marital status

hv116

currently, formerly, never m.

hv117

eligibility for female interview

hv118

eligibility for male interview

hv120

children eligibility for height/weight and hemoglobin

hv121

member attended school during current school year

hv122

educational level during current school year

hv123

grade of education during current school year

hv124

education in single years - current school year

hv125

member attended school during previous school year

hv126

educational level during pr-na

hv127

grade of education during p-na

hv128

education in single years --na

hv129

school attendance status

hv130

member has been very sick f-na

hv131

mother has been very sick f-na

hv132

father has been very sick f-na

hv133

mother/father dead or been -na

hv134

both parents alive -na

hv135

has brothers/sisters under -na

hv136

brothers/sisters under 18 t-na

hv137

member has a blanket -na

hv138

member has a pair of shoes -na

hv139

member has 2+ sets of cloth-na

hv140

member has a birth certificate

idxh4

index to household schedule

sh03

relationship to head

sh11

education levels

sh17

registered with nadra

ha0

na-index to household schedule

ha1

na-women's age in years

ha2

na-respondent's weight (kilos-1d)

ha3

na-respondent's height (cms-1d)

ha4

na-ht/a percentile (resp.)

ha5

na-ht/a standard deviations (resp

ha6

na-ht/a percent ref. median (resp

ha11

na-wt/ht std deviations(resp) dhs

ha12

na-wt/ht percent ref. median (dhs

ha12a

na-wt/ht percent ref. median (fog

ha12b

na-wt/ht percent ref. median (who

ha13

na-result of measurement of resp.

ha32

na-date of birth (cmc)

ha33

na-completeness of information

ha35

na-smoking

ha40

na-body mass index for respondent

ha41

na-rohrer's index for respondent

ha50

na-under age 18

ha51

na-line no. of parent/caretaker

ha52

na-read consent statement

ha53

na-hemoglobin level (g/dl - 1 decimal)

ha54

na-currently pregnant

ha55

na-result of measuring (hemoglobin)

ha56

na-hemoglobin level adjusted by altitude (g/dl - 1 decimal)

ha57

na-anemia level

ha58

na-agrees to referral

ha60

na-marital status

ha61

na-read consent statement (hiv)

ha62

na-blood sample id number

ha63

na-result of measuring (hiv)

ha64

na-consent for additional tests

ha65

na-result of woman individual interview

ha66

na-woman's highest educational level

ha67

na-woman's highest year of education

ha68

na-highest educational level (cs for prelimnary & final repo

ha69

na-hiv weight (6 decimals)

hc0

na-index to household schedule

hc1

na-age in months

hc2

na-weight in kilograms (1 dec.)

hc3

na-height in centimeters (1 dec.)

hc4

na-ht/a percentile

hc5

na-ht/a standard deviations

hc6

na-ht/a percent of ref. median

hc7

na-wt/a percentile

hc8

na-wt/a standard deviations

hc9

na-wt/a percent of ref. median

hc10

na-wt/ht percentile

hc11

na-wt/ht standard deviations

hc12

na-wt/ht percent of ref. median

hc13

na-reason not measured

hc15

na-height: lying or standing

hc16

na-day of birth of child

hc17

na-date measured (day)

hc18

na-date measured (month)

hc19

na-date measured (year)

hc27

na-sex

hc30

na-month of birth of child

hc31

na-year of birth of child

hc32

na-date of birth (cmc)

hc33

na-completeness of information

hc51

na-line no. of parent/caretaker

hc52

na-read consent statement

hc53

na-hemoglobin level (g/dl - 1 decimal)

hc55

na-result of measuring (hemoglobin)

hc56

na-hemoglobin level adjusted by altitude (g/dl - 1 decimal)

hc57

na-anemia level

hc58

na-agrees to referral

hc60

na-mother's line number from woman's questionnaire

hc61

na-mother's highest educational level

hc62

na-mother's highest year of education

hc63

na-preceding birth interval

hc64

na-birth order number

hc68

na-highest educational level (cs for prelimnary & final repo

hc70

na-ht/a standard deviations (according to who)

hc71

na-wt/a standard deviations (according to who)

hc72

na-wt/ht standard deviations (according to who)

hc73

na-bmi standard deviations (according to who)

hmlidx

net number

hml3

net observed -na

hml4

months ago obtained mosquit-na

hml5

was net treated with insecticide when bought

hml6

mosquito net treated with i-na

hml7

brand of bednet -na

hml8

bednet treated since receiving

hml9

months ago re-treated

hml10

itn net

hml11

number of persons who slept under this net

hml21

did someone sleep under bednet last night

hmla

person who slept in mosquito net

hmlb

person who slept in mosquito net

hmlc

person who slept in mosquito net

hmld

person who slept in mosquit-na

hmle

person who slept in mosquit-na

hmhidx

index to household schedule

hml12

type of bednet(s) person slept under last night

hml13

1st net number for net person slept under last night

hml14

2nd net number for net pers-na

hml15

3rd net number for net pers-na

hml16

corrected age from individu-na

hml17

flag for ah07 age from indi-na

hml18

pregnancy from individual q-na

hml19

person sleep under an ever treated bednet

sh20

line number

sh22

gender

sh23m

month of birth

sh23y

year of birth

sh24

live birth

sh25

currently alive

sh26

line number in household roster

sh23c

cmc birth

sh23f

cmc flag

hhid

case identification

hvidx

line number

hv000

country code and phase

hv001

cluster number

hv002

household number

hv003

respondent's line number

hv004

ultimate area unit

hv005

sample weight

hv006

month of interview

hv007

year of interview

hv008

date of interview (cmc)

hv009

number of household members

hv010

number of eligible women in hh

hv011

number of eligible men in hh

hv012

number of de jure members

hv013

number of de facto members

hv014

number of children 5 and under

hv015

result of household interview

hv016

day of interview

hv017

number of visits

hv018

interviewer identification

hv019

keyer identification

hv020

ever-married sample

hv021

primary sampling unit

hv022

sample stratum number

hv023

sample domain

hv024

region

hv025

type of place of residence

hv026

place of residence

hv027

selection for male/husb. int.

hv028

sample weight for male subsamp

hv030

field supervisor

hv031

field editor

hv032

office editor

hv033

ultimate area selection pr-na

hv035

number of eligible children for height & weight

hv040

cluster altitude in meters -na

hv041

total adults measured

hv042

selection for hemoglobin

hv043

selection for women's statu-na

hv044

selection for domestic viol-na

hv201

source of drinking water

hv202

source of non-drinking wate-na

hv204

time to get to water source

hv205

type of toilet facility

hv206

has electricity

hv207

has radio

hv208

has television

hv209

has refrigerator

hv210

has bicycle

hv211

has motorcycle/scooter

hv212

has car/truck

hv213

main floor material

hv214

main wall material

hv215

main roof material

hv216

rooms used for sleeping

hv217

relationship structure

hv218

line number of head of househ.

hv219

sex of head of household

hv220

age of head of household

hv221

has telephone

hv225

share toilet with other households

hv226

type of cooking fuel

hv227

have bednet for sleeping

hv228

children under 5 slept under bednet last night

hv234

test salt for iodine -na

hv235

location of source for wate-na

hv236

person fetching water -na

hv237

anything done to water to make safe to drink

hv237a

water usually treated by: boil

hv237b

water usually treated by: add bleach/chlorine

hv237c

water usually treated by: strain through a cloth

hv237d

water usually treated by: use water filter

hv237e

water usually treated by: solar disinfection

hv237f

water usually treated by: let it stand and settle

hv237g

water usually treated by: c-na

hv237h

water usually treated by: c-na

hv237i

water usually treated by: c-na

hv237j

water usually treated by: c-na

hv237k

water usually treated by: c-na

hv237x

water usually treated by: other

hv237z

water usually treated by: don't know

hv238

number of households sharin-na

hv239

food cooked on stove or ope-na

hv240

household has a chimney, ho-na

hv241

food cooked in the house / -na

hv242

household has separate room-na

hv243a

has a mobile telephone -na

hv243b

has a watch

hv243c

has an animal-drawn cart

hv243d

has a boat with a motor

hv244

own land usable for agriculture

hv245

hectares for agricultural l-na

hv246

livestock, herds or farm animals

hv246a

cattle own -na

hv246b

cows, bulls own

hv246c

horses, donkeys, mules own

hv246d

goats own

hv246e

sheep own

hv246f

chickens own

hv246g

buffalo own

hv246h

camels own

hv246i

cs own -na

hv246j

cs own -na

hv246k

cs own -na

hv247

owns a bank account -na

hv248

number of sick people 18-59-na

hv249

member of the hh died last -na

hv250

number of members who died -na

hv251

number of orphans and vulne-na

hv270

wealth index

hv271

wealth index factor score (5 decimals)

hml1

number of mosquito nets

hml1a

number of mosquito nets with specific information

hml2

number of children under bednet previous night

shdist

district

shtype

type of questionnaire

shv005

hh weight for all households

shv218

line number of head of household (all households).

shv219

sex of head of household (all households)

shv220

age of head of household (all households)

shbirths

total births in household since january 2003

shdeaths

total deaths in household since january 2003

shcdeath

total deaths < 5/sbs from q38

shfdeath

total female deaths age 12-49 from q39

sh107f

room cooler, air conditioner owns

sh107g

washing machine owns

sh107h

water pump owns

sh107i

bed owns

sh107j

chairs owns

sh107k

almirah/cabinet owns

sh107l

clock owns

sh107m

sofa owns

sh107n

sewing machine owns

sh107o

camera owns

sh107p

personal computer owns

sh113

status of house

sh127a

hh uses coil to avoid mosquitos

sh127b

hh uses mats to avoid mosquitos

sh127c

hh uses spray to avoid mosquitos

sh127d

hh uses electric spray repellant to avoid mosquitos

sh127e

hh uses insect repellant to avoid mosquitos

sh127f

hh uses smoke to avoid mosquitos

sh127g

hh uses fan to avoid mosquitos

sh127h

hh uses covered windows/screens to avoid mosquitos

sh127x

hh uses other to avoid mosquitos

sh127y

hh uses nothing to avoid mosquitos

sh128

currently have any medicines for treating malaria in the hou

hv101

relationship to head

hv102

usual resident

hv103

slept last night

hv104

sex of household member

hv105

age of household members

hv106

highest educational level

hv107

highest year of education

hv108

education in single years

hv109

educational attainment

hv110

member still in school -na

hv111

mother alive

hv112

mother's line number -na

hv113

father alive

hv114

father's line number -na

hv115

current marital status

hv116

currently, formerly, never m.

hv117

eligibility for female interview

hv118

eligibility for male interview

hv120

children eligibility for height/weight and hemoglobin

hv121

member attended school during current school year

hv122

educational level during current school year

hv123

grade of education during current school year

hv124

education in single years - current school year

hv125

member attended school during previous school year

hv126

educational level during pr-na

hv127

grade of education during p-na

hv128

education in single years --na

hv129

school attendance status

hv130

member has been very sick f-na

hv131

mother has been very sick f-na

hv132

father has been very sick f-na

hv133

mother/father dead or been -na

hv134

both parents alive -na

hv135

has brothers/sisters under -na

hv136

brothers/sisters under 18 t-na

hv137

member has a blanket -na

hv138

member has a pair of shoes -na

hv139

member has 2+ sets of cloth-na

hv140

member has a birth certificate

idxh4

index to household schedule

sh03

relationship to head

sh11

education levels

sh17

registered with nadra

ha0

na-index to household schedule

ha1

na-women's age in years

ha2

na-respondent's weight (kilos-1d)

ha3

na-respondent's height (cms-1d)

ha4

na-ht/a percentile (resp.)

ha5

na-ht/a standard deviations (resp

ha6

na-ht/a percent ref. median (resp

ha11

na-wt/ht std deviations(resp) dhs

ha12

na-wt/ht percent ref. median (dhs

ha12a

na-wt/ht percent ref. median (fog

ha12b

na-wt/ht percent ref. median (who

ha13

na-result of measurement of resp.

ha32

na-date of birth (cmc)

ha33

na-completeness of information

ha35

na-smoking

ha40

na-body mass index for respondent

ha41

na-rohrer's index for respondent

ha50

na-under age 18

ha51

na-line no. of parent/caretaker

ha52

na-read consent statement

ha53

na-hemoglobin level (g/dl - 1 decimal)

ha54

na-currently pregnant

ha55

na-result of measuring (hemoglobin)

ha56

na-hemoglobin level adjusted by altitude (g/dl - 1 decimal)

ha57

na-anemia level

ha58

na-agrees to referral

ha60

na-marital status

ha61

na-read consent statement (hiv)

ha62

na-blood sample id number

ha63

na-result of measuring (hiv)

ha64

na-consent for additional tests

ha65

na-result of woman individual interview

ha66

na-woman's highest educational level

ha67

na-woman's highest year of education

ha68

na-highest educational level (cs for prelimnary & final repo

ha69

na-hiv weight (6 decimals)

hc0

na-index to household schedule

hc1

na-age in months

hc2

na-weight in kilograms (1 dec.)

hc3

na-height in centimeters (1 dec.)

hc4

na-ht/a percentile

hc5

na-ht/a standard deviations

hc6

na-ht/a percent of ref. median

hc7

na-wt/a percentile

hc8

na-wt/a standard deviations

hc9

na-wt/a percent of ref. median

hc10

na-wt/ht percentile

hc11

na-wt/ht standard deviations

hc12

na-wt/ht percent of ref. median

hc13

na-reason not measured

hc15

na-height: lying or standing

hc16

na-day of birth of child

hc17

na-date measured (day)

hc18

na-date measured (month)

hc19

na-date measured (year)

hc27

na-sex

hc30

na-month of birth of child

hc31

na-year of birth of child

hc32

na-date of birth (cmc)

hc33

na-completeness of information

hc51

na-line no. of parent/caretaker

hc52

na-read consent statement

hc53

na-hemoglobin level (g/dl - 1 decimal)

hc55

na-result of measuring (hemoglobin)

hc56

na-hemoglobin level adjusted by altitude (g/dl - 1 decimal)

hc57

na-anemia level

hc58

na-agrees to referral

hc60

na-mother's line number from woman's questionnaire

hc61

na-mother's highest educational level

hc62

na-mother's highest year of education

hc63

na-preceding birth interval

hc64

na-birth order number

hc68

na-highest educational level (cs for prelimnary & final repo

hc70

na-ht/a standard deviations (according to who)

hc71

na-wt/a standard deviations (according to who)

hc72

na-wt/ht standard deviations (according to who)

hc73

na-bmi standard deviations (according to who)

hmlidx

net number

hml3

net observed -na

hml4

months ago obtained mosquit-na

hml5

was net treated with insecticide when bought

hml6

mosquito net treated with i-na

hml7

brand of bednet -na

hml8

bednet treated since receiving

hml9

months ago re-treated

hml10

itn net

hml11

number of persons who slept under this net

hml21

did someone sleep under bednet last night

hmla

person who slept in mosquito net

hmlb

person who slept in mosquito net

hmlc

person who slept in mosquito net

hmld

person who slept in mosquit-na

hmle

person who slept in mosquit-na

hmhidx

index to household schedule

hml12

type of bednet(s) person slept under last night

hml13

1st net number for net person slept under last night

hml14

2nd net number for net pers-na

hml15

3rd net number for net pers-na

hml16

corrected age from individu-na

hml17

flag for ah07 age from indi-na

hml18

pregnancy from individual q-na

hml19

person sleep under an ever treated bednet

sh20

line number

sh22

gender

sh23m

month of birth

sh23y

year of birth

sh24

live birth

sh25

currently alive

sh26

line number in household roster

sh23c

cmc birth

sh23f

cmc flag

References

National Institute of Population Studies (NIPS), Pakistan (http://nips.org.pk/Home.htm).

Examples

data("PKPR53")
PKPR53

#> # A tibble: 727,493 x 302
#>    hhid  hvidx hv000 hv001 hv002 hv003 hv004 hv005 hv006 hv007 hv008 hv009 hv010
#>    <chr> <int> <chr> <int> <int> <int> <int> <int> <int> <int> <int> <int> <int>
#>  1 "   …     1 PK5       1     1     2     1     0     2  2007  1286     7     0
#>  2 "   …     2 PK5       1     1     2     1     0     2  2007  1286     7     0
#>  3 "   …     3 PK5       1     1     2     1     0     2  2007  1286     7     0
#>  4 "   …     4 PK5       1     1     2     1     0     2  2007  1286     7     0
#>  5 "   …     5 PK5       1     1     2     1     0     2  2007  1286     7     0
#>  6 "   …     6 PK5       1     1     2     1     0     2  2007  1286     7     0
#>  7 "   …     7 PK5       1     1     2     1     0     2  2007  1286     7     0
#>  8 "   …     1 PK5       1     2     2     1     0     2  2007  1286     7     0
#>  9 "   …     2 PK5       1     2     2     1     0     2  2007  1286     7     0
#> 10 "   …     3 PK5       1     2     2     1     0     2  2007  1286     7     0
#> # … with 727,483 more rows, and 289 more variables: hv011 <int>, hv012 <int>,
#> #   hv013 <int>, hv014 <int>, hv015 <fct>, hv016 <int>, hv017 <int>,
#> #   hv018 <int>, hv019 <int>, hv020 <fct>, hv021 <int>, hv022 <int>,
#> #   hv023 <fct>, hv024 <fct>, hv025 <fct>, hv026 <fct>, hv027 <fct>,
#> #   hv028 <int>, hv030 <int>, hv031 <int>, hv032 <int>, hv033 <int>,
#> #   hv035 <int>, hv040 <int>, hv041 <int>, hv042 <fct>, hv043 <fct>,
#> #   hv044 <fct>, hv201 <int>, hv202 <fct>, hv204 <int>, hv205 <int>,
#> #   hv206 <int>, hv207 <int>, hv208 <int>, hv209 <int>, hv210 <int>,
#> #   hv211 <int>, hv212 <int>, hv213 <int>, hv214 <int>, hv215 <int>,
#> #   hv216 <int>, hv217 <fct>, hv218 <int>, hv219 <fct>, hv220 <int>,
#> #   hv221 <int>, hv225 <int>, hv226 <int>, hv227 <int>, hv228 <fct>,
#> #   hv234 <fct>, hv235 <fct>, hv236 <fct>, hv237 <int>, hv237a <int>,
#> #   hv237b <int>, hv237c <int>, hv237d <int>, hv237e <int>, hv237f <int>,
#> #   hv237g <fct>, hv237h <fct>, hv237i <fct>, hv237j <fct>, hv237k <fct>,
#> #   hv237x <int>, hv237z <int>, hv238 <fct>, hv239 <fct>, hv240 <fct>,
#> #   hv241 <fct>, hv242 <fct>, hv243a <fct>, hv243b <int>, hv243c <int>,
#> #   hv243d <int>, hv244 <int>, hv245 <fct>, hv246 <int>, hv246a <fct>,
#> #   hv246b <int>, hv246c <int>, hv246d <int>, hv246e <int>, hv246f <int>,
#> #   hv246g <int>, hv246h <int>, hv246i <fct>, hv246j <fct>, hv246k <fct>,
#> #   hv247 <fct>, hv248 <int>, hv249 <fct>, hv250 <fct>, hv251 <int>,
#> #   hv270 <fct>, hv271 <int>, hml1 <int>, …