Pakistan Demographic and Health Survey 2012-13 Household Member Recode Data

PKPR61 data of Household Member Recode from Pakistan Demographic and Health Survey 2012-13.

hhid

Case Identification

hvidx

Line number

hv000

Country code and phase

hv001

Cluster number

hv002

Household number

hv003

Respondent's line number (answering Household questionnaire)

hv004

Ultimate area unit

hv005

Household sample weight (6 decimals)

hv006

Month of interview

hv007

Year of interview

hv008

Date of interview (CMC)

hv009

Number of household members

hv010

Number of eligible women in household

hv011

Number of eligible men in household

hv012

Number of de jure members

hv013

Number of de facto members

hv014

Number of children 5 and under (de jure)

hv015

Result of household interview

hv016

Day of interview

hv017

Number of visits

hv018

Interviewer identification

hv019

Keyer identification

hv020

Ever-married sample

hv021

Primary sampling unit

hv022

Sample strata for sampling errors

hv023

Stratification used in sample design

hv024

Region

hv025

Type of place of residence

hv026

Place of residence

hv027

Household selected for male interview

hv028

Household weight for male subsample (6 decimals)

hv030

Field supervisor

hv031

Field editor

hv032

NA - Office editor

hv035

Number of eligible children for height and weight

hv040

NA - Cluster altitude in meters

hv041

Total adults measured

hv042

NA - Household selected for hemoglobin

hv044

Household selected for Domestic Violence module

hv201

Source of drinking water

hv202

NA - Source of non-drinking water

hv204

Time to get to water source (minutes)

hv205

Type of toilet facility

hv206

Has electricity

hv207

Has radio

hv208

Has television

hv209

Has refrigerator

hv210

Has bicycle

hv211

Has motorcycle/scooter

hv212

Has car/truck

hv213

Main floor material

hv214

Main wall material

hv215

Main roof material

hv216

Number of rooms used for sleeping

hv217

Relationship structure

hv218

Line number of head of household

hv219

Sex of head of household

hv220

Age of head of household

hv221

Has telephone (land-line)

hv225

Share toilet with other households

hv226

Type of cooking fuel

hv227

Has mosquito bed net for sleeping

hv228

NA - Children under 5 slept under mosquito bed net last night

hv230a

Place where household members wash their hands

hv230b

Presence of water at hand washing place

hv232

Items present: Soap or detergent

hv232b

Items present: Ash, mud, sand

hv232c

NA - Items present: CS

hv232d

NA - Items present: CS

hv232e

NA - Items present: CS

hv232y

Items present: None

hv234

NA - Result of salt test for iodine (PPM)

hv234a

NA - Result of salt test for iodine

hv235

Location of source for water

hv236

NA - Person fetching water

hv237

Anything done to water to make safe to drink

hv237a

Water usually treated by: boil

hv237b

Water usually treated by: add bleach/chlorine

hv237c

Water usually treated by: strain through a cloth

hv237d

Water usually treated by: use water filter

hv237e

Water usually treated by: solar disinfection

hv237f

Water usually treated by: let it stand and settle

hv237g

NA - Water usually treated by: CS

hv237h

NA - Water usually treated by: CS

hv237i

NA - Water usually treated by: CS

hv237j

NA - Water usually treated by: CS

hv237k

NA - Water usually treated by: CS

hv237x

Water usually treated by: other

hv237z

Water usually treated by: don't know

hv238

Number of households sharing toilet

hv239

NA - Food cooked on stove or open fire

hv240

NA - Household has a chimney, hood or neither

hv241

Food cooked in the house/ separate building/ outdoors

hv242

Household has separate room used as kitchen

hv243a

Has mobile telephone

hv243b

Has watch

hv243c

Has animal-drawn cart

hv243d

Has boat with a motor

hv244

Owns land usable for agriculture

hv245

Hectares of agricultural land (1 decimal)

hv246

Owns livestock, herds or farm animals

hv246a

NA - Owns cattle

hv246b

Owns cows/ bulls

hv246c

Owns horses/ donkeys/ mules

hv246d

Owns goats

hv246e

Owns sheep

hv246f

Owns chickens

hv246g

Owns buffalos

hv246h

Owns camels

hv246i

NA - Owns CS

hv246j

NA - Owns CS

hv246k

NA - Owns CS

hv247

Has bank account

hv252

Frequency household members smoke inside the house

hv253

Has dwelling been sprayed against mosquitoes in last 12 months

hv253a

Dwelling sprayed by: government worker/program

hv253b

Dwelling sprayed by: private company

hv253c

Dwelling sprayed by: NGO

hv253d

NA - Dwelling sprayed by: CS

hv253e

NA - Dwelling sprayed by: CS

hv253f

NA - Dwelling sprayed by: CS

hv253g

NA - Dwelling sprayed by: CS

hv253h

NA - Dwelling sprayed by: CS

hv253x

Dwelling sprayed by: other

hv253z

Dwelling sprayed by: don't know

hv270

Wealth index

hv271

Wealth index factor score (5 decimals)

hml1

Number of mosquito bed nets

hml1a

NA - Number of mosquito bed nets with specific information

hml2

NA - Number of children under mosquito bed net previous night

shdist

District

sh26

Any members of HH who lived here in the past 10 years but who have since moved

sh33a

Total number of migrants

sh110f

Has an almirah/cabinet

sh110g

Has a chair

sh110h

Has a room cooler

sh110i

Has air conditioner

sh110j

Has a washing machine

sh110k

Has a water pump

sh110l

Has a bed

sh110m

Has a clock

sh110n

Has a sofa

sh110o

Has a camera

sh110p

Has a sewing machine

sh110q

Has a computer

sh110r

Has internet connection

sh118g

Has a Tractor

sh118i

Has a boat without a motor

sh120u

Area for agricultural land: Unit

sh120n

Area for agricultural land: Number ( 1 decimal implicit)

sh128

Number of mosquito nets are insecticide treated

sh129

Usually soak the mosquito nets in a liquid

sh130

Number of months mosquito net are soaked or dipped

sh131

Use of insecticide treated nets reduce the incidence of malaria

sh132

Household has other ways to avoid mosquitoes

sh133a

Ways to avoid mosquitoes: coil

sh133b

Ways to avoid mosquitoes: mats

sh133c

Ways to avoid mosquitoes: spray

sh133d

Ways to avoid mosquitoes: electric spray repellent

sh133e

Ways to avoid mosquitoes: insect repellent

sh133f

Ways to avoid mosquitoes: infrared electric device

sh133g

Ways to avoid mosquitoes: smoke

sh133h

Ways to avoid mosquitoes: membrane

sh133x

Ways to avoid mosquitoes: other

hv101

Relationship to head

hv102

Usual resident

hv103

Slept last night

hv104

Sex of household member

hv105

Age of household members

hv106

Highest educational level attained

hv107

Highest year of education completed

hv108

Education completed in single years

hv109

Educational attainment

hv110

NA - Member still in school

hv111

Mother alive

hv112

Mother's line number

hv113

Father alive

hv114

Father's line number

hv115

Current marital status

hv116

Currently, formerly, never married

hv117

Eligibility for female interview

hv118

Eligibility for male interview

hv120

Children eligibility for height/weight and hemoglobin

hv121

Member attended school during current school year

hv122

Educational level during current school year

hv123

Grade of education during current school year

hv124

Education in single years - current school year

hv125

NA - Member attended school during previous school year

hv126

NA - Educational level during previous school year

hv127

NA - Grade of education during previous school year

hv128

NA - Education in single years - previous school year

hv129

NA - School attendance status

hv140

Member has a birth certificate

idxh4

Index to Household Schedule

sheduc

Person's educational level for report

sh13

Highest grade completed

sh15

Grade attending this year

sh16

Reasons for not attending school

sh21

Registered with NADRA

sh22

Born in this village/city

sh23u

Place moved from to this village/city: Unit

sh23n

Place moved from to this village/city: Code

sh24

Year when first move to this village/city

sh25

Primary reason of migration to this village/city

ha0

Index to Household Schedule

ha1

Woman's age in years

ha2

Woman's weight in kilograms (1 decimal)

ha3

Woman's height in centimeters (1 decimal)

ha4

Height/Age percentile

ha5

Height/Age standard deviation

ha6

Height/Age percent ref. median

ha11

Weight/Height standard deviation (DHS)

ha12

Weight/Height percent ref. median (DHS)

ha12a

Weight/Height percent ref. median (Fog)

ha12b

Weight/Height percent ref. median (WHO)

ha13

Result of measurement - height/weight

ha32

Date of birth (CMC)

ha33

Completeness of HA32 information

ha35

Smoking (cigarettes in last 24 hours)

ha40

Body Mass Index

ha41

Rohrer's index

ha50

NA - Under age 18

ha51

NA - Line number of parent/caretaker

ha52

NA - Read consent statement - hemoglobin

ha53

NA - Hemoglobin level (g/dl - 1 decimal)

ha54

Currently pregnant

ha55

NA - Result of measurement - hemoglobin

ha56

NA - Hemoglobin level adjusted for altitude and smoking (g/dl - 1 decimal)

ha57

NA - Anemia level

ha58

NA - Agrees to referral - anemia

ha60

NA - Marital status

ha61

NA - Read consent statement - HIV

ha62

Bar code for HIV blood sample

ha63

NA - Result of measurement - HIV

ha64

NA - Consent for additional tests

ha65

Result of woman's individual interview

ha66

Woman's highest educational level

ha67

Woman's highest year of education

ha68

Woman's highest educational level (for preliminary and final report)

ha69

NA - HIV weight (6 decimals)

ha70

NA - Interviewer that took blood for HIV testing

hc0

Index to Household Schedule

hc1

Child's age in months

hc2

Child's weight in kilograms (1 decimal)

hc3

Child's height in centimeters (1 decimal)

hc4

Height/Age percentile

hc5

Height/Age standard deviation

hc6

Height/Age percent of ref. median

hc7

Weight/Age percentile

hc8

Weight/Age standard deviation

hc9

Weight/Age percent of ref. median

hc10

Weight/Height percentile

hc11

Weight/Height standard deviation

hc12

Weight/Height percent of ref. median

hc13

Result of measurement - height/weight

hc15

Height: lying or standing

hc16

Day of birth

hc17

Date measured (day)

hc18

Date measured (month)

hc19

Date measured (year)

hc27

Sex

hc30

Month of birth

hc31

Year of birth

hc32

Date of birth (CMC)

hc33

Completeness of HC32 information

hc51

NA - Line number of parent/caretaker

hc52

NA - Read consent statement - hemoglobin

hc53

NA - Hemoglobin level (g/dl - 1 decimal)

hc55

NA - Result of measurement - hemoglobin

hc56

NA - Hemoglobin level adjusted for altitude (g/dl - 1 decimal)

hc57

NA - Anemia level

hc58

NA - Agrees to referral - anemia

hc60

Mother's line number (woman's questionnaire)

hc61

Mother's highest educational level

hc62

Mother's highest year of education

hc63

Preceding birth interval (months)

hc64

Birth order number

hc68

Mother's highest educational level (for preliminary and final report)

hc70

Height/Age standard deviation (new WHO)

hc71

Weight/Age standard deviation (new WHO)

hc72

Weight/Height standard deviation (new WHO)

hc73

BMI standard deviation (new WHO)

References

National Institute of Population Studies (NIPS), Pakistan (http://nips.org.pk/Home.htm).

Examples

data("PKPR61")
PKPR61

#> # A tibble: 94,169 x 282
#>    hhid  hvidx hv000 hv001 hv002 hv003 hv004  hv005 hv006 hv007 hv008 hv009
#>    <chr> <int> <chr> <int> <int> <int> <int>  <int> <int> <int> <int> <int>
#>  1 1 1       1 PK6       1     1     1     1 2.26e6    11  2012  1355     3
#>  2 1 1       2 PK6       1     1     1     1 2.26e6    11  2012  1355     3
#>  3 1 1       3 PK6       1     1     1     1 2.26e6    11  2012  1355     3
#>  4 1 2       1 PK6       1     2     2     1 2.26e6    11  2012  1355    11
#>  5 1 2       2 PK6       1     2     2     1 2.26e6    11  2012  1355    11
#>  6 1 2       3 PK6       1     2     2     1 2.26e6    11  2012  1355    11
#>  7 1 2       4 PK6       1     2     2     1 2.26e6    11  2012  1355    11
#>  8 1 2       5 PK6       1     2     2     1 2.26e6    11  2012  1355    11
#>  9 1 2       6 PK6       1     2     2     1 2.26e6    11  2012  1355    11
#> 10 1 2       7 PK6       1     2     2     1 2.26e6    11  2012  1355    11
#> # … with 94,159 more rows, and 270 more variables: hv010 <int>, hv011 <int>,
#> #   hv012 <int>, hv013 <int>, hv014 <int>, hv015 <fct>, hv016 <int>,
#> #   hv017 <int>, hv018 <int>, hv019 <int>, hv020 <fct>, hv021 <int>,
#> #   hv022 <int>, hv023 <fct>, hv024 <fct>, hv025 <fct>, hv026 <fct>,
#> #   hv027 <fct>, hv028 <int>, hv030 <int>, hv031 <int>, hv032 <int>,
#> #   hv035 <int>, hv040 <int>, hv041 <int>, hv042 <fct>, hv044 <fct>,
#> #   hv201 <int>, hv202 <fct>, hv204 <int>, hv205 <int>, hv206 <int>,
#> #   hv207 <int>, hv208 <int>, hv209 <int>, hv210 <int>, hv211 <int>,
#> #   hv212 <int>, hv213 <int>, hv214 <int>, hv215 <int>, hv216 <int>,
#> #   hv217 <fct>, hv218 <int>, hv219 <fct>, hv220 <int>, hv221 <int>,
#> #   hv225 <int>, hv226 <int>, hv227 <int>, hv228 <fct>, hv230a <int>,
#> #   hv230b <int>, hv232 <fct>, hv232b <fct>, hv232c <fct>, hv232d <fct>,
#> #   hv232e <fct>, hv232y <fct>, hv234 <fct>, hv234a <fct>, hv235 <int>,
#> #   hv236 <fct>, hv237 <int>, hv237a <int>, hv237b <int>, hv237c <int>,
#> #   hv237d <int>, hv237e <int>, hv237f <int>, hv237g <fct>, hv237h <fct>,
#> #   hv237i <fct>, hv237j <fct>, hv237k <fct>, hv237x <int>, hv237z <int>,
#> #   hv238 <int>, hv239 <fct>, hv240 <fct>, hv241 <int>, hv242 <int>,
#> #   hv243a <int>, hv243b <int>, hv243c <int>, hv243d <int>, hv244 <int>,
#> #   hv245 <int>, hv246 <int>, hv246a <fct>, hv246b <int>, hv246c <int>,
#> #   hv246d <int>, hv246e <int>, hv246f <int>, hv246g <int>, hv246h <int>,
#> #   hv246i <fct>, hv246j <fct>, hv246k <fct>, …