Pakistan Demographic and Health Survey 2017-18 Household Member Recode Data

PKPR71 data of Household Member Recode from Pakistan Demographic and Health Survey 2017-18.

HHID

Case Identification

HVIDX

Line number

HV000

Country code and phase

HV001

Cluster number

HV002

Household number

HV003

Respondent's line number (answering Household questionnaire)

HV004

Ultimate area unit

HV005

Household sample weight (6 decimals)

HV006

Month of interview

HV007

Year of interview

HV008

Date of interview (CMC)

HV008A

Date of interview Century Day Code (CDC)

HV009

Number of household members

HV010

Number of eligible women in household

HV011

Number of eligible men in household

HV012

Number of de jure members

HV013

Number of de facto members

HV014

Number of children 5 and under (de jure)

HV015

Result of household interview

HV016

Day of interview

HV017

Number of visits

HV018

Interviewer identification

HV019

Keyer identification

HV020

Ever-married sample

HV021

Primary sampling unit

HV022

Sample strata for sampling errors

HV023

Stratification used in sample design

HV024

Region

HV025

Type of place of residence

HV026

NA - Place of residence

HV027

Household selected for male interview

HV028

Household weight for male subsample (6 decimals)

HV030

Field supervisor

HV031

Field editor

HV032

NA - Office editor

HV035

Number of eligible children for height and weight

HV040

NA - Cluster altitude in meters

HV041

Total adults measured

HV042

Household selected for hemoglobin

HV044

Household selected for Domestic Violence module

HV045A

Language of questionnaire

HV045B

Language of interview

HV045C

Native language of respondent

HV046

Translator used

HV801

Time interview started (hhmm - 24 hour clock)

HV802

Time interview ended (hhmm - 24 hour clock)

HV803

Length of interview in minutes

HV804

Number of visits for biomarker

HV807D

Day of biomarker visit

HV807M

Month of biomarker visit

HV807Y

Year of biomarker visit

HV807C

Date of biomarker (CMC)

HV807A

Date of biomarker Century Day Code (CDC)

HV201

Source of drinking water

HV202

Source of non-drinking water

HV201A

Water not available for at least a day last two weeks

HV204

Time to get to water source (minutes)

HV205

Type of toilet facility

HV206

Has electricity

HV207

Has radio

HV208

Has television

HV209

Has refrigerator

HV210

Has bicycle

HV211

Has motorcycle/scooter

HV212

Has car/truck

HV213

Main floor material

HV214

Main wall material

HV215

Main roof material

HV216

Number of rooms used for sleeping

HV217

Relationship structure

HV218

Line number of head of household

HV219

Sex of head of household

HV220

Age of head of household

HV221

Has telephone (land-line)

HV225

Share toilet with other households

HV226

Type of cooking fuel

HV227

Has mosquito bed net for sleeping

HV228

Children under 5 slept under mosquito bed net last night

HV230A

Place where household members wash their hands

HV230B

Presence of water at hand washing place

HV232

Items present: Soap or detergent

HV232B

Items present: Ash, mud, sand

HV232C

NA - Items present: CS

HV232D

NA - Items present: CS

HV232E

NA - Items present: CS

HV232Y

Items present: None

HV234

NA - Result of salt test for iodine (PPM)

HV234A

NA - Result of salt test for iodine

HV235

Location of source for water

HV236

NA - Person fetching water

HV237

Anything done to water to make safe to drink

HV237A

Water usually treated by: boil

HV237B

Water usually treated by: add bleach/chlorine

HV237C

Water usually treated by: strain through a cloth

HV237D

Water usually treated by: use water filter

HV237E

Water usually treated by: solar disinfection

HV237F

Water usually treated by: let it stand and settle

HV237G

NA - Water usually treated by: CS

HV237H

NA - Water usually treated by: CS

HV237I

NA - Water usually treated by: CS

HV237J

NA - Water usually treated by: CS

HV237K

NA - Water usually treated by: CS

HV237X

Water usually treated by: other

HV237Z

Water usually treated by: don't know

HV238

Number of households sharing toilet

HV238A

Location of toilet facility

HV239

NA - Food cooked on stove or open fire

HV240

NA - Household has a chimney, hood or neither

HV241

Food cooked in the house/ separate building/ outdoors

HV242

Household has separate room used as kitchen

HV243A

Has mobile telephone

HV243B

Has watch

HV243C

Has animal-drawn cart

HV243D

Has boat with a motor

HV243E

Has a computer

HV244

Owns land usable for agriculture

HV245

Hectares of agricultural land (1 decimal)

HV246

Owns livestock, herds or farm animals

HV246A

Owns cattle

HV246B

Owns cows/ bulls

HV246C

Owns horses/ donkeys/ mules

HV246D

Owns goats

HV246E

Owns sheep

HV246F

Owns chickens/poultry

HV246G

Owns Camels

HV246H

NA - Owns CS

HV246I

NA - Owns CS

HV246J

NA - Owns CS

HV246K

NA - Owns CS

HV247

Has bank account

HV252

Frequency household members smoke inside the house

HV253

Has dwelling been sprayed against mosquitoes in last 12 months

HV253A

Dwelling sprayed by: government worker/program

HV253B

Dwelling sprayed by: private company

HV253C

Dwelling sprayed by: NGO

HV253D

NA - Dwelling sprayed by: CS

HV253E

NA - Dwelling sprayed by: CS

HV253F

NA - Dwelling sprayed by: CS

HV253G

NA - Dwelling sprayed by: CS

HV253H

NA - Dwelling sprayed by: CS

HV253X

Dwelling sprayed by: other

HV253Z

Dwelling sprayed by: don't know

HV270

Wealth index combined

HV271

Wealth index factor score combined (5 decimals)

HV270A

Wealth index for urban/rural

HV271A

Wealth index factor score for urban/rural (5 decimals)

HML1

Number of mosquito bed nets

HML1A

Number of mosquito bed nets with specific information

HML2

Number of children under mosquito bed net previous night

SHDIST

District

SH35

Any member of the household who lived here in past 10 years and moved away

SH42

Total number of migrants

SH120

Area for agricultural land (Acres and Kanal, 1 decimal)

SH121F

Almirah/Cabinet

SH121G

Chair

SH121H

Room cooler

SH121I

Air conditioner

SH121J

Washing machine

SH121K

Water pump

SH121L

Bed

SH121M

Clock

SH121N

Sofa

SH121O

Camera

SH121P

Sewing machine

SH121R

Internet connection

SH122G

Tractor

SH122I

Boat without motor

SH122J

Rickshaw/chingchi

SHV005

Household sample weight (AJK/GB, 6 decimals)

SHV028

Household weight for male subsample (AJK/GB, 6 decimals)

HV101

Relationship to head

HV102

Usual resident

HV103

Slept last night

HV104

Sex of household member

HV105

Age of household members

HV106

Highest educational level attained

HV107

Highest year of education completed

HV108

Education completed in single years

HV109

Educational attainment

HV110

NA - Member still in school

HV111

Mother alive

HV112

Mother's line number

HV113

Father alive

HV114

Father's line number

HV115

Current marital status

HV116

Currently, formerly, never married

HV117

Eligibility for female interview

HV118

Eligibility for male interview

HV120

Children eligibility for height/weight and hemoglobin

HV121

Member attended school during current school year

HV122

Educational level during current school year

HV123

Grade of education during current school year

HV124

Education in single years - current school year

HV125

NA - Member attended school during previous school year

HV126

NA - Educational level during previous school year

HV127

NA - Grade of education during previous school year

HV128

NA - Education in single years - previous school year

HV129

NA - School attendance status

HV140

Member has a birth certificate

IDXH4

Index to Household Schedule

SHEDUC

Person's educational level for report

SH17

Highest grade completed

SH19

Grade attending this year

SH19A

Reasons for not attending school

SH20

Birth certificate

SH20A

Have NIC card

SH21

Born in this village/city

SH21AU

Village/city was born: Unit

SH21AN

Village/city/country was born: Code

SH22U

Place from where move to this village/city/country: Unit

SH22N

Place from where move to this village/city/country: Code

SH23

Year when first move to this village/city

SH24

Primary reason to move to this village/city

HA0

Index to Household Schedule

HA1

Woman's age in years

HA2

Woman's weight in kilograms (1 decimal)

HA3

Woman's height in centimeters (1 decimal)

HA4

Height/Age percentile

HA5

Height/Age standard deviation

HA6

Height/Age percent ref. median

HA11

Weight/Height standard deviation (DHS)

HA12

Weight/Height percent ref. median (DHS)

HA12A

Weight/Height percent ref. median (Fog)

HA12B

Weight/Height percent ref. median (WHO)

HA13

Result of measurement - height/weight

HA21

Fieldworker measurer code

HA32

Date of birth (CMC)

HA33

Completeness of HA32 information

HA35

Smoking (cigarettes in last 24 hours)

HA40

Body Mass Index

HA41

Rohrer's index

HA50

NA - Under age 18

HA51

NA - Line number of parent/caretaker

HA52

NA - Read consent statement - hemoglobin

HA53

NA - Hemoglobin level (g/dl - 1 decimal)

HA54

Currently pregnant

HA55

NA - Result of measurement - hemoglobin

HA56

NA - Hemoglobin level adjusted for altitude and smoking (g/dl - 1 decimal)

HA57

NA - Anemia level

HA58

NA - Agrees to referral - anemia

HA60

NA - Marital status

HA61

NA - Read consent statement - HIV

HA62

Bar code for HIV blood sample

HA63

NA - Result of measurement - HIV

HA64

NA - Consent for additional tests

HA65

Result of woman's individual interview

HA66

Woman's highest educational level

HA67

Woman's highest year of education

HA68

Woman's highest educational level (for preliminary and final report)

HA69

NA - HIV weight (6 decimals)

HA70

NA - Interviewer that took blood for HIV testing

HC0

Index to Household Schedule

HC1

Child's age in months

HC1A

Child's age in days

HC2

Child's weight in kilograms (1 decimal)

HC3

Child's height in centimeters (1 decimal)

HC4

Height/Age percentile

HC5

Height/Age standard deviation

HC6

Height/Age percent of ref. median

HC7

Weight/Age percentile

HC8

Weight/Age standard deviation

HC9

Weight/Age percent of ref. median

HC10

Weight/Height percentile

HC11

Weight/Height standard deviation

HC12

Weight/Height percent of ref. median

HC13

Result of measurement - height/weight

HC15

Height: lying or standing

HC16

Day of birth

HC17

Date measured (day)

HC18

Date measured (month)

HC19

Date measured (year)

HC20

Century Day Code (CDC) of measurement

HC21

Fieldworker measurer code

HC27

Sex

HC30

Month of birth

HC31

Year of birth

HC32

Date of birth (CMC)

HC32A

Century Day Code (CDC) of birth

HC33

Completeness of HC32 information

HC51

NA - Line number of parent/caretaker

HC52

NA - Read consent statement - hemoglobin

HC53

NA - Hemoglobin level (g/dl - 1 decimal)

HC55

NA - Result of measurement - hemoglobin

HC56

NA - Hemoglobin level adjusted for altitude (g/dl - 1 decimal)

HC57

NA - Anemia level

HC58

NA - Agrees to referral - anemia

HC60

Mother's line number (woman's questionnaire)

HC61

Mother's highest educational level

HC62

Mother's highest year of education

HC63

Preceding birth interval (months)

HC64

Birth order number

HC68

Mother's highest educational level (for preliminary and final report)

HC70

Height/Age standard deviation (new WHO)

HC71

Weight/Age standard deviation (new WHO)

HC72

Weight/Height standard deviation (new WHO)

HC73

BMI standard deviation (new WHO)

HMLIDX

Mosquito Bed Net Designation Number

HML3

Net observed by interviewer

HML4

Months ago net obtained

HML5

NA - Net treated with insecticide when bought

HML6

NA - Net treatment status

HML7

Brand of net

HML8

Net treated since received

HML9

NA - Time since last re-treatment (months)

HML10

Insecticide-Treated Net (ITN)

HML11

Number of persons who slept under this net

HML21

Someone slept under this net last night

HMLA

Line number of person who slept in this net

HMLB

Line number of person who slept in this net

HMLC

Line number of person who slept in this net

HMLD

Line number of person who slept in this net

HMLE

NA - Line number of person who slept in this net

HML22

Obtained net from campaign, antenatal or immunization visit

HML23

Place where net was obtained

HMHIDX

Index to Household Schedule

HML12

Type of Mosquito Bed Net(s) person slept under last night

HML13

Net Designation Number (HMLIDX) for 1st net person slept under last night

HML14

Net Designation Number (HMLIDX) for 2nd net person slept under last night

HML15

NA - Net Designation Number (HMLIDX) for 3rd net person slept under last night

HML16

Corrected age from Individual file

HML16A

NA - Age in months (for children)

HML17

Flag for age from Individual file

HML18

Pregnancy status from Individual file

HML19

Person slept under an ever-treated net

HML20

Person slept under an LLIN net

HML30

NA - Line number of parent/caretaker (for malaria testing)

HML31

NA - Read consent statement for malaria

HML32

NA - Final result of malaria from blood smear test

HML32A

NA - Presence of species: falciparum (Pf)

HML32B

NA - Presence of species: malariae (Pm)

HML32C

NA - Presence of species: ovale (Po)

HML32D

NA - Presence of species: vivax (Pv)

HML32E

NA - Presence of species: CS

HML32F

NA - Presence of species: CS

HML32G

NA - Presence of species: CS

HML33

NA - Result of malaria measurement

HML34

Bar code for blood smear sample

HML35

NA - Result of malaria rapid test

HML36

NA - Fieldworker measurer code for malaria

IDXDIS

Index to household schedule

HDIS1

Wear glasses or contact lenses

HDIS2

Have difficulty seeing

HDIS3

Wear a hearing aid

HDIS4

Have difficulty hearing

HDIS5

Have difficulty communicating using usual language

HDIS6

Have difficulty remembering or concentrating

HDIS7

Have difficulty walking or climbing steps

HDIS8

Have difficulty washing all over or dressing

HDIS9

Highest degree of difficulty for any of the impairments

References

National Institute of Population Studies (NIPS), Pakistan (http://nips.org.pk/Home.htm).

Examples

load("PKPR71")

#> Warning: cannot open compressed file 'PKPR71', probable reason 'No such file or directory'

#> Error in readChar(con, 5L, useBytes = TRUE): cannot open the connection

PKPR71

#> # A tibble: 100,869 x 349
#>    HHID  HVIDX HV000 HV001 HV002 HV003 HV004  HV005 HV006 HV007 HV008 HV008A
#>    <fct> <int> <fct> <int> <int> <fct> <int>  <int> <int> <int> <int>  <int>
#>  1 "   …     1 PK7       1     1 2         1 849785     2  2018  1418  43135
#>  2 "   …     2 PK7       1     1 2         1 849785     2  2018  1418  43135
#>  3 "   …     3 PK7       1     1 2         1 849785     2  2018  1418  43135
#>  4 "   …     4 PK7       1     1 2         1 849785     2  2018  1418  43135
#>  5 "   …     5 PK7       1     1 2         1 849785     2  2018  1418  43135
#>  6 "   …     6 PK7       1     1 2         1 849785     2  2018  1418  43135
#>  7 "   …     7 PK7       1     1 2         1 849785     2  2018  1418  43135
#>  8 "   …     1 PK7       1     2 2         1 849785     2  2018  1418  43133
#>  9 "   …     2 PK7       1     2 2         1 849785     2  2018  1418  43133
#> 10 "   …     1 PK7       1     3 3         1 849785     2  2018  1418  43133
#> # … with 100,859 more rows, and 337 more variables: HV009 <int>, HV010 <int>,
#> #   HV011 <int>, HV012 <int>, HV013 <int>, HV014 <int>, HV015 <fct>,
#> #   HV016 <int>, HV017 <int>, HV018 <int>, HV019 <int>, HV020 <fct>,
#> #   HV021 <int>, HV022 <fct>, HV023 <fct>, HV024 <fct>, HV025 <fct>,
#> #   HV026 <fct>, HV027 <fct>, HV028 <int>, HV030 <int>, HV031 <int>,
#> #   HV032 <int>, HV035 <int>, HV040 <int>, HV041 <int>, HV042 <fct>,
#> #   HV044 <fct>, HV045A <fct>, HV045B <fct>, HV045C <fct>, HV046 <fct>,
#> #   HV801 <int>, HV802 <int>, HV803 <fct>, HV804 <int>, HV807D <int>,
#> #   HV807M <int>, HV807Y <int>, HV807C <int>, HV807A <int>, HV201 <fct>,
#> #   HV202 <fct>, HV201A <fct>, HV204 <fct>, HV205 <fct>, HV206 <fct>,
#> #   HV207 <fct>, HV208 <fct>, HV209 <fct>, HV210 <fct>, HV211 <fct>,
#> #   HV212 <fct>, HV213 <fct>, HV214 <fct>, HV215 <fct>, HV216 <int>,
#> #   HV217 <fct>, HV218 <int>, HV219 <fct>, HV220 <fct>, HV221 <fct>,
#> #   HV225 <fct>, HV226 <fct>, HV227 <fct>, HV228 <fct>, HV230A <fct>,
#> #   HV230B <fct>, HV232 <fct>, HV232B <fct>, HV232C <fct>, HV232D <fct>,
#> #   HV232E <fct>, HV232Y <fct>, HV234 <fct>, HV234A <fct>, HV235 <fct>,
#> #   HV236 <fct>, HV237 <fct>, HV237A <fct>, HV237B <fct>, HV237C <fct>,
#> #   HV237D <fct>, HV237E <fct>, HV237F <fct>, HV237G <fct>, HV237H <fct>,
#> #   HV237I <fct>, HV237J <fct>, HV237K <fct>, HV237X <fct>, HV237Z <fct>,
#> #   HV238 <fct>, HV238A <fct>, HV239 <fct>, HV240 <fct>, HV241 <fct>,
#> #   HV242 <fct>, HV243A <fct>, HV243B <fct>, …