Pakistan Demographic and Health Survey 2006-07 Vaccination of Children Data

PKVC51 data of Vaccination of Children from Pakistan Demographic and Health Survey 2006-07.

qhclust

cluster number

qhnumber

household number

qqtype

type of questionnaire (numeric)

qline

line number of woman/deceased woman/child/stillbirth

rec.type

qcluster

cluster number

qnumber

household number

qregion

region of residence

qdist

district

qlocal3

city \ town \ countryside - 3 values

qlocal4

city \ town \ countryside

typeq

type of questionnaire (alpha)

qintd

day of interview

qintm

month of interview

qinty

year of interview

qintnum

interviewer number

qresult

result of individual interview

qvisits

total number of visits

qqlang

language of questionaire

qilang

language of interview

qwlang

language woman speaks a home

qsuperv

supervisor number

qfedit

field editor number

qoedit

office editor number

qkeyer

data entry clerk

qintc

date of interview (cmc)

qweight

sample weight (6 decimals)

qtype

type of place of residence

qres

usual resident or visitor

c101.1

line number

c101.2

line number

c101.3

line number

c101.4

line number

c101.5

line number

c101.6

line number

c103.1

sex of respondent

c103.2

sex of respondent

c103.3

sex of respondent

c103.4

sex of respondent

c103.5

sex of respondent

c103.6

sex of respondent

c104.1

age of respondent

c104.2

age of respondent

c104.3

age of respondent

c104.4

age of respondent

c104.5

age of respondent

c104.6

age of respondent

c105.1

relationship with deceased child

c105.2

relationship with deceased child

c105.3

relationship with deceased child

c105.4

relationship with deceased child

c105.5

relationship with deceased child

c105.6

relationship with deceased child

c106.1

number of classes passed

c106.2

number of classes passed

c106.3

number of classes passed

c106.4

number of classes passed

c106.5

number of classes passed

c106.6

number of classes passed

c107.1

respondent present at delivery

c107.2

respondent present at delivery

c107.3

respondent present at delivery

c107.4

respondent present at delivery

c107.5

respondent present at delivery

c107.6

respondent present at delivery

c108.1

respondent present when ill

c108.2

respondent present when ill

c108.3

respondent present when ill

c108.4

respondent present when ill

c108.5

respondent present when ill

c108.6

respondent present when ill

c109.1

present when deceased died

c109.2

present when deceased died

c109.3

present when deceased died

c109.4

present when deceased died

c109.5

present when deceased died

c109.6

present when deceased died

c110.1

line number of main respondent

c110.2

line number of main respondent

c110.3

line number of main respondent

c110.4

line number of main respondent

c110.5

line number of main respondent

c110.6

line number of main respondent

c100

number of respondents for child

c111h

start time of interview (hours)

c111m

start time of interview (minutes)

fc112

q112 (filter): check 105 - clerk must enter

c113

mother of deceased child still alive

c114

line number of mother

c115

line number of father

c116m

date of birth of deceased's mother (month)

c116y

date of birth of deceased's mother (year)

c117m

date of marriage of deceased's mother (month)

c117y

date of marriage of deceased's mother (years)

c118

total number of pregnancy for mother

c119d

date of birth of deceased's child (day)

c119m

date of birth of deceased's child (month)

c119y

date of birth of decesed's child (year)

c120

child born alive or dead

c121

baby ever cry

c122

baby ever show movement

c123

baby ever breathe

c124d

date of death of deceased child (day)

c124m

date of death of deceased child (month)

c124y

date of death of deceased child (year)

c125u

age of deceased child at the time of death (unit)

c125n

age of deceased child at the time of death (period)

c126

sex of deceased child

kcod.1

cod

kcod.2

cod

kcod.3

cod

kcod.4

cod

kgrpdi.1

grpdig

kgrpdi.2

grpdig

kgrpdi.3

grpdig

kgrpdi.4

grpdig

kdoc.1

doctype

kdoc.2

doctype

kdoc.3

doctype

kdoc.4

doctype

c119c

cmc death

c119f

cmc flag

c124c

cmc death

c124f

cmc flag

coda

cod

codb

cod

codc

cod

grpdxa

grpdig

grpdxb

grpdig

grpdxc

grpdig

compdx

computer dx

commnt

comment

qualdxf

????? quality of decision?

dxfin

final dx

dthtype

type of death per docs

wkv

work variable (not appl)

c201n.01

label for c201

c201n.02

label for c201

c201n.03

label for c201

c201n.04

label for c201

c201n.05

label for c201

c201n.06

label for c201

c201n.07

label for c201

c201n.08

label for c201

c201n.09

label for c201

c201n.10

label for c201

c201n.11

label for c201

c201n.12

label for c201

c201n.13

label for c201

c201.01

complications during pregnancy

c201.02

complications during pregnancy

c201.03

complications during pregnancy

c201.04

complications during pregnancy

c201.05

complications during pregnancy

c201.06

complications during pregnancy

c201.07

complications during pregnancy

c201.08

complications during pregnancy

c201.09

complications during pregnancy

c201.10

complications during pregnancy

c201.11

complications during pregnancy

c201.12

complications during pregnancy

c201.13

complications during pregnancy

c202n.01

label for c202

c202n.02

label for c202

c202n.03

label for c202

c202n.04

label for c202

c202n.05

label for c202

c202n.06

label for c202

c202n.07

label for c202

c202n.08

label for c202

c202n.09

label for c202

c202n.10

label for c202

c202n.11

label for c202

c202n.12

label for c202

c202n.13

label for c202

c202n.14

label for c202

c202n.15

label for c202

c202n.16

label for c202

c202n.17

label for c202

c202n.18

label for c202

c202n.19

label for c202

c202.01

complications during labour and delivery

c202.02

complications during labour and delivery

c202.03

complications during labour and delivery

c202.04

complications during labour and delivery

c202.05

complications during labour and delivery

c202.06

complications during labour and delivery

c202.07

complications during labour and delivery

c202.08

complications during labour and delivery

c202.09

complications during labour and delivery

c202.10

complications during labour and delivery

c202.11

complications during labour and delivery

c202.12

complications during labour and delivery

c202.13

complications during labour and delivery

c202.14

complications during labour and delivery

c202.15

complications during labour and delivery

c202.16

complications during labour and delivery

c202.17

complications during labour and delivery

c202.18

complications during labour and delivery

c202.19

complications during labour and delivery

fc203

q203 (filter): check 126

c204n.01

label for c204

c204n.02

label for c204

c204n.03

label for c204

c204n.04

label for c204

c204n.05

label for c204

c204n.06

label for c204

c204n.07

label for c204

c204n.08

label for c204

c204n.09

label for c204

c204n.10

label for c204

c204n.11

label for c204

c204n.12

label for c204

c204n.13

label for c204

c204n.14

label for c204

c204n.15

label for c204

c204n.16

label for c204

c204n.17

label for c204

c204n.18

label for c204

c204n.19

label for c204

c204n.20

label for c204

c204n.21

label for c204

c204n.22

label for c204

c204n.23

label for c204

c204n.24

label for c204

c204n.25

label for c204

c204n.26

label for c204

c204n.27

label for c204

c204n.28

label for c204

c204n.29

label for c204

c204.01

early complications for the baby

c204.02

early complications for the baby

c204.03

early complications for the baby

c204.04

early complications for the baby

c204.05

early complications for the baby

c204.06

early complications for the baby

c204.07

early complications for the baby

c204.08

early complications for the baby

c204.09

early complications for the baby

c204.10

early complications for the baby

c204.11

early complications for the baby

c204.12

early complications for the baby

c204.13

early complications for the baby

c204.14

early complications for the baby

c204.15

early complications for the baby

c204.16

early complications for the baby

c204.17

early complications for the baby

c204.18

early complications for the baby

c204.19

early complications for the baby

c204.20

early complications for the baby

c204.21

early complications for the baby

c204.22

early complications for the baby

c204.23

early complications for the baby

c204.24

early complications for the baby

c204.25

early complications for the baby

c204.26

early complications for the baby

c204.27

early complications for the baby

c204.28

early complications for the baby

c204.29

early complications for the baby

c204.30

complication during illness - infection

c301a

illnesses during pregnancy

c301n.01

number of illness

c301n.02

number of illness

c301n.03

number of illness

c301n.04

number of illness

c301n.05

number of illness

c301n.06

number of illness

c301n.07

number of illness

c301n.08

number of illness

c301n.09

number of illness

c301n.10

number of illness

c301n.11

number of illness

c301n.12

number of illness

c301n.13

number of illness

c301b.01

duration of illness

c301b.02

duration of illness

c301b.03

duration of illness

c301b.04

duration of illness

c301b.05

duration of illness

c301b.06

duration of illness

c301b.07

duration of illness

c301b.08

duration of illness

c301b.09

duration of illness

c301b.10

duration of illness

c301b.11

duration of illness

c301b.12

duration of illness

c301b.13

duration of illness

c301c.01

stage of pregnancy

c301c.02

stage of pregnancy

c301c.03

stage of pregnancy

c301c.04

stage of pregnancy

c301c.05

stage of pregnancy

c301c.06

stage of pregnancy

c301c.07

stage of pregnancy

c301c.08

stage of pregnancy

c301c.09

stage of pregnancy

c301c.10

stage of pregnancy

c301c.11

stage of pregnancy

c301c.12

stage of pregnancy

c301c.13

stage of pregnancy

fc302

q302 (filter): check 301

c303

health care sought from health care provider/facility

c304

place from where health care sought

c305

vaginal problem before delivery for any previous pregnancy

c306

health care sought for vaginal problem for that pregnancy

c307

place from where seek health care for the problem

c308

child was single/multiple birth

c309

born after full term or pre-mature or after term

c310

month sof pregnancy completed

c311

plae of brith of deceased child

c312

person who attended the delivery

c313

duration of labour

c314

any intervention or procedure done during the labor

c315

type of interventions performed during labor

c316u

time taken in the delivery (unit)

c316n

time taken in the delivery (period)

fc317

q317 (filter): chekc 126

c318

weight of baby at birth

c319

size of birth (show photo)

c320n.01

number of vaccination

c320n.02

number of vaccination

c320n.03

number of vaccination

c320n.04

number of vaccination

c320n.05

number of vaccination

c320n.06

number of vaccination

c320n.07

number of vaccination

c320n.08

number of vaccination

c320n.09

number of vaccination

c320n.10

number of vaccination

c320n.11

number of vaccination

c320n.12

number of vaccination

c320.01

received vaccination

c320.02

received vaccination

c320.03

received vaccination

c320.04

received vaccination

c320.05

received vaccination

c320.06

received vaccination

c320.07

received vaccination

c320.08

received vaccination

c320.09

received vaccination

c320.10

received vaccination

c320.11

received vaccination

c320.12

received vaccination

c321

mother recieved tt vaccination

c322

number of tt injection recieved

fc323

q323 (filter): check 125u and n

c301b.130

during pregnancy, mother experienced vomiting (duration week

c301c.130

during pregnancy, mother experienced vomiting at stage of pr

c301b.14

during pregnancy, mother experienced uti (duration weeks)

c301c.14

during pregnancy, mother experienced uti at stage of preg in

c401

baby moved few days before the delivery

c402u

time when feel movements of the baby (unit)

c402n

time when feel movements of the baby (period)

c403

mother having excess fluid in the womb

c404

color of liqour during delivery

c405

liqour foul smelling

c406

part of the baby came out first

c407

cord around the neck of the dead baby

c408

gross physical deformity in the dead baby

c409

head not properly formed

c410

mass or defect on the back

c411

any cleft lip or cleft palate in the dead baby

c412

dead baby maerated

c501

died from an injury or accident

c502

kind of injury or accident

fc503

q503 (filter): check 126

c601

water bag rapture during labor ain

c602u

time before labor started when bag rapture

c602n

time before labor started when bag rapture (period)

c603

color of liqour

c604

color of child at birth

c605

color of child after 5-minutes of birth

c606

baby breathe immediately after birth

c607

child cry after birth

c608

baby cry feeble or strong

c609

any green marks of meconium on the child's body

c610

any bruises or marks of injury on the child's body

c611

newborn have swelling(s) over the skull

c612

any physical deformity in the baby

c613

head size not properly formed

c614

mass or defect on the back of head or spine

c615

any cleft lip or cleft palate

c616

any limp defect

c617

child limp/flaccid during first 72 hours

c618u

child start sucking on the breast or feed bottle after birth

c618n

child start sucking on the breast or feed bottle after birth

c619u

child stop sucking or bottle-feeding before death (unit)

c619n

child stop sucking or bottle-feeding before death (period)

c620u

how long before death, infant stop crying (unit)

c620n

how long before death, infant stop crying (period)

c621

body of child become stiff with the back of arched

c622

new born have convulsions during first 24-hours

c623

safe delivery kit using during the delivery

c624

used to tie the umbilical cord

c625

used to cut the umbilical cord

c626

instruments boiled before using it

c627

child have tetanus

c628

child become unresponsive/unconcious during the illness

c629

child have bulging fontanelle during hte illness

c630

child have jaundice or yellow discolorationof skin

c631

child have redness or pus oozing from the umbilical cord

c632

child have areas of skin that were red and hot

c633

child have skin rash with pus

c634

skin peel off after the rash started on the body of child

c635

child have ear discharge

c636

child become lethargic at any stage of the illness

c637

child have a fever at any stage of the illness

c638u

how long did fever last (unit)

c638n

how long did fever last (period)

c639

child have frequent loose or watery stools

c640u

how long did fever last (unit)

c640n

how long did fever last (period)

c641

visible blood in the stools in the stage of diarrhea

c642

child have abdominal distension in the stage of diarhea

c643

child take any liquids during loose or watery stools

c644

child take ors during loose and watery stools

c645

child have cough

c646u

how long cough last (unit)

c646n

how long cough last (period)

c647

child have difficulty in breathing

c648u

how long difficult breathing last (unit)

c648n

how long difficult breathing last (period)

c649

child have fast breathing

c650u

how long difficult fast breathing last (unit)

c650n

how long difficult fast breathing last (period)

c651

child ever stop breathing for a long time and start again

c652

child have chest in-drawing

c653

child have noisy breathing (stridor)

c654

child have noisy breathing (grunting)

c655

child have noisy breathing (wheezing)

c656

child nostril's flare with breathing

c657

child have pneumonia

c658

child become cold at the beginning /during the illness

c701

child have a fever

c702u

how long fever last (unit)

c702n

how long fever last (period)

c703

hild become unresponsive/inconcious

c704

child have tetanus

c705

child have bulging fontanile

c706

child have stiff neck

c707

child have convulsions

c708

child have frequent loose or water stools

c709u

how long diarrhea last (unit)

c709n

how long diarrhea last (period)

c710

visible blood in the stools during diarrhea

c711

child have abdominal distension during diarrhea

c712

child drink anything during loose of watery stools

c712a

child take ors dring diarrhea

c713

child have cough

c714u

how long cough last (unit)

c714n

how long cough last (period)

c715

severity of cough

c716

child have difficult breathing

c717u

how long difficult breathing last (unit)

c717n

how long difficult breathing last (period)

c718

child have fast breathing

c719u

how long child have fast breathing last (unit)

c719n

how long child have fast breathing last (period)

c720

child ever stop breathing fo a long time and start again

c721

child have chest in-drawing

c722

child have noisy breathing (stridor)

c723

child have noisy breathing (glunting)

c724

child have noisy breathing (wheezing)

c725

child's nostris flare with breathing

c726

child have pneumonia

c727

child become cold at the begining of illness

c728

child become cold during the illness

c729

child have yellow eyes

c730

child have jaundice or yellow coloured skin

c731

child have skin rash

c732

rash all over the child's body

c733

rash also on child's face

c734

number of days rash last

c735

rash contain clear fluid

c736

skin crack or peel off after the rash started

c737

child have measles

c738

child become very thin

c739

child become very weak

c740

child have swelling on legs or feet

c741

days swelling lasted

c742

child's skin flake off in patches

c743

child's hair color change to reddish

c744

child have kwashiorkor

c745

child have "maramus"

c746

child suffer from "lack of blood" or "pallor"

c747

child have pale palms

c748

child have white nails

c749

child have swellings in the armpits

c750

child have swelling in the groin

c751

child have a whitish rash inside the mouth/tongue

c801.1

medicines received

c801.2

medicines received

c801.3

medicines received

c801.4

medicines received

c801.5

medicines received

c802

have any health records that belong to the child

c803

records for child is available at the time of interview

c804w1

record most recent weights of the child-1 (weight)

c8041d

record most recent weights of the child-1 (date-day)

c8041m

record most recent weights of the child-1 (date-month)

c8041y

record most recent weights of the child-1 (date-year)

c804w2

record most recent weights of the child-2 (weight)

c8042d

record most recent weights of the child-2 (date-day)

c8042m

record most recent weights of the child-2 (date-month)

c8042y

record most recent weights of the child-2 (date-year)

c805

death certificate sought for the deceased child

c806

record the immediate cause of death from the certificate

c807a

record the other causes from death certificate (1)

c807b

record the other causes from death certificate (2)

c807c

record the other causes from death certificate (3)

c801.6

child received injection during illness

c801.7

child received oxygen during illness

c901

mother have any complications, problems during last trimeste

c902a

nature of illness which mother faced during last 3-months of

c902b

nature of illness which mother faced during last 3-months of

c902c

nature of illness which mother faced during last 3-months of

c903

mother seek advice/treatment for the problem(s)

c904

person to whom seek advice/treatment

c905

mother seek anyone for this problem

c906

place of delivery

c907

reason for not delivering baby in the health facility

c908u

age when noticed that child not well (unit)

c908n

age when noticed that child not well (period)

c909a

signs and symptoms noticed in the baby(1)

c909b

signs and symptoms noticed in the baby(2)

c909c

signs and symptoms noticed in the baby(3)

c910

did on the first day of illness of a child

c911

reason for doing nothing on the first day of illness

c912

reason for seek home care on the first day

c913

persons to whom seek care on the first day

c914

what did care provider do

c915

position of child on the second day of illness

c916

did on the second day of illness

c917

reason for doing nothing on second day of illness

c918

reason for seek home care on the second day

c919

person to whom seek care on the second day

c920

what did the care provider do

c921

position of child between 4-8 days of illness

c922

did on the third day of illness

c923

reason for doing nothing on 4-8 days of illness

c924

reason for seek home care on the 4-8 days

c925

person to whom seek care on the 4-8 days

c926

what did care provider do on the 4-8 days of illness

c927

position of child on the last day of illness

c928

did on the last day of illness

c929

reason for doing nothing on the last day of illness

c930

reason for seek home care on the last day of illness

c931

person to whom seek care on the last day of illness

c932

what did care provider do on the last day of illness

c933

time when mother share with spouse about the illness of a ba

c934a

reaction of spouse in reply(1)

c934b

reaction of spouse in reply(2)

c934c

reaction of spouse in reply(3)

c935

person who mainly takes the care seeking decisions

c936a

reason why person (who take final say) take decisions(1)

c936b

reason why person (who take final say) take decisions(2)

c936c

reason why person (who take final say) take decisions(3)

c937

persons who took the child for seeking care

c938a

reason for person who accompanied go for seeking care(1)

c938b

reason for person who accompanied go for seeking care(2)

c938c

reason for person who accompanied go for seeking care(3)

c939

person who decided the care should be sought outside home

c940

respondent's view about illness

c941

place where usually go to recieve health care

c942

distance of nearest health facility from home

c943

mode of transports use to reach the nearest health facility

c944

transportation cost to reach the nearest health facility

c945u

time when household member last time visit a health care fac

c945n

time when household member last time visit a health care fac

fc946

q946 (filter): check 916, 922, 928

c947

easy to see health care provider regarding child

c948

conduct of physician( who saw the child towards the baby)

c949

conduct of physician( who saw the child towards the caretake

c950

easily able to purchase/acquire the drugs needed for the chi

c951

reasons for not able to purchase.acquire the drugs

c952

overall satisfied with the qaulity of care provided at the h

c953a

reasons for satisfactions(1)

c953b

reasons for satisfactions(2)

c953c

reasons for satisfactions(3)

References

National Institute of Population Studies (NIPS), Pakistan (http://nips.org.pk/Home.htm).

Examples

load("PKVC51")

#> Warning: cannot open compressed file 'PKVC51', probable reason 'No such file or directory'

#> Error in readChar(con, 5L, useBytes = TRUE): cannot open the connection

PKVC51

#> # A tibble: 4,438 x 579
#>    qhclust qhnumber qqtype qline rec.type qcluster qnumber qregion qdist qlocal3
#>      <int>    <int> <fct>  <int> <chr>       <int>   <int> <fct>   <int> <fct>  
#>  1       1       89 still…     2 CVA             1      89 punjab      1 other …
#>  2       2       32 still…     1 CVA             2      32 punjab      1 other …
#>  3       2       60 still…     1 CVA             2      60 punjab      1 other …
#>  4       3       50 still…     3 CVA             3      50 punjab      1 rural  
#>  5       3       50 child…     1 CVA             3      50 punjab      1 rural  
#>  6       3       50 child…     2 CVA             3      50 punjab      1 rural  
#>  7       3       54 child…     1 CVA             3      54 punjab      1 rural  
#>  8       3       54 child…     2 CVA             3      54 punjab      1 rural  
#>  9       3       76 child…     1 CVA             3      76 punjab      1 rural  
#> 10       4        2 child…     1 CVA             4       2 punjab      1 rural  
#> # … with 4,428 more rows, and 569 more variables: qlocal4 <fct>, typeq <chr>,
#> #   qintd <int>, qintm <int>, qinty <int>, qintnum <int>, qresult <fct>,
#> #   qvisits <int>, qqlang <fct>, qilang <fct>, qwlang <fct>, qsuperv <int>,
#> #   qfedit <int>, qoedit <int>, qkeyer <int>, qintc <int>, qweight <int>,
#> #   qtype <fct>, qres <fct>, c101.1 <int>, c101.2 <int>, c101.3 <int>,
#> #   c101.4 <int>, c101.5 <int>, c101.6 <int>, c103.1 <int>, c103.2 <int>,
#> #   c103.3 <int>, c103.4 <int>, c103.5 <int>, c103.6 <int>, c104.1 <int>,
#> #   c104.2 <int>, c104.3 <int>, c104.4 <int>, c104.5 <int>, c104.6 <int>,
#> #   c105.1 <int>, c105.2 <int>, c105.3 <int>, c105.4 <int>, c105.5 <int>,
#> #   c105.6 <int>, c106.1 <int>, c106.2 <int>, c106.3 <int>, c106.4 <int>,
#> #   c106.5 <int>, c106.6 <int>, c107.1 <int>, c107.2 <int>, c107.3 <int>,
#> #   c107.4 <fct>, c107.5 <fct>, c107.6 <fct>, c108.1 <int>, c108.2 <int>,
#> #   c108.3 <int>, c108.4 <int>, c108.5 <fct>, c108.6 <fct>, c109.1 <int>,
#> #   c109.2 <int>, c109.3 <int>, c109.4 <int>, c109.5 <fct>, c109.6 <fct>,
#> #   c110.1 <int>, c110.2 <int>, c110.3 <int>, c110.4 <int>, c110.5 <int>,
#> #   c110.6 <fct>, c100 <int>, c111h <int>, c111m <int>, fc112 <fct>,
#> #   c113 <int>, c114 <int>, c115 <int>, c116m <int>, c116y <int>, c117m <int>,
#> #   c117y <int>, c118 <int>, c119d <int>, c119m <int>, c119y <int>, c120 <fct>,
#> #   c121 <int>, c122 <int>, c123 <int>, c124d <int>, c124m <int>, c124y <int>,
#> #   c125u <fct>, c125n <int>, c126 <int>, kcod.1 <fct>, kcod.2 <fct>, …