Pakistan Demographic and Health Survey 2006-07 Vaccination of Women Data

PKVW51 data of Vaccination of Women from Pakistan Demographic and Health Survey 2006-07.

qhclust

cluster number

qhnumber

household number

qqtype

type of questionnaire (numeric)

qline

line number of woman/deceased woman/child/stillbirth

rec.type

qcluster

cluster number

qnumber

household number

qregion

region of residence

qdist

district

qlocal3

city \ town \ countryside - 3 values

qlocal4

city \ town \ countryside

typeq

type of questionnaire (alpha)

qintd

day of interview

qintm

month of interview

qinty

year of interview

qintnum

interviewer number

qresult

result of individual interview

qvisits

total number of visits

qqlang

language of questionaire

qilang

language of interview

qwlang

language woman speaks a home

qsuperv

supervisor number

qfedit

field editor number

qoedit

office editor number

qkeyer

data entry clerk

qintc

date of interview (cmc)

qweight

sample weight (6 decimals)

qtype

type of place of residence

qres

usual resident or visitor

m101.1

line number

m101.2

line number

m101.3

line number

m101.4

line number

m101.5

line number

m101.6

line number

m103.1

sex of respondent

m103.2

sex of respondent

m103.3

sex of respondent

m103.4

sex of respondent

m103.5

sex of respondent

m103.6

sex of respondent

m104.1

age of respondent

m104.2

age of respondent

m104.3

age of respondent

m104.4

age of respondent

m104.5

age of respondent

m104.6

age of respondent

m105.1

relationship with deceased woman

m105.2

relationship with deceased woman

m105.3

relationship with deceased woman

m105.4

relationship with deceased woman

m105.5

relationship with deceased woman

m105.6

relationship with deceased woman

m106.1

number of classes passed

m106.2

number of classes passed

m106.3

number of classes passed

m106.4

number of classes passed

m106.5

number of classes passed

m106.6

number of classes passed

m107.1

respondent present when deceased first fell ill

m107.2

respondent present when deceased first fell ill

m107.3

respondent present when deceased first fell ill

m107.4

respondent present when deceased first fell ill

m107.5

respondent present when deceased first fell ill

m107.6

respondent present when deceased first fell ill

m108.1

respondent present when deceased was taken to hospital

m108.2

respondent present when deceased was taken to hospital

m108.3

respondent present when deceased was taken to hospital

m108.4

respondent present when deceased was taken to hospital

m108.5

respondent present when deceased was taken to hospital

m108.6

respondent present when deceased was taken to hospital

m109.1

present when deceased died

m109.2

present when deceased died

m109.3

present when deceased died

m109.4

present when deceased died

m109.5

present when deceased died

m109.6

present when deceased died

m110.1

line number of main respondent

m110.2

line number of main respondent

m110.3

line number of main respondent

m110.4

line number of main respondent

m110.5

line number of main respondent

m110.6

line number of main respondent

m100

number of respondents for woman

m201h

start of interview (hours)

m201m

start of interview (minutes)

m204m

month when deceased died

m204y

year when deceased died

m205

age at the time death

m206

at the time of death, usual member of household

m207

ever attended school

m208

highest classes passed

m209

working for wages or salary when she died

m210

occupation of deceased woman

m211

mother tongue of deceased

m212

marital status at the time death

m214

husband's age at the time deceased death

m215

deceased's husband ever attended school

m216

highest class passed by deceased's husband

m217

deceased's husband occupation

dcat.1

******* direct, indirect, incidental cause of death

dcat.2

******* direct, indirect, incidental cause of death

dcat.3

******* direct, indirect, incidental cause of death

dcat.4

******* direct, indirect, incidental cause of death

vacmi.1

most probable immediate cause

vacmi.2

most probable immediate cause

vacmi.3

most probable immediate cause

vacmi.4

most probable immediate cause

vacmu.1

most probable underlying cause (see vacmi for labels)

vacmu.2

most probable underlying cause (see vacmi for labels)

vacmu.3

most probable underlying cause (see vacmi for labels)

vacmu.4

most probable underlying cause (see vacmi for labels)

vacpi.1

possible immediate cause (see vacmi for labels)

vacpi.2

possible immediate cause (see vacmi for labels)

vacpi.3

possible immediate cause (see vacmi for labels)

vacpi.4

possible immediate cause (see vacmi for labels)

vacpu.1

possible underlying cause (see vacmi for labels)

vacpu.2

possible underlying cause (see vacmi for labels)

vacpu.3

possible underlying cause (see vacmi for labels)

vacpu.4

possible underlying cause (see vacmi for labels)

del1.1

first delay - to get transport

del1.2

first delay - to get transport

del1.3

first delay - to get transport

del1.4

first delay - to get transport

del2.1

second delay - to reach health facility

del2.2

second delay - to reach health facility

del2.3

second delay - to reach health facility

del2.4

second delay - to reach health facility

del3.1

third delay - to get service from provider

del3.2

third delay - to get service from provider

del3.3

third delay - to get service from provider

del3.4

third delay - to get service from provider

dqual.1

f. reviewer's assessment of the quality of data (0-10)

dqual.2

f. reviewer's assessment of the quality of data (0-10)

dqual.3

f. reviewer's assessment of the quality of data (0-10)

dqual.4

f. reviewer's assessment of the quality of data (0-10)

m204c

cmc death

m204f

cmc flag

gprobi

grouped probable immediate cause (2nd set from pakistan)

gprobu

grouped probable underlying cause (see gprobi for labels)

gpossi

grouped possible immediate cause (see gprobi for labels)

gpossu

grouped possible underlying cause (see gprobi for labels)

oldcause

cause of death

findcat

******* direct, indirect, incidental cause of death (origin

gcause1

immediate cause of death (icd-10 codes) final set of codes b

gcause2

immediate cause of death (group)

gcause3

immediate cause of death (group final)

gcause1r

immediate cause of death (icd-10 codes) work for maternal de

gcause1n

immediate cause of death (icd-10 codes) maternal deaths (fin

gcausemr

immediate cause of death (icd-10 codes) work for maternal de

gcausemn

immediate cause of death (icd-10 codes) maternal deaths (mat

vimmn.1

immediate cause of death numeric

vimmn.2

immediate cause of death numeric

vimmn.3

immediate cause of death numeric

vimmn.4

immediate cause of death numeric

vimmgn.1

immed group numeric

vimmgn.2

immed group numeric

vimmgn.3

immed group numeric

vimmgn.4

immed group numeric

vundn.1

underlying cause of death numeric ( see vimmn for labels)

vundn.2

underlying cause of death numeric ( see vimmn for labels)

vundn.3

underlying cause of death numeric ( see vimmn for labels)

vundn.4

underlying cause of death numeric ( see vimmn for labels)

vunngn.1

under numeric group

vunngn.2

under numeric group

vunngn.3

under numeric group

vunngn.4

under numeric group

vfidx

number of underlying cause coded by reviewers

matdth

when died according to doctorrs' evaluation (from pakistan)

gcauseu1

underlying cause of death (icd-10 codes)

gcauseu3

underlying cause of death (group final)

gcausu1r

underlying cause of death (icd-10 codes) work for maternal d

gcausu1n

underlying cause of death (icd-10 codes) maternal deaths (fi

gcauseur

underlying cause of death (icd-10 codes) work for maternal d

gcauseun

underlying cause of death (icd-10 codes) maternal deaths (ma

mideath

maternal immediate cause of death for direct/indirect matern

mudeath

maternal underlying cause of death for direct/indirect mater

fm300

never married but with data in section 3

m301

ever have live birth

m302

number of live births she had

m303t

number of total surviving children

m303b

number of surviving boys

m303g

number of surviving girls

m304

ever have miscarriage, abortion etc

m305

number of pregnancy losses

fm306

q306 (filter): check 301 and 304

m307

had caesarean section

m308

had pregnancy during last 3-yrs, regardless of outcome

m309u

time before her death,last pregnancy end (unit)

m309n

time before her death, last pregnancy end (period)

m310

outcome of last pregnancy

m311

last born child still alive

m312

age of last alive child

m313u

age of child at the time deceased's death (unit)

m313n

age of child at the time deceased's death (period)

m314u

time after her last birth/delivery/abortion etc (unit)

m314n

time after her last birth/delivery/abortion etc (period)

m315

pregnant at the time death

m316

months of pregnancy when died

m317

died during delivery, abortion or miscarriage

m318

died before labour pain begans, bfore birth or during aborti

m319

died after delivery, abortion or miscarriage

m320u

period after delivery, abortion or miscarriage died (unit)

m320n

period after delivery, abortion or miscarriage died (period)

m321

died within 40-days of delivery, abortion or miscarriage

m501

place of death

m502

name of hospital/clinic where she died

m503

anyone at the hospital told the reason of death

m504a

reasons given by the hospital/clinic(1)

m504b

reasons given by the hospital/clinic(2)

m504c

reasons given by the hospital/clinic(3)

m505a

main cause of death in respondent's view(1)

m505b

main cause of death in respondent's view(2)

m505c

main cause of death in respondent's view(3)

m506.01

high blood pressure or hypertension

m506.02

high blood pressure or hypertension

m506.03

high blood pressure or hypertension

m506.04

high blood pressure or hypertension

m506.05

high blood pressure or hypertension

m506.06

high blood pressure or hypertension

m506.07

high blood pressure or hypertension

m506.08

high blood pressure or hypertension

m506.09

high blood pressure or hypertension

m506.10

high blood pressure or hypertension

m506.11

high blood pressure or hypertension

m506.12

high blood pressure or hypertension

m506.13

high blood pressure or hypertension

m507

deceased was hospitalized

m508u

duration of hospitalization before death (unit)

m508n

duration of hospitalization before death (period)

m509a

reason or hospitalization(1)

m509b

reason or hospitalization(2)

m509c

reason or hospitalization(3)

m510

have operation before death

m511

have fever

m511u

days/months her fever last (unit)

m511n

days/months her fever last (period)

m511b

was fever continous or on and off

m512

was breathless doing light work

m512a

was breathless when she was lying down/asleep

m513

have rapid heart beat (palpitations)

m514

have wheezing

m515

have a cough

m515u

how long have a cough (unit)

m515n

how long have a cough (period)

m515b

cough produce sputum

m515c

cough has a blood

m516

have a chest pain

m516u

days/months has chest pain (unit)

m516n

days/months has chest pain (period)

m516b

chest pain mild, moderate or severe

m516c

chest pain starts suddenly or gradually

m516d

pain at or near the center of chest

m517

have diarrhea

m517a

number of times a day have loose motion

m517b

have blood in the stools

m518

have poor appetite or loss of apetite

m518u

how long have poor apetite (unit)

m518n

how long have poor apetite (period)

m519

pain in swallowing

m520

difficulty in swallowing

m521

have headache

m521u

days/months have headache (units)

m521n

days/months have headache (period)

m522

passed blood in her urine

m522u

days/months passed blood in urine (units)

m522n

days/months passed blood in urine (period)

m523

pain while urinating

m523u

days/months have pain when urinating (units)

m523n

days/months have pain when urinating (period)

m524

unable to pass urine

m525

urinate many times in a day

m526

have any type of pain anywhere in the body

m527

have abdominal pain

m527u

how long abdominal pain last (units)

m527n

how long abdominal pain last (period)

m527b

abdominal pain mild, moderate or severe

m527c

abdominal pian in upper belly , lower or all over belly

m528

have abdominal distension

m528u

days/months was abdomen distened (units)

m528n

days/months was abdomen distened (period)

m528b

distension come rapidly within days or slowly over several w

m529

have a mass in her abdomen

m530

have vomiting

m530u

days/months have vomitting (unit)

m530n

days/months have vomiting (period)

m530b

vomitted blood

m531

become mentally confuse

m532

loose consciousness

m532u

how long remained unconcious (units)

m532n

how long remained unconcious (period)

m532b

became unconcious suddenly or gradually

m533

became paralyze before death

m533u

how long the paralyze last (units)

m533n

how long the paralyze last (period)

m533b

paralysis on only one side or both side of body

m534

have stiffness in whole body

m535

have neck pain

m536

have fits or convulsions

m536u

days/months have fits (units)

m536n

days/months have fits (period)

m536b

frequency of fits in a day

m537

have an ulcer or swelling in the breast

m538

have vaginal bleeding when was not having mensrual period

m538u

days/months did have bleeding (units)

m538n

days/months did have bledding (period)

m538b

did the bleeding persist till death

m539

have abnormal vaginal discharge

m540

have swelling on her ankles

m541

have swelling or puffiness on hands and/or face

m541u

days/months have swelling on hands or face (units)

m541n

days/months have swelling on hands or face (period)

m542

lose weight

m543

have sores in the mouth

m544

looked pale

m545

have any skin disease

m546

eyes yellowish in color due to jaundice

m546u

days/months have yellow eyes (units)

m546n

days/months have yellow eyes (period)

m547

ever complain of having blurred vision

m548

have difficulty in openng mouth

m549

have difficulty in passing stools

m550

feel dizzy

m551

have general weakness or fatigue

m551u

days/months have weakness (units)

m551n

days/month have weakness (period)

m552

have any ulcers on the body

m553a

other symptoms not elsewhere mention(1)

m553b

other symptoms not elsewhere mention(2)

m553c

other symptoms not elsewhere mention(3)

m554

people think that she ha an evil eye or shadow

m555

faith healer called or give amulets or spiritual healing

m556a

give details(1)

m556b

give details(2)

m556c

give details(3)

fm601

q601 (filter): check 511

m602u

how long before she died did the fever start (unit)

m602n

how long before she died did the fever start (period)

m603u

how long did the fever last (unit)

m603n

how long did the fever last (period)

m604

was the fever very high

m605

have fever with chills

m606

was prescribed anti-malarial tablets for fever

m607

colour changed during her last illness

m607a

deceased's colour was

m608

had vomiting during last illness

m608u

how long before she died did the vomiting start (unit)

m608n

how long before she died did the vomiting start (period)

m609

ever vomit pure blood

m610

have any difficulty with urination

m610a.1

urination problems

m610a.2

urination problems

m610a.3

urination problems

m610a.4

urination problems

m610a.5

urination problems

m610a.6

urination problems

m611

when did the feve start

m612u

how long before/after child birth, miscarriage or abortion f

m612n

how long before/after cild birth, miscarriage or abortion fe

m613

have convulsions with fever

fm614

q614 (filter): check 515

m615u

how long before she died did the cough start (units)

m615n

how long before she died did the cough start (period)

m616

was any sputum when she coughed

m617

coughed with blood

m617a

details for coughing(1)

m617b

details for coughing(2)

m617c

details for coughing(3)

m618

loose weight during illness

m619

ever have any fever

m619a

how much fever

m620

was short of breath

m620u

how long have short of breath (units)

m620n

how long have short of breath (period)

fm621

q621 (filter): check 526

m622

kind of pain

m623

site(s) of the pain

m624u

how long before she died did the pain start (units)

m624n

how long before she died did the pain start (period)

m625u

how long did it last the pain (units)

m625n

how long did it last the pain (period)

m626a

if it was in abdomen, which specific side(1)

m626b

if it was in abdomen, which specific side(2)

m627

was any pain in th elower abdomen

m628

was the pain accompanied by fever

m629

was the fever mild, moderate or high

m630

was the pain accompanied by vomiting

m631

when did the pain start

fm632

q632 (filter): check 536

m633

have a history of convulsions or epilepsy

m634

have convulsions in last illness

m634u

how long before she died did the pain start (units)

m634n

how long before she died did the pain start (period)

m635

have high blood pressure before death

m636

have severe headache before she died

m637

have change in her vision before she died

m638a

state of conciousness before she died(1)

m638b

state of conciousness before she died(2)

fm639

q639 (filter): check 540 and 541

m640.1

place of site of swelling on the body

m640.2

place of site of swelling on the body

m640.3

place of site of swelling on the body

m640.4

place of site of swelling on the body

m641u

how long before she died did have this swelling (units)

m641n

how long before she died did have this swelling (period)

m642u

time when swelling start in relation to delivery of the baby

m642n

time when swelling start in relation to delivery of the baby

m643

time of death, was short of breath

m644

colour changed during her last illness

m645

have also any urinary problems

m645a.1

urinary problems

m645a.2

urinary problems

m645a.3

urinary problems

m645a.4

urinary problems

m645a.5

urinary problems

fm701

q701 (filter): check 308

m702

during last pregnancy, see anyone for antenatal care

m703

whom to see antenatal care

m704

first visit for antenatal care, for any problem or for check

m705a

type of problem to visit for antenatal care(1)

m705b

type of problem to visit for antenatal care(2)

m705c

type of problem to visit for antenatal care(3)

m706

number of pregnant month on first visit for antenatal care

m707

number of visits for antenatal care in the last pregnancy

m708

was referred by health care provider to go to a specialist

m709

reasons for not antenatal care during last pregnancy

m710

have tetanus injection during last pregnancy

m711

have blood pressure measure during last pregnancy

m712

status of blood pressure during last pregnancy

m713

prescribed medicines for blood pressure

m714

using any family planning method before she became pregnant

m715

wanted the last pregnancy

m716

did any thing to end last pregnancy

m717

steps taken to end the last pregnancy

m718

person whom to seek help to end last pregnancy

fm719

q719 (filter): check 310

m720

have a foul-smelling discharge from vagina after miscarriage

m721

have a fever after miscarriage/abortion

m722

have abdominal distension after the miscarriage/abortion

m723

during last pregnancy, have swelling around ankles

m724

face puffy during pregnancy

m725

during last pregnancy, complained about blurred vision

m726

during last pregnancy, have bleeding from vagina

m727

bleeding wet her clothes

m728

bleeding wet the bed

m729

was there so much blood as to wet the floor

m730

was in pain while bleeding

m731

did the bleeding contains clots

m732

someone examine her internally during last pregnancy

m733

during vaginal examination, cause any bleeding or make th eb

m734

did the bleeding persist until she died

fm801

q801 (filter): check 317, 319, 321

m802

have bleeding from vagina

m803

bleeding start before or after the birth/delivery

m804

was she in pain while bleeding

m805

pain start before the labor pains started

m806

place of delivery

m807

person who assisted with the delivery

m808

any instruments used to assist in last pregnancy

m809

during last pregnancy, someone cut vagina to make room for t

m810

was a tear in vagina after last delivery

m811

during the delivery, birth attendant examine vagina

m812

water bag break before the labor pains started

m813

how long was in labor

m814

was given any drugs just before or during labor

m815

what were the drugs used for

m816

have a lot bleeding before delivering the baby

m817

the bleeding wet her clothes

m818

the bleeding wet the bed

m819

was there so much blood as to wet the floor

m820

died before or after delivering the baby

m821

have facing difficulty in delivering the baby

m822

part of baby came out first

m823

did the placenta came out

m824u

how long after the baby came did the placenta come out (unit

m824n

how long after the baby came did the placenta come out (peri

m825

all the placenta come out or only part

m826

have a lot of bleeding after delivering the baby

m827

the bleeding wet her clothes

m828

the bleeding wet the bed

m829

was there so much blood as to wet the floor

m830

have a foul-smelling discharge from her vagina after the bab

m831

have a pain in her legs after the baby was born

m832

have a fever after the baby was born

m833u

how long after the delivery did the fever start (units)

m833n

how long after the delivery did the fever start (period)

m834

have any fits or rigors with the fever

m835

the fits stop after the baby was born

m836u

how long did the fever last (units)

m836n

how long did the fever last (period)

m837

was having the fever when she died

m838

was the colour of eyes yellow after the baby was born

m839

after the delivery, birth attendant examine vagina

m901

have any serious accient or injury before she died

m902

details of accident or injury

m903u

how long before she died did this happen (units)

m903n

how long before she died did this happen (period)

m904

hurt herself intentionally or someone else did this

m905

think that was trying to commit suicide

m906

think that injury was the main cause of her death

References

National Institute of Population Studies (NIPS), Pakistan (http://nips.org.pk/Home.htm).

Examples

data("PKVW51")
PKVW51

#> # A tibble: 1,062 x 500
#>    qhclust qhnumber qqtype qline rec.type qcluster qnumber qregion qdist qlocal3
#>      <int>    <int> <fct>  <int> <chr>       <int>   <int> <fct>   <int> <fct>  
#>  1       1       34 woman…     1 WVA             1      34 punjab      1 other …
#>  2       1       87 woman…     1 WVA             1      87 punjab      1 other …
#>  3       2       93 woman…     1 WVA             2      93 punjab      1 other …
#>  4       4       64 woman…     1 WVA             4      64 punjab      1 rural  
#>  5       7        5 woman…     2 WVA             7       5 punjab      1 rural  
#>  6       7        5 woman…     3 WVA             7       5 punjab      1 rural  
#>  7       7       40 woman…     1 WVA             7      40 punjab      1 rural  
#>  8       7       81 woman…     1 WVA             7      81 punjab      1 rural  
#>  9       7       88 woman…     1 WVA             7      88 punjab      1 rural  
#> 10       7       97 woman…     2 WVA             7      97 punjab      1 rural  
#> # … with 1,052 more rows, and 490 more variables: qlocal4 <fct>, typeq <chr>,
#> #   qintd <int>, qintm <int>, qinty <int>, qintnum <int>, qresult <fct>,
#> #   qvisits <int>, qqlang <fct>, qilang <fct>, qwlang <fct>, qsuperv <int>,
#> #   qfedit <int>, qoedit <int>, qkeyer <int>, qintc <int>, qweight <int>,
#> #   qtype <fct>, qres <fct>, m101.1 <int>, m101.2 <int>, m101.3 <int>,
#> #   m101.4 <int>, m101.5 <int>, m101.6 <int>, m103.1 <int>, m103.2 <int>,
#> #   m103.3 <int>, m103.4 <int>, m103.5 <int>, m103.6 <int>, m104.1 <int>,
#> #   m104.2 <int>, m104.3 <int>, m104.4 <int>, m104.5 <int>, m104.6 <int>,
#> #   m105.1 <int>, m105.2 <int>, m105.3 <int>, m105.4 <int>, m105.5 <int>,
#> #   m105.6 <int>, m106.1 <int>, m106.2 <int>, m106.3 <int>, m106.4 <int>,
#> #   m106.5 <int>, m106.6 <int>, m107.1 <int>, m107.2 <int>, m107.3 <fct>,
#> #   m107.4 <int>, m107.5 <fct>, m107.6 <fct>, m108.1 <int>, m108.2 <int>,
#> #   m108.3 <fct>, m108.4 <fct>, m108.5 <fct>, m108.6 <fct>, m109.1 <int>,
#> #   m109.2 <int>, m109.3 <fct>, m109.4 <int>, m109.5 <fct>, m109.6 <int>,
#> #   m110.1 <int>, m110.2 <int>, m110.3 <int>, m110.4 <int>, m110.5 <int>,
#> #   m110.6 <int>, m100 <int>, m201h <int>, m201m <int>, m204m <int>,
#> #   m204y <int>, m205 <int>, m206 <int>, m207 <fct>, m208 <int>, m209 <fct>,
#> #   m210 <int>, m211 <int>, m212 <int>, m214 <int>, m215 <fct>, m216 <int>,
#> #   m217 <int>, dcat.1 <fct>, dcat.2 <fct>, dcat.3 <fct>, dcat.4 <fct>,
#> #   vacmi.1 <chr>, vacmi.2 <chr>, vacmi.3 <chr>, vacmi.4 <chr>, vacmu.1 <chr>,
#> #   vacmu.2 <chr>, …